% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_metrics_sfs.R
\name{alpha_metrics_sfs}
\alias{alpha_metrics_sfs}
\title{compute alpha metrics during sfs}
\usage{
alpha_metrics_sfs(
  alpha_metrics,
  obs2optimize,
  SSValid,
  corrMethod,
  nb_clusters,
  Hill_order,
  pcelim,
  nbPlots_total
)
}
\arguments{
\item{alpha_metrics}{character}

\item{obs2optimize}{numeric}

\item{SSValid}{numeric}

\item{corrMethod}{character}

\item{nb_clusters}{numeric}

\item{Hill_order}{numeric}

\item{pcelim}{numeric}

\item{nbPlots_total}{numeric}
}
\value{
assessed value and correlation
}
\description{
compute alpha metrics during sfs
}
