% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabeta_chunk.R
\name{alphabeta_chunk}
\alias{alphabeta_chunk}
\title{apply alphabeta_window to a list of lists}
\usage{
alphabeta_chunk(
  SSwindow,
  nb_clusters,
  Beta_info = NULL,
  alpha_metrics = "shannon",
  Hill_order = 1,
  beta_metrics = "bray",
  diss_output = FALSE,
  pcelim = 0.02,
  p = NULL,
  ...
)
}
\arguments{
\item{SSwindow}{dataframe. spectral species corresponding to a raster subset
(window = elementary spatial unit of process)}

\item{nb_clusters}{numeric. number of clusters used in kmeans}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics}

\item{alpha_metrics}{list. alpha diversity metrics}

\item{Hill_order}{numeric. Hill order}

\item{beta_metrics}{list. beta diversity metrics}

\item{diss_output}{boolean set TRUE to get dissimilarity matrices as outputs}

\item{pcelim}{numeric. min proportion of pixels to consider spectral species}

\item{p}{list. progressor object for progress bar}

\item{...}{list. additional parameters}
}
\value{
list of alpha and beta diversity metrics
}
\description{
apply alphabeta_window to a list of lists
}
