% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabeta_window.R
\name{alphabeta_window}
\alias{alphabeta_window}
\title{compute alpha and beta diversity metrics from pixel data corresponding to
spectral species extracted from a window}
\usage{
alphabeta_window(
  SSwindow,
  nb_clusters,
  Beta_info,
  alpha_metrics,
  beta_metrics,
  pcelim = 0.02,
  Hill_order = 1,
  diss_output = FALSE,
  p = NULL
)
}
\arguments{
\item{SSwindow}{dataframe. spectral species corresponding to a raster subset
(window = elementary spatial unit of process)}

\item{nb_clusters}{numeric. number of clusters used in kmeans}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics}

\item{alpha_metrics}{list. alpha diversity metrics}

\item{beta_metrics}{list. beta diversity metrics}

\item{pcelim}{numeric. min proportion of pixels to consider spectral species}

\item{Hill_order}{numeric. Hill order}

\item{diss_output}{boolean set TRUE to get dissimilarity matrices as outputs}

\item{p}{list. progressor object for progress bar}
}
\value{
list of alpha and beta diversity metrics corresponding to SSwindow
}
\description{
compute alpha and beta diversity metrics from pixel data corresponding to
spectral species extracted from a window
}
