% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabeta_window_list.R
\name{alphabeta_window_list}
\alias{alphabeta_window_list}
\title{apply alphabeta_window to a list of lists}
\usage{
alphabeta_window_list(
  SSwindow,
  nb_clusters,
  Beta_info = NULL,
  alpha_metrics = "shannon",
  Hill_order = 1,
  beta_metrics = "bray",
  diss_output = FALSE,
  pcelim = 0.02
)
}
\arguments{
\item{SSwindow}{dataframe. spectral species corresponding to a raster subset
(window = elementary spatial unit of process)}

\item{nb_clusters}{numeric. number of clusters used in kmeans}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics}

\item{alpha_metrics}{list. alpha diversity metrics}

\item{Hill_order}{numeric. Hill order}

\item{beta_metrics}{list. beta diversity metrics}

\item{diss_output}{boolean set TRUE to get dissimilarity matrices as outputs}

\item{pcelim}{numeric. min proportion of pixels to consider spectral species}
}
\value{
list of alpha and beta diversity metrics
}
\description{
apply alphabeta_window to a list of lists
}
