% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabeta_window_sdm.R
\name{alphabeta_window_sdm}
\alias{alphabeta_window_sdm}
\title{compute alpha and beta diversity metrics from pixel data corresponding to
spectral species extracted from a window}
\usage{
alphabeta_window_sdm(
  ssd,
  Beta_info,
  alpha_metrics,
  beta_metrics,
  Hill_order = 1,
  nbCPU = 1
)
}
\arguments{
\item{ssd}{list.}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics}

\item{alpha_metrics}{list. alpha diversity metrics}

\item{beta_metrics}{list. beta diversity metrics}

\item{Hill_order}{numeric. Hill order}

\item{nbCPU}{numeric. Number of CPUs available}
}
\value{
list of alpha and beta diversity metrics corresponding to ssd
}
\description{
compute alpha and beta diversity metrics from pixel data corresponding to
spectral species extracted from a window
}
