% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ContinuumRemoval.R
\name{Apply.Continuum.Removal}
\alias{Apply.Continuum.Removal}
\title{prepares data to run multithreaded continuum removal}
\usage{
Apply.Continuum.Removal(Spectral.Data, Spectral, nbCPU = 1)
}
\arguments{
\item{Spectral.Data}{initial data matrix (nb samples x nb bands)}

\item{Spectral}{information about spectral bands}
}
\value{
samples from image and updated number of pixels to sampel if necessary
}
\description{
prepares data to run multithreaded continuum removal
}
