% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapSpectralSpecies.R
\name{Apply.Kmeans}
\alias{Apply.Kmeans}
\title{Applies Kmeans clustering to PCA image}
\usage{
Apply.Kmeans(PCA.Path, PC.Select, ImPathShade, Kmeans.info,
  Spectral.Species.Path, nbCPU = 1, MaxRAM = FALSE)
}
\arguments{
\item{PCA.Path}{path for the PCA image}

\item{PC.Select}{PCs selected from PCA}

\item{ImPathShade}{shade mask}

\item{Kmeans.info}{information about kmeans computed in previous step}

\item{Spectral.Species.Path}{path for spectral species file to be written}
}
\description{
Applies Kmeans clustering to PCA image
}
