% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_bigRaster.R
\name{apply_bigRaster}
\alias{apply_bigRaster}
\title{This function performs big raster processing
It requires user to write a function expecting vectors / matrices / dataframes
corresponding to raster data read sequentially, a list of input files, a list
of output files, and it is good to go.}
\usage{
apply_bigRaster(
  funct,
  input_rasters,
  input_args = NULL,
  output_rasters,
  output_lyrs = NULL,
  filetype = "EHdr",
  bandNames = NULL,
  unit_nrows = NULL,
  maxRows = NULL
)
}
\arguments{
\item{funct}{function. function to be applied to process chunks of big rasters}

\item{input_rasters}{list. path for input files to be processed}

\item{input_args}{list. additional arguments sent to funct}

\item{output_rasters}{list. path for output file to be written}

\item{output_lyrs}{numeric. number of bands for output raster}

\item{filetype}{character. name of GDAL driver}

\item{bandNames}{list. list of band names corresponding to output files}

\item{unit_nrows}{numeric. define to read a number of lines multiple of unit_nrows}

\item{maxRows}{numeric. maximumnumber of rows to process at once}
}
\value{
list.
}
\description{
This function performs big raster processing
It requires user to write a function expecting vectors / matrices / dataframes
corresponding to raster data read sequentially, a list of input files, a list
of output files, and it is good to go.
}
