% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_continuum_removal.R
\name{apply_continuum_removal}
\alias{apply_continuum_removal}
\title{prepares data into list to apply continuum removal}
\usage{
apply_continuum_removal(spectral_data, spectral)
}
\arguments{
\item{spectral_data}{numeric. initial data matrix (nb samples x nb bands)}

\item{spectral}{list. information about spectral bands}
}
\value{
samples from image and updated number of pixels to sample
}
\description{
prepares data into list to apply continuum removal
}
