% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_spectral_species.R
\name{apply_spectral_species}
\alias{apply_spectral_species}
\title{compute spectral species from a raster}
\usage{
apply_spectral_species(
  input_rast,
  input_mask = NULL,
  selected_bands = NULL,
  Kmeans_info,
  output_dir,
  filetype = "GTiff",
  overwrite = TRUE
)
}
\arguments{
\item{input_rast}{spatRaster. data raster}

\item{input_mask}{spatRaster. mask raster}

\item{selected_bands}{numeric. bands selected from input data}

\item{Kmeans_info}{list. kmeans description obtained from function get_kmeans}

\item{output_dir}{path where to save outputs}

\item{filetype}{character. gdal driver for output raster}

\item{overwrite}{boolean.}
}
\value{
none
}
\description{
compute spectral species from a raster
}
