% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_spectral_species_plot.R
\name{apply_spectral_species_plot}
\alias{apply_spectral_species_plot}
\title{apply biodivMapR to a set of plots identified by a field '\emph{id}'
produced with preprocS2 function get_s2_tiling}
\usage{
apply_spectral_species_plot(
  id,
  feature_dir,
  mask_dir = NULL,
  list_features,
  Kmeans_info,
  output_dir,
  selected_bands = NULL,
  filetype = "GTiff",
  overwrite = TRUE,
  p = NULL
)
}
\arguments{
\item{id}{character. ID for plot}

\item{feature_dir}{directory where spectral indices are for each plot}

\item{mask_dir}{character.}

\item{list_features}{character.}

\item{Kmeans_info}{list. kmeans description obtained from function get_kmeans}

\item{output_dir}{path where to save outputs}

\item{selected_bands}{numeric. bands selected from input data}

\item{filetype}{character. gdal driver for output raster}

\item{overwrite}{boolean.}

\item{p}{list. progressbar}
}
\value{
output_raster_full_name character. path for output raster file
}
\description{
apply biodivMapR to a set of plots identified by a field '\emph{id}'
produced with preprocS2 function get_s2_tiling
}
