% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_to_poly.R
\name{bbox_to_poly}
\alias{bbox_to_poly}
\title{converts a bbox to a polygon}
\usage{
bbox_to_poly(x, crs = 4326)
}
\arguments{
\item{x}{bbox}

\item{crs}{crs of the polygon}
}
\value{
polyg sf polygon
}
\description{
converts a bbox to a polygon
}
