% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR.R
\name{biodivMapR}
\alias{biodivMapR}
\title{computes diversity metrics from raster}
\usage{
biodivMapR(
  input_raster_path,
  input_mask_path = NULL,
  output_dir,
  window_size,
  selected_bands = NULL,
  Kmeans_info_save = NULL,
  Kmeans_info_read = NULL,
  Beta_info_save = NULL,
  Beta_info_read = NULL,
  nbCPU = 1,
  options = NULL
)
}
\arguments{
\item{input_raster_path}{character. path for the input rasters}

\item{input_mask_path}{character. path for mask file}

\item{output_dir}{character. path for the output files}

\item{window_size}{numeric. window size for square plots}

\item{selected_bands}{numeric. bands selected from input_rast}

\item{Kmeans_info_save}{character. path where to save Kmeans_info}

\item{Kmeans_info_read}{character. path where to read Kmeans_info}

\item{Beta_info_save}{character. path where to save Beta_info}

\item{Beta_info_read}{character. path where to read Beta_info}

\item{nbCPU}{numeric. Number of CPUs available}

\item{options}{list. includes options
\itemize{
\item alpha_metrics list. alpha diversity metrics: richness, shannon, simpson
\item Hill_order numeric. Hill order
\item beta_metrics list. beta diversity metrics: bray, brayturn, simpson_diss, jaccard, jaccardturn, sorensen
\item fd_metrics character. list of functional metrics
\item nb_samples_alpha numeric. max number of pixels extracted for kmeans
\item nb_samples_beta numeric. number of samples to compute beta diversity
\item nb_clusters numeric. number of clusters used in kmeans
\item nb_iter numeric. nb of iterations averaged to compute diversity indices
\item pcelim numeric. minimum proportion of pixels to consider spectral species
\item maxRows numeric. max number of rows processed once by each CPU
\item moving_window boolean. should diversity be computed on moving window?
\item min_sun numeric. minimum amount of sunlit pixels in the plots
\item dimPCoA numeric. number of dimensions of PCoA
\item progressbar boolean. set true for progress bar during clustering
\item filetype character. driver for output diversity raster data
}}
}
\value{
path for diversity_maps, Kmeans_info and Beta_info
}
\description{
computes diversity metrics from raster
}
