% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_classif.R
\name{biodivMapR_classif}
\alias{biodivMapR_classif}
\title{computes diversity metrics from a classification raster containing information
on a specific class for each pixel}
\usage{
biodivMapR_classif(
  input_raster_path,
  output_dir,
  window_size,
  input_mask_path = NULL,
  site_name = NULL,
  alpha_metrics = c("richness", "shannon", "simpson", "hill"),
  beta_metrics = c("bray", "brayturn", "simpson_diss", "sorensen", "jaccard",
    "jaccardturn"),
  pcelim = 0.02,
  compute_beta = TRUE,
  nb_samples_beta = 1000
)
}
\arguments{
\item{input_raster_path}{character. path for the input rasters}

\item{output_dir}{character. path for the output files}

\item{window_size}{numeric. window size for square plots}

\item{input_mask_path}{character. path for mask file}

\item{site_name}{character. nname of site}

\item{alpha_metrics}{list. alpha diversity metrics}

\item{beta_metrics}{list. beta diversity metrics}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{compute_beta}{boolean. set TRUE if beta is to be computed}

\item{nb_samples_beta}{numeric. number of samples to compute beta diversity}
}
\value{
diversity_maps_ground
}
\description{
computes diversity metrics from a classification raster containing information
on a specific class for each pixel
}
