% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_full.R
\name{biodivMapR_full}
\alias{biodivMapR_full}
\title{computes diversity metrics from raster}
\usage{
biodivMapR_full(
  input_raster_path,
  output_dir,
  window_size,
  maxRows = NULL,
  Kmeans_info_save = NULL,
  Kmeans_info_read = NULL,
  Beta_info_save = NULL,
  Beta_info_read = NULL,
  input_mask_path = NULL,
  nbclusters = 50,
  nbSamples_beta = 1000,
  SelectBands = NULL,
  alphametrics = "shannon",
  Hill_order = 1,
  FDmetric = NULL,
  pcelim = 0.02,
  nbCPU = 1,
  nbIter = 20,
  MinSun = 0.25,
  maxPixel_kmeans = 1e+05,
  dimPCoA = 3,
  progressbar = T,
  filetype = "GTiff"
)
}
\arguments{
\item{input_raster_path}{character. path for the input rasters}

\item{output_dir}{character. path for the output files}

\item{window_size}{numeric. window size for square plots}

\item{maxRows}{numeric. max number of rows in each block}

\item{Kmeans_info_save}{character. path where to save Kmeans_info}

\item{Kmeans_info_read}{character. path where to read Kmeans_info}

\item{Beta_info_save}{character. path where to save Beta_info}

\item{Beta_info_read}{character. path where to read Beta_info}

\item{input_mask_path}{character. path for mask file}

\item{nbclusters}{numeric. number of clusters used in kmeans}

\item{nbSamples_beta}{numeric. number of samples to compute beta diversity from}

\item{SelectBands}{numeric. bands selected from input_rast}

\item{alphametrics}{list. alpha diversity metrics: richness, shannon, simpson}

\item{Hill_order}{numeric. Hill order}

\item{FDmetric}{character. list of functional metrics}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{nbCPU}{numeric. Number of CPUs available}

\item{nbIter}{numeric. nb of iterations averaged to compute diversity indices}

\item{MinSun}{numeric. minimum amount of sunlit pixels in the plots}

\item{maxPixel_kmeans}{numeric. max number of pixels extracted for kmeans}

\item{dimPCoA}{numeric. number of dimensions of PCoA}

\item{progressbar}{boolean. set true for progress bar during clustering}

\item{filetype}{character. driver for output diversity raster data}
}
\value{
Kmeans_info and Beta_info
}
\description{
computes diversity metrics from raster
}
