% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_full_classif.R
\name{biodivMapR_full_classif}
\alias{biodivMapR_full_classif}
\title{computes diversity metrics from raster}
\usage{
biodivMapR_full_classif(
  input_raster_path,
  output_dir,
  window_size,
  input_mask_path = NULL,
  site_name = NULL,
  pcelim = 0.02,
  nb_samples_beta = 1000
)
}
\arguments{
\item{input_raster_path}{character. path for the input rasters}

\item{output_dir}{character. path for the output files}

\item{window_size}{numeric. window size for square plots}

\item{input_mask_path}{character. path for mask file}

\item{site_name}{character. nname of site}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{nb_samples_beta}{numeric. number of samples to compute beta diversity}
}
\value{
diversity_maps_ground
}
\description{
computes diversity metrics from raster
}
