% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_OptClusters.R
\name{biodivMapR_OptClusters}
\alias{biodivMapR_OptClusters}
\title{explores performances of biodivMapR for different numbers of clusters}
\usage{
biodivMapR_OptClusters(
  input_raster,
  obs_vect,
  obs2optimize,
  SelectBands,
  obs_criterion = "shannon",
  input_mask = NULL,
  outputdir = "./",
  nbclusters = 50,
  MinSun = 0.25,
  nbIter = 20,
  pcelim = 0.02,
  verbose = T,
  nbRep = 50,
  maxPixel_kmeans = 1e+05,
  Hill_order = 1,
  algorithm = "Hartigan-Wong",
  nbCPU = 1
)
}
\arguments{
\item{input_raster}{SpatRaster or list of SpatRaster}

\item{obs_vect}{SpatVector or SpatVectorCollection}

\item{obs2optimize}{numeric .list of ground obs diversity metrics
corresponding to obs_vect.
Expected values: richness, shannon, simpson, BC}

\item{SelectBands}{numeric. bnds to select from input_raster}

\item{obs_criterion}{character. richness, shannon, simpson or BC}

\item{input_mask}{SpatRaster corresponding to mask}

\item{outputdir}{character. output directory}

\item{nbclusters}{numeric.}

\item{MinSun}{numeric.}

\item{nbIter}{numeric.}

\item{pcelim}{numeric.}

\item{verbose}{boolean.}

\item{nbRep}{numeric.}

\item{maxPixel_kmeans}{numeric.}

\item{Hill_order}{numeric.}

\item{algorithm}{character.}

\item{nbCPU}{numeric.}
}
\value{
list including performances (correlation) of SFS with additional
features and assessed diversity metrics corresponding to each step
}
\description{
explores performances of biodivMapR for different numbers of clusters
}
