% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_SFS.R
\name{biodivMapR_sfs}
\alias{biodivMapR_sfs}
\title{performs SFS to identify combination of input variables maximizing a
criterion}
\usage{
biodivMapR_sfs(
  input_raster,
  obs_vect,
  obs2optimize,
  obs_criterion = "shannon",
  corrMethod = "spearman",
  input_mask = NULL,
  Hill_order = 1,
  nb_clusters = 50,
  min_sun = 0.25,
  nb_pix = 1e+05,
  nb_iter = 10,
  pcelim = 0.02,
  verbose = TRUE,
  nbWorkers = 1,
  algorithm = "Hartigan-Wong",
  nbCPU = 1
)
}
\arguments{
\item{input_raster}{SpatRaster or list of SpatRaster}

\item{obs_vect}{SpatVector or SpatVectorCollection}

\item{obs2optimize}{numeric .list of ground obs diversity metrics
corresponding to obs_vect.
Expected values: richness, shannon, simpson, hill}

\item{obs_criterion}{character. richness, shannon, simpson, hill}

\item{corrMethod}{character. select between method available for cor.test}

\item{input_mask}{SpatRaster corresponding to mask}

\item{Hill_order}{numeric.}

\item{nb_clusters}{numeric.}

\item{min_sun}{numeric.}

\item{nb_pix}{numeric.}

\item{nb_iter}{numeric.}

\item{pcelim}{numeric.}

\item{verbose}{boolean.}

\item{nbWorkers}{numeric.}

\item{algorithm}{character.}

\item{nbCPU}{numeric.}
}
\value{
list including performances (correlation) of SFS with additional
features and assessed diversity metrics corresponding to each step
}
\description{
performs SFS to identify combination of input variables maximizing a
criterion
}
