% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_tiles.R
\name{biodivMapR_tiles}
\alias{biodivMapR_tiles}
\title{computes diversity metrics from raster}
\usage{
biodivMapR_tiles(
  feature_dir,
  list_features,
  mask_dir = NULL,
  output_dir,
  window_size,
  plot_names,
  alpha_metrics = "shannon",
  Hill_order = 1,
  beta_metrics = "bray",
  fd_metrics = NULL,
  nbCPU = 1,
  pcelim = 0.02,
  maxRows = 1200,
  moving_window = FALSE,
  site_name = NULL,
  mosaic_output = TRUE
)
}
\arguments{
\item{feature_dir}{character. path where to get features}

\item{list_features}{character. list of features}

\item{mask_dir}{character. path for masks}

\item{output_dir}{character. path where to save results}

\item{window_size}{numeric. number of clusters used in kmeans}

\item{plot_names}{character. plot names}

\item{alpha_metrics}{character.}

\item{Hill_order}{numeric.}

\item{beta_metrics}{character.}

\item{fd_metrics}{character.}

\item{nbCPU}{numeric. Number of CPUs available}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{maxRows}{numeric. maximum number of rows}

\item{moving_window}{boolean. should moving window be used?}

\item{site_name}{character. name for the output files}

\item{mosaic_output}{boolean. set TRUE if outputs need to be mosaiced}
}
\value{
mosaic_path
}
\description{
computes diversity metrics from raster
}
