% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specdiv_functions.R
\name{bright_norm}
\alias{bright_norm}
\title{Brightness normalization -----}
\usage{
bright_norm(x)
}
\arguments{
\item{x}{numeric. spectrum to be normalized}
}
\value{
normalized spectrum
}
\description{
Brightness normalization -----
}
