% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specdiv_functions.R
\name{bright_norm_df}
\alias{bright_norm_df}
\title{Brightness normalization for data frame ----}
\usage{
bright_norm_df(x)
}
\arguments{
\item{x}{dataframe. spectrum to be normalized}
}
\value{
dataframe including normalized spectrum
}
\description{
Brightness normalization for data frame ----
}
