% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Build.Image.From.List}
\alias{Build.Image.From.List}
\title{rebuild full image from list of subsets}
\usage{
Build.Image.From.List(Image.Subsets, ipix, jpix, nbBands)
}
\arguments{
\item{Image.Subsets}{subsets of an image. list expected}

\item{ipix}{nb of lines for the full image}

\item{jpix}{nb of columns for the full image}

\item{nbBands}{nb of bands for the full image & subsets}
}
\value{
image full size in 3 dimensions
}
\description{
rebuild full image from list of subsets
}
