% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{build_image_from_list}
\alias{build_image_from_list}
\title{rebuilds full image from list of subsets}
\usage{
build_image_from_list(Image_Subsets, ipix, jpix, nbBands)
}
\arguments{
\item{Image_Subsets}{list. subsets of an image}

\item{ipix}{numeric. nb of lines for the full image}

\item{jpix}{numeric. nb of columns for the full image}

\item{nbBands}{numeric. nb of bands for the full image & subsets}
}
\value{
Image numeric. full size in 3 dimensions
}
\description{
rebuilds full image from list of subsets
}
