% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Center.Reduce}
\alias{Center.Reduce}
\title{center and reduce data matrix based on known mean and SD}
\usage{
Center.Reduce(X, m, sig)
}
\arguments{
\item{X}{data matrix (each column is centered/reduced)}

\item{m}{mean of each variable in the data matrix}

\item{sig}{SD of each variable in the data matrix}
}
\value{
Centered matrix
}
\description{
center and reduce data matrix based on known mean and SD
}
