% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Change.Resolution.HDR}
\alias{Change.Resolution.HDR}
\title{change resolution in a HDR file}
\usage{
Change.Resolution.HDR(HDR, Spatial.Unit)
}
\arguments{
\item{HDR}{information read from a header file}

\item{Spatial.Unit}{multiplying factor for initial resolution}
}
\value{
updated HDR information
}
\description{
change resolution in a HDR file
}
