% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{change_resolution_HDR}
\alias{change_resolution_HDR}
\title{change resolution in a HDR file}
\usage{
change_resolution_HDR(HDR, window_size)
}
\arguments{
\item{HDR}{information read from a header file}

\item{window_size}{multiplying factor for initial resolution}
}
\value{
updated HDR information
}
\description{
change resolution in a HDR file
}
