% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_CheckConvertData.R
\name{Check.Data.Format}
\alias{Check.Data.Format}
\title{Checks if the data to be processed has the format type expected}
\usage{
Check.Data.Format(Raster.Path, Mask = FALSE)
}
\arguments{
\item{Raster.Path}{full path for the raster to be converted}

\item{Mask}{is the raster a mask?}
}
\value{
Output.Path path for the image converted into ENVI BIL format
}
\description{
Checks if the data to be processed has the format type expected
}
