% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Checks if the data to be processed has the format type expected}
\usage{
check_data(input_data, arguments)
}
\arguments{
\item{input_data}{character. which check?}

\item{arguments}{list. additional arguments}
}
\value{
None
}
\description{
Checks if the data to be processed has the format type expected
}
