% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{check_update_mask_format}
\alias{check_update_mask_format}
\title{check if the format of the mask is as expected: integer, coded in Bytes, same
dimensions as input image}
\usage{
check_update_mask_format(Input_Mask_File, Input_Image_File)
}
\arguments{
\item{Input_Mask_File}{character. path for original mask}

\item{Input_Image_File}{character. path for image file corresponding to the mask}
}
\value{
MaskPath_Update
}
\description{
check if the format of the mask is as expected: integer, coded in Bytes, same
dimensions as input image
}
