% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_NAsInf.R
\name{clean_NAsInf}
\alias{clean_NAsInf}
\title{cleans dataframe from NAs and Inf values}
\usage{
clean_NAsInf(df)
}
\arguments{
\item{df}{dataframe. df to be cleaned}
}
\value{
df dataframe.
}
\description{
cleans dataframe from NAs and Inf values
}
