% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{Compute.Alpha.Diversity}
\alias{Compute.Alpha.Diversity}
\title{Map alpha diversity metrics based on spectral species}
\usage{
Compute.Alpha.Diversity(Spectral.Species.Path, Spatial.Unit, nbclusters,
  MinSun, pcelim, nbCPU = FALSE, MaxRAM = FALSE,
  Index.Alpha = "Shannon")
}
\arguments{
\item{Spectral.Species.Path}{path for spectral species file to be written}

\item{Spatial.Unit}{size of spatial units (in pixels) to compute diversity}

\item{nbclusters}{number of clusters defined in k-Means}

\item{MinSun}{minimum proportion of sunlit pixels required to consider plot}
}
\value{
list of mean and SD of alpha diversity metrics
}
\description{
Map alpha diversity metrics based on spectral species
}
