% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{Compute.BCdiss}
\alias{Compute.BCdiss}
\title{compute the bray curtis dissimilarity matrix corresponding to a list of kernels
(rows) defined by their spectral species (columns)
SSDList is a list containing spectral species distribution for two sets of kernels ([[1]] and [[2]])
pcelim is the threshold for minimum contributin of a spctral species to be kept}
\usage{
Compute.BCdiss(SSDList, pcelim)
}
\arguments{
\item{SSDList}{list of 2 groups to compute BC dissimilarity from}

\item{pcelim}{minimum proportion required for a species to be included (currently deactivated)}
}
\value{
MatBC matrix of bray curtis dissimilarity matrix
}
\description{
compute the bray curtis dissimilarity matrix corresponding to a list of kernels
(rows) defined by their spectral species (columns)
SSDList is a list containing spectral species distribution for two sets of kernels ([[1]] and [[2]])
pcelim is the threshold for minimum contributin of a spctral species to be kept
}
