% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{Compute.BetaDiversity}
\alias{Compute.BetaDiversity}
\title{computes beta diversity}
\usage{
Compute.BetaDiversity(Output.Dir, MinSun, Nb.Units.Ordin, NbIter,
  nbclusters, pcelim, scaling = "PCO", nbCPU = FALSE, MaxRAM = FALSE)
}
\arguments{
\item{Output.Dir}{directory where spectral species are stored}

\item{MinSun}{minimum proportion of sunlit pixels required to consider plot}

\item{Nb.Units.Ordin}{maximum number of spatial units to be processed in Ordination}

\item{NbIter}{number of iterations}

\item{nbclusters}{number of clusters defined in k-Means}

\item{pcelim}{min proprtion for a spectral species in a spatial unit to be considerd}
}
\description{
computes beta diversity
}
