% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{Compute.NMDS}
\alias{Compute.NMDS}
\title{computes NMDS}
\usage{
Compute.NMDS(MatBCdist)
}
\arguments{
\item{MatBCdist}{BC dissimilarity matrix}
}
\value{
BetaNMDS.sel
}
\description{
computes NMDS
}
