% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{Compute.NN.From.Ordination}
\alias{Compute.NN.From.Ordination}
\title{compute the nearest neighbors among kernels used in NMDS}
\usage{
Compute.NN.From.Ordination(MatBC3, knn, BetaDiversity0)
}
\arguments{
\item{MatBC3}{matrix of BC dissimilarity between the kernels excluded from Ordination (rows)}

\item{knn}{number of neighbors}

\item{BetaDiversity0}{}
}
\value{
estimated NMDS position based on nearest neighbors from NMDS
}
\description{
compute the nearest neighbors among kernels used in NMDS
}
