% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapSpectralSpecies.R
\name{Compute.Spectral.Species}
\alias{Compute.Spectral.Species}
\title{this function reads PCA file and defines the spectral species for each pixel
based on the set of cluster centroids defined for each iteration
applies kmeans --> closest cluster corresponds to the "spectral species"}
\usage{
Compute.Spectral.Species(PCA.Path, ImPathShade, Spectral.Species.Path,
  Location.RW, PC.Select, Kmeans.info, nbCPU = 1)
}
\arguments{
\item{PCA.Path}{path for the PCA image}

\item{ImPathShade}{shade mask}

\item{Spectral.Species.Path}{path for spectral species file to be written}

\item{Location.RW}{where to read/write information in binary file}

\item{PC.Select}{PCs selected from PCA}

\item{Kmeans.info}{information about kmeans computed in previous step}
}
\description{
this function reads PCA file and defines the spectral species for each pixel
based on the set of cluster centroids defined for each iteration
applies kmeans --> closest cluster corresponds to the "spectral species"
}
