% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{Compute.SSD}
\alias{Compute.SSD}
\title{compute spectral species distribution from original spectral species map}
\usage{
Compute.SSD(Image.Chunk, Spatial.Unit, nbclusters, MinSun, pcelim,
  Index.Alpha = "Shannon")
}
\arguments{
\item{Image.Chunk}{3D image chunk of spectral species}

\item{Spatial.Unit}{size of spatial units (in pixels) to compute diversity}

\item{nbclusters}{number of clusters defined in k-Means}

\item{MinSun}{minimum proportion of sunlit pixels required to consider plot}

\item{pcelim}{minimum proportion for a spectral species to be included}
}
\value{
list of alpha diversity metrics for each iteration
}
\description{
compute spectral species distribution from original spectral species map
}
