% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{compute_ALPHA_FromPlot}
\alias{compute_ALPHA_FromPlot}
\title{compute alpha diversity from spectral species computed for a plot
expecting a matrix of spectral species (n pixels x p repetitions)}
\usage{
compute_ALPHA_FromPlot(SpectralSpecies_Plot, pcelim = 0.02)
}
\arguments{
\item{SpectralSpecies_Plot}{numeric. matrix of spectral species}

\item{pcelim}{minimum contribution of spectral species to estimation of diversity
each spectral species with a proprtion < pcelim is eliminated before computation of diversity}
}
\value{
list of alpha diversity metrics
}
\description{
compute alpha diversity from spectral species computed for a plot
expecting a matrix of spectral species (n pixels x p repetitions)
}
