% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{compute_alpha_metrics}
\alias{compute_alpha_metrics}
\title{Computes alpha diversity metrics based on spectral species}
\usage{
compute_alpha_metrics(
  Spectral_Species_Path,
  window_size,
  nbclusters,
  MinSun = 0.25,
  pcelim = 0.02,
  nbCPU = 1,
  MaxRAM = 0.25,
  Index_Alpha = "Shannon"
)
}
\arguments{
\item{Spectral_Species_Path}{character. path for spectral species file to be written}

\item{window_size}{numeric. size of spatial units (in pixels) to compute diversity}

\item{nbclusters}{numeric. number of clusters defined in k-Means}

\item{MinSun}{numeric. minimum proportion of sunlit pixels required to consider plot}

\item{pcelim}{numeric. percentage of occurence of a cluster below which cluster is eliminated}

\item{nbCPU}{numeric. number of CPUs available}

\item{MaxRAM}{numeric. maximum RAM available}

\item{Index_Alpha}{list. list of alpha diversity indices to be computed from spectral species}
}
\value{
list of mean and SD of alpha diversity metrics
}
\description{
Computes alpha diversity metrics based on spectral species
}
