% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{Compute_ALPHA_per_window}
\alias{Compute_ALPHA_per_window}
\title{computes alpha diversity metrics and update spectral species distribution for
a window defined in biodivmapR}
\usage{
Compute_ALPHA_per_window(
  SSD,
  nbPix_Sunlit,
  alphaIdx,
  nbclusters,
  pcelim = 0.02
)
}
\arguments{
\item{SSD}{numeric.}

\item{nbPix_Sunlit}{numeric.}

\item{alphaIdx}{list. list of diversity metrics to be computed}

\item{nbclusters}{numeric. number of clusters defined in k-Means}

\item{pcelim}{minimum proportion for a spectral species to be included}
}
\value{
shannon, Simpson, Fisher, SSDMap
}
\description{
computes alpha diversity metrics and update spectral species distribution for
a window defined in biodivmapR
}
