% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{Compute_ALPHA_SSD_per_window}
\alias{Compute_ALPHA_SSD_per_window}
\title{computes alpha diversity metrics and spectral species distribution for a
window (individual spatial unit) defined in biodivmapR}
\usage{
Compute_ALPHA_SSD_per_window(
  listAlpha,
  nb_partitions,
  nbclusters,
  alphaIdx,
  MinSun = 0.25,
  pcelim = 0.02
)
}
\arguments{
\item{listAlpha}{list.}

\item{nb_partitions}{numeric. number of iterations performed with biodivMapR}

\item{nbclusters}{numeric. number of clusters defined in k-Means}

\item{alphaIdx}{list.}

\item{MinSun}{minimum proportion of sunlit pixels required to consider plot}

\item{pcelim}{minimum proportion for a spectral species to be included}
}
\value{
res list of alpha metrics and spectral species distribution
corresponding to the window
}
\description{
computes alpha diversity metrics and spectral species distribution for a
window (individual spatial unit) defined in biodivmapR
}
