% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{Compute_ALPHA_SSD_per_window_list}
\alias{Compute_ALPHA_SSD_per_window_list}
\title{computes alpha diversity metrics and spectral species distribution for a list
of windows (individual spatial unit) defined in biodivmapR}
\usage{
Compute_ALPHA_SSD_per_window_list(
  listAlpha,
  nb_partitions,
  nbclusters,
  alphaIdx,
  MinSun = 0.25,
  pcelim = 0.02,
  p = NULL
)
}
\arguments{
\item{listAlpha}{list.}

\item{nb_partitions}{numeric. number of iterations performed with biodivMapR}

\item{nbclusters}{numeric. number of clusters defined in k-Means}

\item{alphaIdx}{list.}

\item{MinSun}{minimum proportion of sunlit pixels required to consider plot}

\item{pcelim}{minimum proportion for a spectral species to be included}

\item{p}{list. progressor object for progress bar}
}
\value{
list_assd list of alpha metrics and spectral species distribution
corresponding to the list of windows
}
\description{
computes alpha diversity metrics and spectral species distribution for a list
of windows (individual spatial unit) defined in biodivmapR
}
