% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bc_diss.R
\name{compute_bc_diss}
\alias{compute_bc_diss}
\title{compute bray curtis dissimilarity matrix corresponding to a list of kernels
(rows) defined by their spectral species (columns)
ssd_list is a list containing spectral species distribution for two sets of kernels
pcelim is the threshold for minimum contribution of a spectral species to be kept}
\usage{
compute_bc_diss(ssd_list, pcelim = 0.02)
}
\arguments{
\item{ssd_list}{list. list of 2 groups to compute BC dissimilarity from}

\item{pcelim}{numeric. minimum proportion required for a species to be included}
}
\value{
mat_bc matrix of bray curtis dissimilarity matrix
}
\description{
compute bray curtis dissimilarity matrix corresponding to a list of kernels
(rows) defined by their spectral species (columns)
ssd_list is a list containing spectral species distribution for two sets of kernels
pcelim is the threshold for minimum contribution of a spectral species to be kept
}
