% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_BCdiss.R
\name{compute_BCdiss}
\alias{compute_BCdiss}
\title{compute bray curtis dissimilarity matrix corresponding to a list of kernels
(rows) defined by their spectral species (columns)
SSDList is a list containing spectral species distribution for two sets of kernels
pcelim is the threshold for minimum contributin of a spctral species to be kept}
\usage{
compute_BCdiss(SSDList, pcelim = 0.02)
}
\arguments{
\item{SSDList}{list. list of 2 groups to compute BC dissimilarity from}

\item{pcelim}{numeric. minimum proportion required for a species to be included}
}
\value{
MatBC matrix of bray curtis dissimilarity matrix
}
\description{
compute bray curtis dissimilarity matrix corresponding to a list of kernels
(rows) defined by their spectral species (columns)
SSDList is a list containing spectral species distribution for two sets of kernels
pcelim is the threshold for minimum contributin of a spctral species to be kept
}
