% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{compute_BETA_FromPlots}
\alias{compute_BETA_FromPlots}
\title{compute beta diversity from spectral species computed for a plot
expecting a matrix of spectral species (n pixels x p repetitions)}
\usage{
compute_BETA_FromPlots(SpectralSpecies_Plots, nbclusters, pcelim = 0.02)
}
\arguments{
\item{SpectralSpecies_Plots}{list. list of matrices of spectral species corresponding to plots}

\item{nbclusters}{numeric. number of clusters}

\item{pcelim}{numeric. Minimum contribution (in %) required for a spectral species
each spectral species with a proprtion < pcelim is eliminated before computation of diversity}
}
\value{
Mean bray curtis dissimilarity matrix for all plots, and individual BC matrices corresponding to each repetitions
}
\description{
compute beta diversity from spectral species computed for a plot
expecting a matrix of spectral species (n pixels x p repetitions)
}
