% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{compute_beta_metrics}
\alias{compute_beta_metrics}
\title{computes beta diversity metrics}
\usage{
compute_beta_metrics(
  ClusterMap_Path,
  MinSun,
  Nb_Units_Ordin,
  nb_partitions,
  nbclusters,
  pcelim,
  scaling = "PCO",
  nbCPU = FALSE,
  MaxRAM = FALSE,
  dimMDS = 3
)
}
\arguments{
\item{ClusterMap_Path}{character. File containing spectral species or classes from prior classification}

\item{MinSun}{numeric. minimum proportion of sunlit pixels required to consider plot}

\item{Nb_Units_Ordin}{numeric. maximum number of spatial units to be processed in Ordination}

\item{nb_partitions}{numeric. number of iterations}

\item{nbclusters}{numeric. number of clusters defined in k-Means}

\item{pcelim}{numeric. Minimum contribution in percents required for a spectral species}

\item{scaling}{character. scaling method, PCO or NMDS}

\item{nbCPU}{numeric. Number of CPUs to use in parallel.}

\item{MaxRAM}{numeric. MaxRAM maximum size of chunk in GB to limit RAM allocation when reading image file.}

\item{dimMDS}{numeric. number of dimensions for the scaling.}
}
\description{
computes beta diversity metrics
}
