% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dissimilarity.R
\name{compute_dissimilarity}
\alias{compute_dissimilarity}
\title{compute dissimilarity matrix corresponding to two distinct matrices, each of
them including a list of samples (rows) defined by their spectral species (columns)
metrics available: bray, brayturn, simpson_diss, jaccard, jaccardturn, sorensen}
\usage{
compute_dissimilarity(A, B, beta_metrics)
}
\arguments{
\item{A}{A numeric matrix of size N x p.}

\item{B}{A numeric matrix of size M x p.}

\item{beta_metrics}{list. beta diversity metrics}
}
\value{
mat_diss dissimilarity matrix
}
\description{
compute dissimilarity matrix corresponding to two distinct matrices, each of
them including a list of samples (rows) defined by their spectral species (columns)
metrics available: bray, brayturn, simpson_diss, jaccard, jaccardturn, sorensen
}
