% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapFunctionalDiversity.R
\name{compute_FD}
\alias{compute_FD}
\title{compute functional diversity metrics for an array, given a specific window size}
\usage{
compute_FD(
  Image_Chunk,
  window_size,
  MinSun,
  FDmetric = c("FRic", "FEve", "FDiv"),
  nbCPU
)
}
\arguments{
\item{Image_Chunk}{numeric. 3D image chunk of spectral species}

\item{window_size}{numeric. size of spatial units (in pixels) to compute diversity}

\item{MinSun}{numeric. minimum proportion of sunlit pixels required to consider plot}

\item{FDmetric}{character. Functional diversity metric}

\item{nbCPU}{numeric. nb of CPU to process data}
}
\value{
list of functional diversity metrics corresponding to image chunk
}
\description{
compute functional diversity metrics for an array, given a specific window size
}
