% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapFunctionalDiversity.R
\name{compute_Functional_metrics}
\alias{compute_Functional_metrics}
\title{Map functional diversity metrics based on spectral species}
\usage{
compute_Functional_metrics(
  Functional_File,
  Functional_Map_Path,
  Selected_Features,
  window_size,
  MinSun,
  nbCPU = 1,
  MaxRAM = 0.25
)
}
\arguments{
\item{Functional_File}{character. Path and name of the image processed to be used to compute functional diversity}

\item{Functional_Map_Path}{character. Path and name of the resulting functional diversity raster}

\item{Selected_Features}{numeric. list of features from Functional_File to be included in the analysis}

\item{window_size}{numeric. size of spatial units (in pixels) to compute diversity}

\item{MinSun}{numeric. minimum proportion of sunlit pixels required to consider plot}

\item{nbCPU}{numeric. Number of CPUs to use in parallel.}

\item{MaxRAM}{numeric. MaxRAM maximum size of chunk in GB to limit RAM allocation when reading image file.}
}
\value{
list of mean and SD of alpha diversity metrics
}
\description{
Map functional diversity metrics based on spectral species
}
