% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{compute_NMDS}
\alias{compute_NMDS}
\title{computes NMDS}
\usage{
compute_NMDS(MatBCdist, dimMDS = 3)
}
\arguments{
\item{MatBCdist}{BC dissimilarity matrix}

\item{dimMDS}{numeric. number of dimensions of the NMDS}
}
\value{
BetaNMDS_sel
}
\description{
computes NMDS
}
