% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nn_from_ordination.R
\name{compute_nn_from_ordination}
\alias{compute_nn_from_ordination}
\title{compute the nearest neighbors among kernels}
\usage{
compute_nn_from_ordination(mat_bc, knn, pcoa_train)
}
\arguments{
\item{mat_bc}{matrix of BC dissimilarity between the kernels excluded from Ordination (rows)}

\item{knn}{numeric. number of neighbors}

\item{pcoa_train}{numeric. PCoA coordinates computed from dissimilarity matrix}
}
\value{
ordin_est estimated PCoA position based on nearest neighbors from
PCoA space defined by training set
}
\description{
compute the nearest neighbors among kernels
}
