% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapSpectralSpecies.R
\name{compute_spectral_species}
\alias{compute_spectral_species}
\title{this function reads PCA file and defines the spectral species for each pixel
based on the set of cluster centroids defined for each iteration
applies kmeans --> closest cluster corresponds to the "spectral species"}
\usage{
compute_spectral_species(
  PCA_Path,
  Input_Mask_File,
  Spectral_Species_Path,
  Location_RW,
  PC_Select,
  Kmeans_info,
  nbCPU = 1
)
}
\arguments{
\item{PCA_Path}{character. path for the PCA image}

\item{Input_Mask_File}{character. Path for the mask}

\item{Spectral_Species_Path}{character. path for spectral species file to be written}

\item{Location_RW}{numeric. where to read/write information in binary file}

\item{PC_Select}{numeric. PCs selected from PCA}

\item{Kmeans_info}{list. information about kmeans computed in previous step}

\item{nbCPU}{numeric. number of CPUs available}
}
\value{
None
}
\description{
this function reads PCA file and defines the spectral species for each pixel
based on the set of cluster centroids defined for each iteration
applies kmeans --> closest cluster corresponds to the "spectral species"
}
