% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapSpectralSpecies.R
\name{compute_spectral_species_FieldPlots}
\alias{compute_spectral_species_FieldPlots}
\title{compute spectral species for a subset of pixels provided in a list, each element
corresponding to a polygon}
\usage{
compute_spectral_species_FieldPlots(
  subset_Raster,
  List_FieldPlot,
  nb_partitions,
  Pix_Per_Partition,
  nbclusters,
  PC_Select = NULL,
  nbCPU = 1,
  progressbar = FALSE,
  algorithm = "Hartigan-Wong"
)
}
\arguments{
\item{subset_Raster}{numeric. Subset of a raster file on which computation of spectral species sould be performed}

\item{List_FieldPlot}{list. list of information from same file as subset, corresponding to field plots}

\item{nb_partitions}{numeric. number of repetitions of kmeans}

\item{Pix_Per_Partition}{numeric. number of pixels per partition}

\item{nbclusters}{numeric. number of clusters / spectral species}

\item{PC_Select}{numeric. selection of components from subset_Raster and List_FieldPlot on which spctral species are computed}

\item{nbCPU}{numeric. number of CPU on which kmeans is computed}

\item{progressbar}{boolean. should progress bar be displayed (set to TRUE only if no conflict of parallel process)}

\item{algorithm}{character. algorithm used in the kmeans clustering}
}
\value{
list. vector_coordinates and vector_ID for each element in the vector file
}
\description{
compute spectral species for a subset of pixels provided in a list, each element
corresponding to a polygon
}
