% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{compute_SSD}
\alias{compute_SSD}
\title{compute spectral species distribution from original spectral species map}
\usage{
compute_SSD(
  Image_Chunk,
  window_size,
  nbclusters,
  MinSun = 0.25,
  pcelim = 0.02,
  Index_Alpha = "Shannon",
  nbCPU = 1,
  nbPieces = 1
)
}
\arguments{
\item{Image_Chunk}{numeric. 3D image chunk of spectral species}

\item{window_size}{numeric. size of spatial units (in pixels) to compute diversity}

\item{nbclusters}{number of clusters defined in k-Means}

\item{MinSun}{minimum proportion of sunlit pixels required to consider plot}

\item{pcelim}{minimum proportion for a spectral species to be included}

\item{Index_Alpha}{list. list of alpha diversity indices to be computed from spectral species}

\item{nbCPU}{numeric. Number of CPUs to use in parallel.}

\item{nbPieces}{numeric. number of pieces in which original image is split}
}
\value{
list of alpha diversity metrics for each iteration
}
\description{
compute spectral species distribution from original spectral species map
}
