% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuum_removal.R
\name{continuum_removal}
\alias{continuum_removal}
\title{Computes continuum removal for matrix shaped data: more efficient than
processing individual spectra
the convex hull is based on the computation of the derivative between R at a
given spectral band and R at the following bands}
\usage{
continuum_removal(mat_init, spectral_bands, p = NULL)
}
\arguments{
\item{mat_init}{numeric. initial data matrix (nb samples x nb bands)}

\item{spectral_bands}{numeric. central wavelength for the spectral bands}

\item{p}{list. progressor object for progress bar}
}
\value{
samples from image and updated number of pixels to sampel if necessary
}
\description{
Computes continuum removal for matrix shaped data: more efficient than
processing individual spectra
the convex hull is based on the computation of the derivative between R at a
given spectral band and R at the following bands
}
