% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ContinuumRemoval.R
\name{ContinuumRemoval}
\alias{ContinuumRemoval}
\title{Computes continuum removal for matrix shaped data: more efficient than
processing individual spectra
the convex hull is based on the computation of the derivative between R at a
given spectral band and R at the following bands}
\usage{
ContinuumRemoval(Minit, Spectral.Bands)
}
\arguments{
\item{Minit}{initial data matrix (nb samples x nb bands)}

\item{Spectral.Bands}{information about spectral bands}
}
\value{
samples from image and updated number of pixels to sampel if necessary
}
\description{
Computes continuum removal for matrix shaped data: more efficient than
processing individual spectra
the convex hull is based on the computation of the derivative between R at a
given spectral band and R at the following bands
}
