% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{Convert.PCA.to.SSD}
\alias{Convert.PCA.to.SSD}
\title{Convert PCA into SSD based on previous clustering}
\usage{
Convert.PCA.to.SSD(ReadWrite, Spectral.Species.Path, HDR.SS, HDR.SSD,
  SS.Format, SSD.Format, ImgFormat, Spatial.Unit, nbclusters, MinSun,
  pcelim, Index.Alpha, SSD.Path, Sunlit.Path, Sunlit.Format)
}
\arguments{
\item{Sunlit.Format}{}
}
\description{
Convert PCA into SSD based on previous clustering
}
