% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_CheckConvertData.R
\name{Convert.Raster2BIL}
\alias{Convert.Raster2BIL}
\title{converts a raster into BIL format as expected by DivMapping codes}
\usage{
Convert.Raster2BIL(Raster.Path, Sensor = "unknown",
  Output.Directory = FALSE, Convert.Integer = TRUE,
  Multiplying.Factor = 1, Multiplying.Factor.Last = 1)
}
\arguments{
\item{Raster.Path}{full path for the raster to be converted}

\item{Sensor}{name of the sensor}

\item{Convert.Integer}{should data be converted into integer ?}

\item{Multiplying.Factor}{multiplying factor (eg convert real reflectance values between 0 and 1 into integer between 0 and 10000)}

\item{Multiplying.Factor.Last}{multiplying factor for last band}
}
\value{
Output.Path path for the image converted into ENVI BIL format
}
\description{
converts a raster into BIL format as expected by DivMapping codes
}
