% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{convert_PCA_to_SSD}
\alias{convert_PCA_to_SSD}
\title{Convert PCA into SSD based on previous clustering}
\usage{
convert_PCA_to_SSD(
  ReadWrite,
  Spectral_Species_Path,
  HDR_SS,
  HDR_SSD,
  SS_Format,
  SSD_Format,
  ImgFormat,
  window_size,
  nbclusters,
  MinSun = 0.25,
  pcelim = 0.02,
  Index_Alpha,
  SSD_Path,
  Sunlit_Path,
  Sunlit_Format,
  p = NULL,
  nbCPU = 1,
  nbPieces = 1
)
}
\arguments{
\item{ReadWrite}{list. includes byte coordinates where to read and write from files}

\item{Spectral_Species_Path}{character. path for Spectral_Species raster file}

\item{HDR_SS}{list. HDR file for spectral species}

\item{HDR_SSD}{list. HDR file for spectral species distribution}

\item{SS_Format}{character. format of data corresponding to spectral species file}

\item{SSD_Format}{character. format of data corresponding to spectral species distribution file}

\item{ImgFormat}{character. image format}

\item{window_size}{numeric. window size}

\item{nbclusters}{numeric. number of clusters}

\item{MinSun}{numeric. Minimum proportion of sunlit pixels required to consider plot.}

\item{pcelim}{numeric. Minimum contribution (in \%) required for a spectral species.}

\item{Index_Alpha}{list. list of spectral metrics to be computed}

\item{SSD_Path}{character. path for spectral species distribution file}

\item{Sunlit_Path}{character. path for sunlit proportion file}

\item{Sunlit_Format}{list. format of raster}

\item{p}{list. progressor object for progress bar}

\item{nbCPU}{numeric. Number of CPUs to use in parallel.}

\item{nbPieces}{numeric. number of pieces to split file read into}
}
\description{
Convert PCA into SSD based on previous clustering
}
