% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PerformPCA.R
\name{coordPix_kernel}
\alias{coordPix_kernel}
\title{XXX}
\usage{
coordPix_kernel(X, kernel)
}
\arguments{
\item{X}{numeric.}

\item{kernel}{numeric.}
}
\value{
coordinates of the pixels
}
\description{
XXX
}
