% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specdiv_functions.R
\name{count_pixels}
\alias{count_pixels}
\title{Count masked/missing pixels -----}
\usage{
count_pixels(cube, fact = 40)
}
\arguments{
\item{cube}{dataframe. data cube}

\item{fact}{numeric. window size}
}
\value{
cube_count
}
\description{
Count masked/missing pixels -----
}
