% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_FilterData.R
\name{Create.Mask.Optical}
\alias{Create.Mask.Optical}
\title{create a mask based on NDVI, Green reflectance and NIR reflectance
NDVI (min) threshold eliminates non vegetated pixels
Blue (max) threshold eliminates Clouds
NIR (min) threshold eliminates shadows
! only valid if Optical data!!}
\usage{
Create.Mask.Optical(ImPath, ImPathShade, ImPathShade.Update, NDVI.Thresh,
  Blue.Thresh, NIR.Thresh, Blue = 480, Red = 700, NIR = 835)
}
\arguments{
\item{ImPath}{full path of a raster file}

\item{ImPathShade}{full path of the raster mask corresponding to the raster file}

\item{ImPathShade.Update}{wavelength (nm) of the spectral bands to be found}

\item{NDVI.Thresh}{NDVI threshold applied to produce a mask (select pixels with NDVI>NDVI.Thresh)}

\item{Blue.Thresh}{Blue threshold applied to produce a mask (select pixels with Blue refl < Blue.Thresh --> filter clouds) refl expected between 0 and 10000}

\item{NIR.Thresh}{NIR threshold applied to produce a mask (select pixels with NIR refl < NIR.Thresh) refl expected between 0 and 10000}
}
\value{
ImPathShade path for the updated shademask produced
}
\description{
create a mask based on NDVI, Green reflectance and NIR reflectance
NDVI (min) threshold eliminates non vegetated pixels
Blue (max) threshold eliminates Clouds
NIR (min) threshold eliminates shadows
! only valid if Optical data!!
}
