% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PerformPCA.R
\name{Create.PCA.Image}
\alias{Create.PCA.Image}
\title{writes an ENVI image corresponding to PCA}
\usage{
Create.PCA.Image(ImPath, ImPathShade, PCA.Path, PCA.model, Spectral,
  Nb.PCs, CR, TypePCA, nbCPU = 1, MaxRAM = 0.25)
}
\arguments{
\item{ImPath}{path for the raster on which PCA is applied}

\item{ImPathShade}{path for the corresponding shade mask}

\item{PCA.Path}{path for resulting PCA}

\item{PCA.model}{PCA model description}

\item{Spectral}{spectral information to be used in the image}

\item{Nb.PCs}{number of components kept in the resulting PCA raster}

\item{CR}{Shoudl continuum removal be performed?}

\item{TypePCA}{PCA, SPCA, NLPCA}

\item{nbCPU}{number of CPUs to process data}

\item{MaxRAM}{max RAM when initial image is read (in Gb)}
}
\description{
writes an ENVI image corresponding to PCA
}
