% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R,
%   R/Lib_Validation_DiversityMappR.R
\name{Create.Shademask}
\alias{Create.Shademask}
\title{create a binary mask file based on a matrix and the corresponding header}
\usage{
Create.Shademask(Mask, HDR, Path.Mask)

Create.Shademask(Mask, HDR, Path.Mask)
}
\arguments{
\item{Mask}{matrix of a binary mask}

\item{HDR}{header corresponding to an image to be masked}

\item{Path.Mask}{path for the mask}

\item{Mask}{matrix of a binary mask}

\item{HDR}{header corresponding to an image to be masked}

\item{Path.Mask}{path for the mask}
}
\description{
create a binary mask file based on a matrix and the corresponding header

create a binary mask file based on a matrix and the corresponding header
}
