% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{create_hdr}
\alias{create_hdr}
\title{create a hdr file for a raster}
\usage{
create_hdr(
  ImPath,
  Sensor = "unknown",
  SpectralBands = NULL,
  BandName = NULL,
  WLunits = NULL
)
}
\arguments{
\item{ImPath}{character. path of the raster image}

\item{Sensor}{character. name of the sensor}

\item{SpectralBands}{numeric. vector of spectral bands for the raster}

\item{BandName}{character. name for each band in the raster}

\item{WLunits}{character. wavelengths unit. 'Micrometers' or 'Nanometers'}
}
\value{
corresponding hdr
}
\description{
create a hdr file for a raster
}
