% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_FilterData.R
\name{create_mask_from_threshold}
\alias{create_mask_from_threshold}
\title{create a mask based on NDVI, Green reflectance and NIR reflectance
NDVI (min) threshold eliminates non vegetated pixels
Blue (max) threshold eliminates Clouds
NIR (min) threshold eliminates shadows
! only valid if Optical data!!}
\usage{
create_mask_from_threshold(
  ImPath,
  MaskPath,
  MaskPath_Update,
  NDVI_Thresh,
  Blue_Thresh,
  NIR_Thresh,
  Blue = 480,
  Red = 690,
  NIR = 835
)
}
\arguments{
\item{ImPath}{character. Full path of a raster file}

\item{MaskPath}{character. Full path of the mask to be used with the raster file}

\item{MaskPath_Update}{character. Full path of the updated mask to be used with the raster file}

\item{NDVI_Thresh}{numeric. NDVI threshold applied to produce a mask (select pixels with NDVI>NDVI_Thresh)}

\item{Blue_Thresh}{numeric. Blue threshold applied to produce a mask (select pixels with Blue refl < Blue_Thresh --> filter clouds) refl expected between 0 and 10000}

\item{NIR_Thresh}{numeric. NIR threshold applied to produce a mask (select pixels with NIR refl < NIR_Thresh) refl expected between 0 and 10000}

\item{Blue}{numeric. spectral band corresponding to the blue channel (in nanometers)}

\item{Red}{numeric. spectral band corresponding to the red channel (in nanometers)}

\item{NIR}{numeric. spectral band corresponding to the NIR channel (in nanometers)}
}
\description{
create a mask based on NDVI, Green reflectance and NIR reflectance
NDVI (min) threshold eliminates non vegetated pixels
Blue (max) threshold eliminates Clouds
NIR (min) threshold eliminates shadows
! only valid if Optical data!!
}
