% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crfun.R
\name{crfun}
\alias{crfun}
\title{computes continuum removal for individual spectra}
\usage{
crfun(x, wav, interpol)
}
\arguments{
\item{x}{numeric. original spectrum}

\item{wav}{numeric. central wavelength for the spectral bands}

\item{interpol}{character. linear or spline}
}
\value{
continuum removed spectrum
}
\description{
adapted from propsectr function
https://github.com/l-ramirez-lopez/prospectr/blob/main/R/continuumRemoval.R
}
