% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Define.Output.Dir}
\alias{Define.Output.Dir}
\title{define output directory and create it if necessary}
\usage{
Define.Output.Dir(Output.Dir, ImPath, TypePCA)
}
\arguments{
\item{Output.Dir}{output directory}

\item{ImPath}{image path}

\item{TypePCA}{Type of PCA (PCA, SPCA, NLPCA...)}
}
\value{
path of the output directory
}
\description{
define output directory and create it if necessary
}
