% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Define.Output.SubDir}
\alias{Define.Output.SubDir}
\title{define output directory and subdirectory and create it if necessary}
\usage{
Define.Output.SubDir(Output.Dir, ImPath, TypePCA, Sub)
}
\arguments{
\item{Output.Dir}{output directory}

\item{ImPath}{image path}

\item{TypePCA}{Type of PCA (PCA, SPCA, NLPCA...)}

\item{Sub}{subdirectory}
}
\value{
path of the output directory
}
\description{
define output directory and subdirectory and create it if necessary
}
