% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PerformPCA.R
\name{Define.Pixels.Per.Iter}
\alias{Define.Pixels.Per.Iter}
\title{defines the number of pixels per iteration based on a trade-off between image size and sample size per iteration}
\usage{
Define.Pixels.Per.Iter(ImNames, NbIter = NbIter)
}
\arguments{
\item{ImNames}{Path and name of the images to be processed}

\item{NbIter}{number of iterations peformed to average diversity indices}
}
\value{
Pix.Per.Iter number of pixels per iteration
}
\description{
defines the number of pixels per iteration based on a trade-off between image size and sample size per iteration
}
