% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_grid.R
\name{define_grid}
\alias{define_grid}
\title{define a grid over a raster}
\usage{
define_grid(raster_path, cellsize)
}
\arguments{
\item{raster_path}{character. path for raster}

\item{cellsize}{numeric. size of grid cell}
}
\value{
list
}
\description{
define a grid over a raster
}
