% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{define_output_directory}
\alias{define_output_directory}
\title{define output directory and create it if necessary}
\usage{
define_output_directory(Output_Dir, ImPath, TypePCA)
}
\arguments{
\item{Output_Dir}{character. output directory}

\item{ImPath}{character. image path}

\item{TypePCA}{character. Type of PCA (PCA, SPCA, NLPCA...)}
}
\value{
path of the output directory
}
\description{
define output directory and create it if necessary
}
