% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{define_output_subdir}
\alias{define_output_subdir}
\title{define output directory and subdirectory and create it if necessary}
\usage{
define_output_subdir(Output_Dir, ImPath, TypePCA, Sub)
}
\arguments{
\item{Output_Dir}{character. output directory}

\item{ImPath}{character. image path}

\item{TypePCA}{character. Type of PCA (PCA, SPCA, NLPCA...)}

\item{Sub}{character. subdirectory}
}
\value{
path of the output directory
}
\description{
define output directory and subdirectory and create it if necessary
}
