% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{biodivMapR_full}
\alias{biodivMapR_full_tiles}
\title{Deprecated functions in biodivMapR}
\usage{
biodivMapR_full(
  input_raster_path,
  input_mask_path = NULL,
  output_dir,
  window_size,
  selected_bands = NULL,
  Kmeans_info_save = NULL,
  Kmeans_info_read = NULL,
  Beta_info_save = NULL,
  Beta_info_read = NULL,
  nbCPU = 1,
  options = NULL
)

biodivMapR_full_tiles(
  feature_dir,
  list_features,
  mask_dir = NULL,
  output_dir,
  window_size,
  plots,
  nbCPU = 1,
  site_name = NULL,
  options = NULL
)
}
\arguments{
\item{input_raster_path}{See the new functions that replace the old ones}

\item{input_mask_path}{See the new functions that replace the old ones}

\item{output_dir}{See the new functions that replace the old ones}

\item{window_size}{See the new functions that replace the old ones}

\item{selected_bands}{See the new functions that replace the old ones}

\item{Kmeans_info_save}{See the new functions that replace the old ones}

\item{Kmeans_info_read}{See the new functions that replace the old ones}

\item{Beta_info_save}{See the new functions that replace the old ones}

\item{Beta_info_read}{See the new functions that replace the old ones}

\item{nbCPU}{See the new functions that replace the old ones}

\item{options}{See the new functions that replace the old ones}

\item{feature_dir}{See the new functions that replace the old ones}

\item{list_features}{See the new functions that replace the old ones}

\item{mask_dir}{See the new functions that replace the old ones}

\item{plots}{See the new functions that replace the old ones}

\item{site_name}{See the new functions that replace the old ones}
}
\description{
These functions are provided for compatibility with older versions of biodivMapR but are deprecated.
}
