% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_biodivMapR.R
\name{diversity_from_plots}
\alias{diversity_from_plots}
\title{gets alpha diversity indicators from plot}
\usage{
diversity_from_plots(
  Raster_SpectralSpecies,
  Plots,
  NbClusters = 50,
  Name_Plot = FALSE
)
}
\arguments{
\item{Raster_SpectralSpecies}{character. path for the SpectralSpecies file computed from DiverstyMapping method}

\item{Plots}{list. list of paths corresponding to shapefiles defining polygons in the raster}

\item{NbClusters}{numeric. Number of clusters defined in k-Means.}

\item{Name_Plot}{character. Name of the plots defined in the shapefiles}
}
\value{
alpha and beta diversity metrics
}
\description{
gets alpha diversity indicators from plot
}
