% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{ENVI.Type2Bytes}
\alias{ENVI.Type2Bytes}
\title{get information corresponding to data type defined in ENVI}
\usage{
ENVI.Type2Bytes(Header)
}
\arguments{
\item{Header}{Path of the hdr file}
}
\value{
description of data format corresponding to ENVI type
}
\description{
get information corresponding to data type defined in ENVI
}
