% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{ENVI_type2bytes}
\alias{ENVI_type2bytes}
\title{get information corresponding to data type defined in ENVI}
\usage{
ENVI_type2bytes(HDR)
}
\arguments{
\item{HDR}{list. header file}
}
\value{
description of data format corresponding to ENVI type
}
\description{
get information corresponding to data type defined in ENVI
}
