% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Exclude.Spectral.Domains}
\alias{Exclude.Spectral.Domains}
\title{define Water Vapor bands based on spectral smapling of original image}
\usage{
Exclude.Spectral.Domains(ImPath, Excluded.WL = FALSE)
}
\arguments{
\item{ImPath}{path of the image}

\item{Excluded.WL}{spectral domains corresponding to water vapor absorption}
}
\value{
bands corresponding to atmospheric water absorption domain
}
\description{
define Water Vapor bands based on spectral smapling of original image
}
