% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{exclude_spectral_domains}
\alias{exclude_spectral_domains}
\title{define Water Vapor bands based on spectral smapling of original image}
\usage{
exclude_spectral_domains(ImPath, Excluded_WL = FALSE)
}
\arguments{
\item{ImPath}{character. path of the image}

\item{Excluded_WL}{numeric. spectral domains corresponding to water vapor absorption}
}
\value{
bands corresponding to atmospheric water absorption domain
}
\description{
define Water Vapor bands based on spectral smapling of original image
}
