% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_dissimilarity.R
\name{explore_dissimilarity}
\alias{explore_dissimilarity}
\title{explore disimilarity metrics for a range of number of clusters}
\usage{
explore_dissimilarity(
  input_rast,
  output_dir,
  window_size,
  Kmeans_info,
  selected_bands = NULL,
  input_mask = NULL,
  nb_samples = 1000,
  nbClust_list = 50,
  min_sun = 0.25,
  pcelim = 0.02,
  dimPCoA = 3,
  nbCPU = 1,
  beta_metrics = "bray",
  verbose = TRUE
)
}
\arguments{
\item{input_rast}{SpatRaster. raster to extract data from}

\item{output_dir}{character. Path for output directory}

\item{window_size}{numeric. window size for square plots}

\item{Kmeans_info}{list. obtained from prepare_init_kmeans}

\item{selected_bands}{numeric. bands selected from input_rast}

\item{input_mask}{SpatRaster. mask corresponding to raster to extract data from}

\item{nb_samples}{numeric. number of samples to be extracted}

\item{nbClust_list}{numeric. number of clusters to be tested}

\item{min_sun}{numeric. minimum percentage of sunlit pixels}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{dimPCoA}{numeric.}

\item{nbCPU}{numeric. Number of CPUs available}

\item{beta_metrics}{character. name of beta dissimilarity metrics
Available options are "bray", "brayturn", "jaccard", "jaccardturn",
"sorensen", "simpson_diss"}

\item{verbose}{boolean. set true for messages}
}
\value{
list including spectral species distribution & BC diss matrix per plot, BetaPCO model
}
\description{
explore disimilarity metrics for a range of number of clusters
}
