% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_kmeans.R
\name{explore_kmeans}
\alias{explore_kmeans}
\title{Compute kmeans from random subset of pixels extracted from an image and a list
of values for k}
\usage{
explore_kmeans(
  input_rast,
  input_mask = NULL,
  SelectBands = NULL,
  nbClust_list = 50,
  nbIter = 20,
  maxPixel_kmeans = 1e+05,
  algorithm = "Hartigan-Wong",
  nbCPU = 1,
  progressbar = F
)
}
\arguments{
\item{input_rast}{character. Path and name of the image to be processed.}

\item{input_mask}{character. Path and name of the mask corresponding to the image}

\item{SelectBands}{numeric. bands selected from input_rast}

\item{nbClust_list}{numeric. number of clusters used in kmeans}

\item{nbIter}{numeric. nb of iterations averaged to compute diversity indices}

\item{maxPixel_kmeans}{numeric. max number of pixels to extract for kmeans}

\item{algorithm}{character. algorithm used in the kmeans clustering}

\item{nbCPU}{numeric. Number of CPUs available}

\item{progressbar}{boolean. set true for progress bar during clustering}
}
\value{
Kmeans_info
}
\description{
Compute kmeans from random subset of pixels extracted from an image and a list
of values for k
}
