% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_DiversityMappR.R
\name{Extract.Pixels.Coordinates.From.OGR}
\alias{Extract.Pixels.Coordinates.From.OGR}
\title{Extracts pixels coordinates from raster corresponding to an area defined by a vector}
\usage{
Extract.Pixels.Coordinates.From.OGR(Path.Raster, OGR.Vector)
}
\arguments{
\item{Path.Raster}{path for the raster file. !! BIL expected}

\item{OGR.Vector}{OGR for the vector file obtained from readOGR}
}
\value{
ColRow list of coordinates of pixels corresponding to each polygon in shp
}
\description{
Extracts pixels coordinates from raster corresponding to an area defined by a vector
}
