% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Extract.Pixels}
\alias{Extract.Pixels}
\title{extracts sample points from binary image file}
\usage{
Extract.Pixels(coordPix.List, ImPath, HDR)
}
\arguments{
\item{coordPix.List}{coordinates of pixels to sample}

\item{ImPath}{path for image}

\item{HDR}{hdr path}
}
\value{
samples from image subset corresponding to coordPix.List
}
\description{
extracts sample points from binary image file
}
