% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Extract.Samples.From.Image}
\alias{Extract.Samples.From.Image}
\title{extracts pixels from image based on their coordinates}
\usage{
Extract.Samples.From.Image(ImPath, coordPix, MaxRAM = FALSE,
  ProgressBar = FALSE, Already.Split = FALSE)
}
\arguments{
\item{ImPath}{path for image}

\item{coordPix}{pixel coordinates}
}
\value{
samples from image corresponding to coordPix
}
\description{
extracts pixels from image based on their coordinates
}
