% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_rast_sfs.R
\name{extract_from_rast_sfs}
\alias{extract_from_rast_sfs}
\title{extract informaton from raster to prepare for SFS}
\usage{
extract_from_rast_sfs(input_raster, input_mask, min_sun, obs_vect)
}
\arguments{
\item{input_raster}{SpatRaster or list of SpatRaster}

\item{input_mask}{SpatRaster corresponding to mask}

\item{min_sun}{numeric.}

\item{obs_vect}{SpatVector or SpatVectorCollection}
}
\value{
list
}
\description{
extract informaton from raster to prepare for SFS
}
