% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{extract_pixels}
\alias{extract_pixels}
\title{extracts sample points from binary image file}
\usage{
extract_pixels(coordPix_List, ImPath, HDR)
}
\arguments{
\item{coordPix_List}{list. coordinates of pixels to sample}

\item{ImPath}{character. path for image}

\item{HDR}{list. hdr path}
}
\value{
samples from image subset corresponding to coordPix_List
}
\description{
extracts sample points from binary image file
}
