% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_biodivMapR.R
\name{extract_pixels_coordinates}
\alias{extract_pixels_coordinates}
\title{Extracts pixels coordinates from raster corresponding to an area defined by a vector}
\usage{
extract_pixels_coordinates(Path.Raster, Path.Vector)
}
\arguments{
\item{Path.Raster}{path for the raster file. !! BIL expected}

\item{Path.Vector}{path for the vector file. !! SHP expected}
}
\value{
ColRow list of coordinates of pixels corresponding to each polygon in shp
}
\description{
Extracts pixels coordinates from raster corresponding to an area defined by a vector
}
