% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{extract_samples_from_image}
\alias{extract_samples_from_image}
\title{extracts pixels from image based on their coordinates}
\usage{
extract_samples_from_image(
  ImPath,
  coordPix,
  MaxRAM = FALSE,
  Already.Split = FALSE
)
}
\arguments{
\item{ImPath}{character. path for image}

\item{coordPix}{numeric. pixel coordinates}

\item{MaxRAM}{numeric.}

\item{Already.Split}{Boolean.}
}
\value{
samples from image corresponding to coordPix
}
\description{
extracts pixels from image based on their coordinates
}
