% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vect_from_rast.R
\name{extract_vect_from_rast}
\alias{extract_vect_from_rast}
\title{extract pixel information from a raster based on vector footprint}
\usage{
extract_vect_from_rast(
  SpatVector,
  input_rast,
  input_mask = NULL,
  min_sun = 0.25,
  prog = TRUE
)
}
\arguments{
\item{SpatVector}{SpatVector. object}

\item{input_rast}{SpatRaster. input raster}

\item{input_mask}{SpatRaster. optional mask}

\item{min_sun}{numeric. minimum sunlit/unmasked proportion}

\item{prog}{boolean. progressbar ?}
}
\value{
list including rast_sample: dataframe corresponding to
information extracted from input_rast, and AttributeTable from SpatVector
}
\description{
extract pixel information from a raster based on vector footprint
}
