% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_metrics_sfs.R
\name{fd_metrics_sfs}
\alias{fd_metrics_sfs}
\title{compute functional metrics during sfs}
\usage{
fd_metrics_sfs(
  fd_metrics,
  obs2optimize,
  rast_val,
  Kmeans_info,
  IDplot,
  corrMethod
)
}
\arguments{
\item{fd_metrics}{character}

\item{obs2optimize}{numeric. number of clusters used in kmeans}

\item{rast_val}{numeric}

\item{Kmeans_info}{list.}

\item{IDplot}{character}

\item{corrMethod}{character}
}
\value{
assesssed value and correlation
}
\description{
compute functional metrics during sfs
}
