% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_path_sans_ext.R
\name{file_path_sans_ext}
\alias{file_path_sans_ext}
\title{Re-implementation of \code{\link{file_path_sans_ext}} in \code{tools}. This
version can handle "." just before the file extenstion, unlike the original
implementation.
taken from https://github.com/cbig/zonator}
\usage{
file_path_sans_ext(x, compression = FALSE)
}
\arguments{
\item{x}{Character vector giving file paths.}

\item{compression}{Logical: should compression extension '.gz', '.bz2' or
'.xz' be removed first?}
}
\value{
File path without the file extension.
}
\description{
Re-implementation of \code{\link{file_path_sans_ext}} in \code{tools}. This
version can handle "." just before the file extenstion, unlike the original
implementation.
taken from https://github.com/cbig/zonator
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
