% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames_explore_cluster_range.R
\name{filenames_explore_cluster_range}
\alias{filenames_explore_cluster_range}
\title{checks if files produced from explore_cluster_range exist}
\usage{
filenames_explore_cluster_range(outputdir, obs_criterion, repet)
}
\arguments{
\item{outputdir}{character}

\item{obs_criterion}{character}

\item{repet}{character}
}
\value{
files_exist boolean if all expected files exist
}
\description{
checks if files produced from explore_cluster_range exist
}
