% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames_opt_clusters.R
\name{filenames_opt_clusters}
\alias{filenames_opt_clusters}
\title{checks if files produced from biodivMapR_opt_cluster exist}
\usage{
filenames_opt_clusters(outputdir, obs_criterion)
}
\arguments{
\item{outputdir}{character}

\item{obs_criterion}{character}
}
\value{
files_exist boolean if all expected files exist
}
\description{
checks if files produced from biodivMapR_opt_cluster exist
}
