% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PerformPCA.R
\name{Filter.PCA}
\alias{Filter.PCA}
\title{perform filtering based on extreme values PCA identified through PCA}
\usage{
Filter.PCA(ImPath, HDR, ImPathShade, Shade.Update, Spectral, CR, PCA.model,
  PCsel, TypePCA, nbCPU = 1, MaxRAM = 0.25)
}
\arguments{
\item{ImPath}{image path}

\item{HDR}{hdr file file correspmonding to ImPath}

\item{ImPathShade}{shade file path}

\item{Shade.Update}{updated shade mask}

\item{Spectral}{spectral information from data}

\item{CR}{logical: does continuum removal need to be performed?}

\item{PCA.model}{general parameters of the PCA}

\item{PCsel}{PCs used to filter out extreme values}
}
\value{
Shade.Update = updated shade mask
}
\description{
perform filtering based on extreme values PCA identified through PCA
}
