% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ContinuumRemoval.R
\name{Filter.Prior.CR}
\alias{Filter.Prior.CR}
\title{Filter data prior to continuum removal:
- values are expected to be real reflectance values between 0 and 10000
- negative values may occur, so a +100 value is applied to avoid negative
- possibly remaining negative values are set to 0
- constant spectra are eliminated}
\usage{
Filter.Prior.CR(Minit, Spectral.Bands)
}
\arguments{
\item{Minit}{initial data matrix, n rows = n samples, p cols = p spectral bands}
}
\value{
updated Minit
}
\description{
Filter data prior to continuum removal:
- values are expected to be real reflectance values between 0 and 10000
- negative values may occur, so a +100 value is applied to avoid negative
- possibly remaining negative values are set to 0
- constant spectra are eliminated
}
