% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PerformPCA.R
\name{filter_PCA}
\alias{filter_PCA}
\title{perform filtering based on extreme values PCA identified through PCA}
\usage{
filter_PCA(
  Input_Image_File,
  HDR,
  Input_Mask_File,
  Shade_Update,
  Spectral,
  Continuum_Removal,
  PCA_model,
  PCsel,
  TypePCA,
  nbCPU = 1,
  MaxRAM = 0.25
)
}
\arguments{
\item{Input_Image_File}{character. Path of the image to be processed}

\item{HDR}{character. Path of the header file corresponding to the image to be processed}

\item{Input_Mask_File}{character. Path of the mask raster corresponding to the image (keeps pixels = 1)}

\item{Shade_Update}{character. Path of the updated mask raster corresponding to the image (keeps pixels = 1)}

\item{Spectral}{list. spectral information from data}

\item{Continuum_Removal}{boolean. set TRUE if continuum removal should be applied}

\item{PCA_model}{dataframe. general parameters of the PCA}

\item{PCsel}{numeric. PCs used to filter out extreme values}

\item{TypePCA}{character. Set to PCA, SPCA or MNF}

\item{nbCPU}{numeric. number of CPUs to be used in parallel}

\item{MaxRAM}{numeric. indicator of RAM to be used to read image file}
}
\value{
Shade_Update = updated shade mask
}
\description{
perform filtering based on extreme values PCA identified through PCA
}
