% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_prior_cr.R
\name{filter_prior_cr}
\alias{filter_prior_cr}
\title{Filter data prior to continuum removal:
\itemize{
\item values are expected to be real reflectance values between 0 and 10000
\item negative values may occur, so a +100 value is applied to avoid negative
\item possibly remaining negative values are set to 0
\item constant spectra are eliminated
}}
\usage{
filter_prior_cr(mat_init, spectral_bands)
}
\arguments{
\item{mat_init}{initial data matrix, n rows = n samples, p cols = p bands}

\item{spectral_bands}{numeric. central wavelength for the spectral bands}
}
\value{
list. updated mat_init
}
\description{
Filter data prior to continuum removal:
\itemize{
\item values are expected to be real reflectance values between 0 and 10000
\item negative values may occur, so a +100 value is applied to avoid negative
\item possibly remaining negative values are set to 0
\item constant spectra are eliminated
}
}
