% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_window_list.R
\name{functional_window_list}
\alias{functional_window_list}
\title{apply functional_window to a list of lists}
\usage{
functional_window_list(SSwindow, fd_metrics = c("FRic", "FEve", "FDiv"))
}
\arguments{
\item{SSwindow}{dataframe. spectral species corresponding to a raster subset
(window = elementary spatial unit of process)}

\item{fd_metrics}{list. functional diversity metrics: FRic, FEve, FDiv}
}
\value{
list of functional diversity metrics
}
\description{
apply functional_window to a list of lists
}
