% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_DiversityMappR.R
\name{Get.Alpha.Metrics}
\alias{Get.Alpha.Metrics}
\title{Computes alpha diversity metrics from distribution}
\usage{
Get.Alpha.Metrics(Distrib)
}
\arguments{
\item{Distrib}{distribution of clusters}
}
\value{
Richness, Fisher, Shannon, Simpson
}
\description{
Computes alpha diversity metrics from distribution
}
