% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Get.Byte.Order}
\alias{Get.Byte.Order}
\title{does the system work with little endians or big endians?}
\usage{
Get.Byte.Order()
}
\value{
ByteOrder
}
\description{
does the system work with little endians or big endians?
}
