% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_DiversityMappR.R
\name{Get.Diversity.From.Plots}
\alias{Get.Diversity.From.Plots}
\title{gets alpha diversity indicators from plot}
\usage{
Get.Diversity.From.Plots(Raster, Plots, NbClusters = 50,
  Name.Plot = FALSE)
}
\arguments{
\item{Raster}{SpectralSpecies file computed from DiverstyMapping method}

\item{Plots}{list of shapefiles included in the raster}
}
\value{
alpha and beta diversity metrics
}
\description{
gets alpha diversity indicators from plot
}
