% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Get.HDR.Name}
\alias{Get.HDR.Name}
\title{get hdr name from image file name, assuming it is BIL format}
\usage{
Get.HDR.Name(ImPath)
}
\arguments{
\item{ImPath}{path of the image}
}
\value{
corresponding hdr
}
\description{
get hdr name from image file name, assuming it is BIL format
}
