% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Get.Image.Bands}
\alias{Get.Image.Bands}
\title{gets rank of spectral bands in an image}
\usage{
Get.Image.Bands(Spectral.Bands, wavelength)
}
\arguments{
\item{Spectral.Bands}{wavelength (nm) of the spectral bands to be found}

\item{wavelength}{wavelength (nm) of all wavelengths in the image}
}
\value{
rank of all spectral bands of interest in the image and corresponding wavelength
}
\description{
gets rank of spectral bands in an image
}
