% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_biodivMapR.R
\name{Get.List.Shp}
\alias{Get.List.Shp}
\title{Get list of shapefiles in a directory}
\usage{
Get.List.Shp(Path.Shp)
}
\arguments{
\item{Path.Shp}{path of the directory containing shapefiles}
}
\value{
list of shapefiles names
}
\description{
Get list of shapefiles in a directory
}
