% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_biodivMapR.R
\name{Get.Projection}
\alias{Get.Projection}
\title{get projection of a raster or a vector}
\usage{
Get.Projection(Path.File, Type = "raster")
}
\arguments{
\item{Path.File}{path for a raster or vector (shapefile)}

\item{Type}{'raster' or 'vector'}
}
\value{
projection
}
\description{
get projection of a raster or a vector
}
