% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Get.Random.Subset.From.Image}
\alias{Get.Random.Subset.From.Image}
\title{Perform random sampling on an image including a shade mask}
\usage{
Get.Random.Subset.From.Image(ImPath, HDR, ImPathShade, NbIter,
  Pix.Per.Iter)
}
\arguments{
\item{ImPath}{path for image}

\item{HDR}{path for hdr file}

\item{ImPathShade}{path for shade mask}

\item{NbIter}{number of iterations}

\item{Pix.Per.Iter}{number of pixels per iteration}
}
\value{
samples from image and updated number of pixels to sampel if necessary
}
\description{
Perform random sampling on an image including a shade mask
}
