% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{Get.Shannon}
\alias{Get.Shannon}
\title{computes shannon index from a distribution}
\usage{
Get.Shannon(Distrib)
}
\arguments{
\item{Distrib}{Distribution}
}
\value{
Shannon index correspnding to the distribution
}
\description{
computes shannon index from a distribution
}
