% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{Get.Simpson}
\alias{Get.Simpson}
\title{computes Simpson index from a distribution}
\usage{
Get.Simpson(Distrib)
}
\arguments{
\item{Distrib}{Distribution}
}
\value{
Simpson index correspnding to the distribution
}
\description{
computes Simpson index from a distribution
}
