% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{Get.Sunlit.Pixels}
\alias{Get.Sunlit.Pixels}
\title{Gets sunlit pixels from SpectralSpecies_Distribution_Sunlit}
\usage{
Get.Sunlit.Pixels(ImPathSunlit, MinSun)
}
\arguments{
\item{ImPathSunlit}{path for SpectralSpecies_Distribution_Sunlit}

\item{MinSun}{minimum proportion of sunlit pixels required}
}
\description{
Gets sunlit pixels from SpectralSpecies_Distribution_Sunlit
}
