% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha.R
\name{get_alpha}
\alias{get_alpha}
\title{computes alpha diversity metrics from ssd_mat corresponding to iterations}
\usage{
get_alpha(ssd_mat, alpha_metrics, Hill_order = 1)
}
\arguments{
\item{ssd_mat}{numeric. spectral species distribution as a matrix:
species = columns, iterations = rows}

\item{alpha_metrics}{list. alpha diversity metrics: richness, shannon, simpson, hill}

\item{Hill_order}{numeric. Hill order}
}
\value{
alpha diversity metrics corresponding to the distribution
}
\description{
computes alpha diversity metrics from ssd_mat corresponding to iterations
}
