% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_biodivMapR.R
\name{get_alpha_metrics}
\alias{get_alpha_metrics}
\title{Computes alpha diversity metrics from distribution}
\usage{
get_alpha_metrics(Distrib)
}
\arguments{
\item{Distrib}{distribution of clusters}
}
\value{
Richness, Fisher, Shannon, Simpson
}
\description{
Computes alpha diversity metrics from distribution
}
