% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asset_path.R
\name{get_asset_path}
\alias{get_asset_path}
\title{This function gets path from an asset in the JSON file}
\usage{
get_asset_path(jsonobj, asset, dateAcq)
}
\arguments{
\item{jsonobj}{character. path for json file}

\item{asset}{character. name of the asset of interest}

\item{dateAcq}{date. date of acquisition}
}
\value{
asset_path
}
\description{
This function gets path from an asset in the JSON file
}
