% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{get_BB}
\alias{get_BB}
\title{This function gets coordinates of a bounding box defined by a vector (optional) and a raster}
\usage{
get_BB(path_raster, path_vector = NULL, Buffer = 0)
}
\arguments{
\item{path_raster}{character. path for raster file}

\item{path_vector}{character. path for vector file}

\item{Buffer}{numeric. buffer applied to vector file (in meters)}
}
\value{
BB_XYcoords list. Coordinates (in pixels) of the upper/lower right/left corners of bounding box
}
\description{
This function gets coordinates of a bounding box defined by a vector (optional) and a raster
}
