% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{get_BB_from_fullImage}
\alias{get_BB_from_fullImage}
\title{This function gets extreme coordinates of a bounding box corresponding to a full image}
\usage{
get_BB_from_fullImage(path_raster)
}
\arguments{
\item{path_raster}{character. path for raster file}
}
\value{
BB_XYcoords list. Coordinates (in pixels) of the upper/lower right/left corners of bounding box
}
\description{
This function gets extreme coordinates of a bounding box corresponding to a full image
}
