% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{get_BB_from_Vector}
\alias{get_BB_from_Vector}
\title{This gets bounding box corresponding to a vector from a raster (UL, UR, LL, LR corners)}
\usage{
get_BB_from_Vector(path_raster, path_vector, Buffer = 0)
}
\arguments{
\item{path_raster}{character. path for raster file}

\item{path_vector}{character. path for vector file}

\item{Buffer}{numeric. buffer applied to vector file (in meters)}
}
\value{
BB_XYcoords list. Coordinates (in pixels) of the upper/lower right/left corners of bounding box
}
\description{
This gets bounding box corresponding to a vector from a raster (UL, UR, LL, LR corners)
}
