% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bc_diss_from_ssd.R
\name{get_bc_diss_from_ssd}
\alias{get_bc_diss_from_ssd}
\title{Computes BC dissimilarity for a list of spectral species distributions}
\usage{
get_bc_diss_from_ssd(ss_dist, nb_clusters, pcelim, p = NULL)
}
\arguments{
\item{ss_dist}{numeric. list of spectral species distribution}

\item{nb_clusters}{numeric. number of clusters used in kmeans}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{p}{list. progressor object for progress bar}
}
\value{
list of spectral species distribution and corresponding  BC
dissimilarity matrix corresponding to Mat1 and Mat2
}
\description{
Computes BC dissimilarity for a list of spectral species distributions
}
