% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BCdiss.R
\name{get_BCdiss}
\alias{get_BCdiss}
\title{Computes BC dissimilarity for a given matrix}
\usage{
get_BCdiss(Mat1, pcelim = 0.02, p = NULL)
}
\arguments{
\item{Mat1}{numeric. matrix of spectral species distribution}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{p}{list. progressor object for progress bar}
}
\value{
BCtmp numeric. BC dissimilarity matrix corresponding to Mat1 and Mat2
}
\description{
Computes BC dissimilarity for a given matrix
}
