% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BCdiss_from_SSD.R
\name{get_BCdiss_from_SSD}
\alias{get_BCdiss_from_SSD}
\title{Computes BC dissimilarity for a list of spectral species distributions}
\usage{
get_BCdiss_from_SSD(SSdist, nbclusters, pcelim, p = NULL)
}
\arguments{
\item{SSdist}{numeric. list of spectral species distribution}

\item{nbclusters}{numeric. number of clusters used in kmeans}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{p}{list. progressor object for progress bar}
}
\value{
list of spectral species distribution and corresponding  BC dissimilarity matrix corresponding to Mat1 and Mat2
}
\description{
Computes BC dissimilarity for a list of spectral species distributions
}
