% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beta_from_ssd.R
\name{get_beta_from_ssd}
\alias{get_beta_from_ssd}
\title{Computes beta diversity for a list of spectral species distributions}
\usage{
get_beta_from_ssd(
  ss_dist,
  beta_metrics = "bray",
  nb_clusters,
  pcelim,
  p = NULL
)
}
\arguments{
\item{ss_dist}{numeric. list of spectral species distribution}

\item{beta_metrics}{character. name of beta dissimilarity metrics
Available options are "bray", "brayturn", "jaccard", "jaccardturn",
"sorensen", "simpson_diss"}

\item{nb_clusters}{numeric. number of clusters used in kmeans}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{p}{list. progressor object for progress bar}
}
\value{
list of spectral species distribution and corresponding  BC
dissimilarity matrix corresponding to Mat1 and Mat2
}
\description{
Computes beta diversity for a list of spectral species distributions
}
