% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{get_byte_order}
\alias{get_byte_order}
\title{does the system work with little endians or big endians?}
\usage{
get_byte_order()
}
\value{
ByteOrder
}
\description{
does the system work with little endians or big endians?
}
