% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date.R
\name{get_date}
\alias{get_date}
\title{This function gets acquisition date from S2 image}
\usage{
get_date(prodName)
}
\arguments{
\item{prodName}{character. original name for the S2 image}
}
\value{
DateAcq character
}
\description{
This function gets acquisition date from S2 image
}
