% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diversity_from_plots_cluster.R
\name{get_diversity_from_plots_cluster}
\alias{get_diversity_from_plots_cluster}
\title{computes diversity metrics from validation plots}
\usage{
get_diversity_from_plots_cluster(
  Kmeans_info,
  rast_sample,
  AttributeTable,
  Hill_order = 1,
  Beta_info = NULL,
  selected_bands = NULL,
  alpha_metrics = c("richness", "shannon", "simpson", "hill"),
  beta_metrics = c("bray"),
  pcelim = 0.02
)
}
\arguments{
\item{Kmeans_info}{list. kmeans description obtained from function get_kmeans}

\item{rast_sample}{list. information extracted from raster}

\item{AttributeTable}{dataframe. corresponds to attributes of original vector file}

\item{Hill_order}{numeric. Hill order}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics from training set}

\item{selected_bands}{numeric. bands selected from input_rast}

\item{alpha_metrics}{character.}

\item{beta_metrics}{character.}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}
}
\value{
SpatVector including diversity metrics and BC dissimilarity for the plots
}
\description{
computes diversity metrics from validation plots
}
