% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{get_gdal_info}
\alias{get_gdal_info}
\title{Get GDAL info as a nested list}
\usage{
get_gdal_info(x)
}
\arguments{
\item{x}{character. Raster image path.}
}
\value{
list. Nested list of raster characteristics.
}
\description{
Get GDAL info as a nested list
}
