% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hdr_name.R
\name{get_hdr_name}
\alias{get_hdr_name}
\title{get hdr name from image file name}
\usage{
get_hdr_name(image_path, showWarnings = TRUE)
}
\arguments{
\item{image_path}{character. ath of the image}

\item{showWarnings}{boolean. set TRUE if warning because HDR does not exist}
}
\value{
path for corresponding hdr
}
\description{
get hdr name from image file name
}
