% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image_bands.R
\name{get_image_bands}
\alias{get_image_bands}
\title{gets rank of spectral bands in an image}
\usage{
get_image_bands(spectral_bands, wavelength)
}
\arguments{
\item{spectral_bands}{wavelength (nm) of the spectral bands to be found}

\item{wavelength}{wavelength (nm) of all wavelengths in the image}
}
\value{
rank of all spectral bands of interest in the image and corresponding wavelength
and distance to wavelength
}
\description{
gets rank of spectral bands in an image
}
