% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_chunk.R
\name{get_input_chunk}
\alias{get_input_chunk}
\title{get input chunk data from a list of rasters}
\usage{
get_input_chunk(r_in, blk, selected_bands, window_size)
}
\arguments{
\item{r_in}{list. list of input SpatRasters}

\item{blk}{list. starting row for reading and nb of rows to read}

\item{selected_bands}{numeric. rank of bands to select}

\item{window_size}{numeric. window size}
}
\value{
list. includes filtered input_data, selected_bands & nb_windows
}
\description{
get input chunk data from a list of rasters
}
