% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_normalized_ssd_list.R
\name{get_normalized_ssd_list}
\alias{get_normalized_ssd_list}
\title{get spectral species distribution from a list for all clusters
including those with 0 abundance}
\usage{
get_normalized_ssd_list(ssd_list, windows_id, nb_clusters, pcelim = 0.02)
}
\arguments{
\item{ssd_list}{dataframe. spectral species distribution}

\item{windows_id}{numeric. window id corresponding to ssd_list}

\item{nb_clusters}{numeric. number of clusters used in kmeans}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}
}
\value{
ssd_map normalized distribution for all clusters including those with 0 abundance
}
\description{
get spectral species distribution from a list for all clusters
including those with 0 abundance
}
