% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_normalized_ssd_mat.R
\name{get_normalized_ssd_mat}
\alias{get_normalized_ssd_mat}
\title{get spectral species distribution for all clusters
including those with 0 abundance}
\usage{
get_normalized_ssd_mat(ssd, nb_clusters, pcelim = 0.02)
}
\arguments{
\item{ssd}{dataframe. spectral species distribution}

\item{nb_clusters}{numeric. number of clusters used in kmeans}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}
}
\value{
ssd_map normalized distribution for all clusters including those with 0 abundance
}
\description{
get spectral species distribution for all clusters
including those with 0 abundance
}
