% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plots_from_tiles.R
\name{get_plots_from_tiles}
\alias{get_plots_from_tiles}
\title{get samples for beta diversity mapping}
\usage{
get_plots_from_tiles(
  plotID,
  plots2sel,
  listfiles,
  feat_list,
  window_size,
  minSun = 0.75,
  p = NULL
)
}
\arguments{
\item{plotID}{list.}

\item{plots2sel}{numeric.}

\item{listfiles}{character.}

\item{feat_list}{list.}

\item{window_size}{numeric.}

\item{minSun}{numeric.}

\item{p}{list.}
}
\value{
samples_beta
}
\description{
get samples for beta diversity mapping
}
