% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{get_polygonCoord_from_Shp}
\alias{get_polygonCoord_from_Shp}
\title{extract coordinates of the footprint of elements of a shapefile in a raster}
\usage{
get_polygonCoord_from_Shp(path_SHP, path_Raster, IDshp = NULL)
}
\arguments{
\item{path_SHP}{character. path for the shapefile}

\item{path_Raster}{character. path for the raster}

\item{IDshp}{character. field in the shapefile determining ID of element}
}
\value{
list. vector_coordinates and vector_ID for each element in the vector file
}
\description{
extract coordinates of the footprint of elements of a shapefile in a raster
}
