% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{get_random_subset_from_image}
\alias{get_random_subset_from_image}
\title{extract random subset of pixels from an image}
\usage{
get_random_subset_from_image(
  ImPath,
  MaskPath,
  nb_partitions,
  Pix_Per_Partition,
  kernel = NULL,
  MaxRAM = 0.5
)
}
\arguments{
\item{ImPath}{character. path for the image to sample}

\item{MaskPath}{character. path for the corresponding mask}

\item{nb_partitions}{numeric. number of repetitions of kmeans}

\item{Pix_Per_Partition}{numeric.}

\item{kernel}{numeric.}

\item{MaxRAM}{numeric.}
}
\value{
list, see Details
}
\description{
extract random subset of pixels from an image
}
\details{
The returned list contains:
- DataSubset: matrix of NxP of N samples and P bands
- nbPix2Sample: integer giving the number of pixels sampled (only central pixel of kernel)
- coordPix: a data.table with columns 'row', 'col' of pixel in the image corresponding to each row of DataSubset, and if kernel is not NULL
Kind (Kernel index) and 'id' the sample ID to be used with the kernel
}
