% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_diversity.R
\name{get_raster_diversity}
\alias{get_raster_diversity}
\title{Computes diversity metrics from raster data}
\usage{
get_raster_diversity(
  input_raster_path,
  Kmeans_info,
  Beta_info,
  input_mask_path = NULL,
  selected_bands = NULL,
  alpha_metrics = "shannon",
  Hill_order = 1,
  beta_metrics = "bray",
  fd_metrics = NULL,
  window_size,
  maxRows = NULL,
  pcelim = 0.02,
  nbCPU = 1,
  min_sun = 0.25
)
}
\arguments{
\item{input_raster_path}{list. list of paths corresponding to input rasters}

\item{Kmeans_info}{list. kmeans description obtained from function get_kmeans}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics from training set}

\item{input_mask_path}{character. path for mask file}

\item{selected_bands}{numeric. bands selected from input_rast}

\item{alpha_metrics}{list. alpha diversity metrics: richness, shannon, simpson}

\item{Hill_order}{numeric. Hill order}

\item{beta_metrics}{list. beta diversity metrics}

\item{fd_metrics}{character. list of functional metrics}

\item{window_size}{numeric. window size for square plots}

\item{maxRows}{numeric. max number of rows processed once by each CPU}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{nbCPU}{numeric. Number of CPUs available}

\item{min_sun}{numeric. minimum amount of sunlit pixels in the plots}
}
\value{
ab_div_metrics list. contains all metrics
}
\description{
Computes diversity metrics from raster data
}
