% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_extent.R
\name{get_raster_extent}
\alias{get_raster_extent}
\title{gets raster extent}
\usage{
get_raster_extent(input_rast)
}
\arguments{
\item{input_rast}{SpatRaster.}
}
\value{
extent_area as SpatVector
}
\description{
gets raster extent
}
