% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_functional.R
\name{get_raster_functional}
\alias{get_raster_functional}
\title{Computes diversity metrics from raster data}
\usage{
get_raster_functional(
  input_raster_path,
  input_mask_path = NULL,
  SelectBands = NULL,
  Functional = c("FRic", "FEve", "FDiv"),
  window_size,
  maxRows = NULL,
  pcelim = 0.02,
  nbCPU = 1,
  MinSun = 0.25
)
}
\arguments{
\item{input_raster_path}{list. list of paths corresponding to input rasters}

\item{input_mask_path}{character. path for mask file}

\item{SelectBands}{numeric. bands selected from input_rast}

\item{Functional}{list. alpha diversity metrics: richness, shannon, simpson}

\item{window_size}{numeric. window size for square plots}

\item{maxRows}{numeric. max number of rows in each block}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{nbCPU}{numeric. Number of CPUs available}

\item{MinSun}{numeric. minimum amount of sunlit pixels in the plots}
}
\value{
ab_div_metrics list. contains all metrics
}
\description{
Computes diversity metrics from raster data
}
