% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples_from_plots.R
\name{get_samples_from_plots}
\alias{get_samples_from_plots}
\title{extracts samples defined by a vector layer from a raster}
\usage{
get_samples_from_plots(x, y)
}
\arguments{
\item{x}{spatVector vector layer}

\item{y}{spatRaster raster}
}
\value{
res
}
\description{
extracts samples defined by a vector layer from a raster
}
