% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Shannon.R
\name{get_Shannon}
\alias{get_Shannon}
\title{computes shannon index from a distribution
(faster than version implemented in vegan package)}
\usage{
get_Shannon(Distrib)
}
\arguments{
\item{Distrib}{Distribution}
}
\value{
Shannon index corresponding to the distribution
}
\description{
computes shannon index from a distribution
(faster than version implemented in vegan package)
}
