% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{get_Simpson}
\alias{get_Simpson}
\title{computes Simpson index from a distribution
(faster than version implemented in vegan package)}
\usage{
get_Simpson(Distrib)
}
\arguments{
\item{Distrib}{Distribution}
}
\value{
Simpson index correspnding to the distribution
}
\description{
computes Simpson index from a distribution
(faster than version implemented in vegan package)
}
