% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SSD_full.R
\name{get_SSD_full}
\alias{get_SSD_full}
\title{get spectral species distribution for all clusters, even those with null
abundance}
\usage{
get_SSD_full(SSD, nbclusters, pcelim = 0.02)
}
\arguments{
\item{SSD}{dataframe. spectral species distribution}

\item{nbclusters}{numeric. number of clusters used in kmeans}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}
}
\value{
SSwindow results of kmeans per window
}
\description{
get spectral species distribution for all clusters, even those with null
abundance
}
