% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{get_SSpath}
\alias{get_SSpath}
\title{defines path for the spectral species raster file. Depends on preprocessing:
- if standard dimensionality reduction (PCA, SPCA, MNF) was used, the
corresponding directory is pointed
- if no dimensionality reduction, or an alternative dimensionality reduction
(spectral indices, tasseled cap... ) was used, user defined subdirectory is used
- Diversity metrics can also be computed from a classification map}
\usage{
get_SSpath(Output_Dir, Input_Image_File, TypePCA, ClassifMap, nbclusters)
}
\arguments{
\item{Output_Dir}{character. Output directory.}

\item{Input_Image_File}{character. Path and name of the image to be processed.}

\item{TypePCA}{character. Type of PCA (PCA, SPCA, NLPCA...).}

\item{ClassifMap}{character. If FALSE, perform standard biodivMapR based on SpectralSpecies.
else corresponds to path for a classification map.}

\item{nbclusters}{numeric. number of clusters (possibly updated)}
}
\value{
list. includes Spectral_Species_Path
}
\description{
defines path for the spectral species raster file. Depends on preprocessing:
- if standard dimensionality reduction (PCA, SPCA, MNF) was used, the
corresponding directory is pointed
- if no dimensionality reduction, or an alternative dimensionality reduction
(spectral indices, tasseled cap... ) was used, user defined subdirectory is used
- Diversity metrics can also be computed from a classification map
}
