% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats_from_tiles.R
\name{get_stats_from_tiles}
\alias{get_stats_from_tiles}
\title{get samples from tiles produced with preprocS2 function get_s2_tiling
in order to produce stats}
\usage{
get_stats_from_tiles(plotID, pix2sel, listfiles, SI_list, p = NULL)
}
\arguments{
\item{plotID}{list.}

\item{pix2sel}{numeric.}

\item{listfiles}{character.}

\item{SI_list}{list.}

\item{p}{list.}
}
\value{
selpix
}
\description{
get samples from tiles produced with preprocS2 function get_s2_tiling
in order to produce stats
}
