% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{get_sunlit_pixels}
\alias{get_sunlit_pixels}
\title{Gets sunlit pixels from SpectralSpecies_Distribution_Sunlit}
\usage{
get_sunlit_pixels(ImPathSunlit, MinSun)
}
\arguments{
\item{ImPathSunlit}{path for SpectralSpecies_Distribution_Sunlit}

\item{MinSun}{minimum proportion of sunlit pixels required}
}
\value{
list of sunlit pixels
}
\description{
Gets sunlit pixels from SpectralSpecies_Distribution_Sunlit
}
