% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sunlitwindows.R
\name{get_sunlitwindows}
\alias{get_sunlitwindows}
\title{eliminate windows with insufficient sunlit pixels}
\usage{
get_sunlitwindows(inputdata, pixperplot, min_sun = 0.25)
}
\arguments{
\item{inputdata}{list. input data with window_ID produced from produce_win_ID}

\item{pixperplot}{numeric. minimum number of pixels per window}

\item{min_sun}{numeric. minimum percentage of sunlit pixels}
}
\value{
inputwindow list of windows with proportion of sunlit pixels > min_sun
}
\description{
eliminate windows with insufficient sunlit pixels
}
