% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valid_pixels.R
\name{get_valid_pixels}
\alias{get_valid_pixels}
\title{get valid pixels from a list of plots produced with
preprocS2 function get_s2_tiling (notNA)}
\usage{
get_valid_pixels(plotID, listfiles, p = NULL)
}
\arguments{
\item{plotID}{list.}

\item{listfiles}{character.}

\item{p}{list.}
}
\value{
nb_pix_valid
}
\description{
get valid pixels from a list of plots produced with
preprocS2 function get_s2_tiling (notNA)
}
