% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xy_samples.R
\name{get_xy_samples}
\alias{get_xy_samples}
\title{get xy of pixels to sample from raster}
\usage{
get_xy_samples(input_rast, nbSamples, input_mask = NULL)
}
\arguments{
\item{input_rast}{SpatRaster. raster to extract data from}

\item{nbSamples}{numeric. number of samples to be extracted}

\item{input_mask}{SpatRaster. mask corresponding to raster to extract data from}
}
\value{
xy matrix
}
\description{
get xy of pixels to sample from raster
}
