% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapFunctionalDiversity.R
\name{getFD}
\alias{getFD}
\title{get functional diversity metrics from dataframe
This function was inspired from FD package}
\usage{
getFD(spectraits, FDmetric = c("FRic", "FEve", "FDiv"), p = NULL)
}
\arguments{
\item{spectraits}{numeric. dataframe containing species in rows and trait values in columns}

\item{FDmetric}{character. Functional diversity metric}

\item{p}{list. progressor object for progress bar}
}
\value{
FDmetrics
}
\description{
get functional diversity metrics from dataframe
This function was inspired from FD package
}
