% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_envi_hdr.R
\name{harmonize_envi_hdr}
\alias{harmonize_envi_hdr}
\title{Harmonize ENVI header file}
\usage{
harmonize_envi_hdr(output_raster, filetype, bandNames)
}
\arguments{
\item{output_raster}{character. path for raster file}

\item{filetype}{character. raster driver}

\item{bandNames}{character. name for bands in raster}
}
\value{
list of the content of the hdr file
}
\description{
Harmonize ENVI header file
}
