% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{ind2sub}
\alias{ind2sub}
\title{convert image coordinates from index to X-Y}
\usage{
ind2sub(Raster, Image.Index)
}
\arguments{
\item{Raster}{image raster object}

\item{Image.Index}{coordinates corresponding to the raster}
}
\description{
convert image coordinates from index to X-Y
}
