% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapSpectralSpecies.R
\name{Init.Kmeans}
\alias{Init.Kmeans}
\title{computes k-means from NbIter subsets taken from dataPCA}
\usage{
Init.Kmeans(dataPCA, Pix.Per.Iter, NbIter, nbclusters, nbCPU = 1)
}
\arguments{
\item{dataPCA}{initial dataset sampled from PCA image}

\item{Pix.Per.Iter}{number of pixels per iteration}

\item{NbIter}{number of iterations}

\item{nbclusters}{number of clusters used in kmeans}
}
\value{
list of centroids and parameters needed to center/reduce data
}
\description{
computes k-means from NbIter subsets taken from dataPCA
}
