% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kmeans_samples.R
\name{init_kmeans_samples}
\alias{init_kmeans_samples}
\title{Compute kmeans from random subset of pixels extracted from an image}
\usage{
init_kmeans_samples(
  rast_sample,
  output_dir,
  selected_bands = NULL,
  nb_clusters = 50,
  nb_iter = 10,
  Kmeans_info_save = NULL,
  clustering = "skm",
  algorithm = "Hartigan-Wong",
  nbCPU = 1,
  verbose = TRUE,
  progressbar = TRUE
)
}
\arguments{
\item{rast_sample}{data frame containing samples to use}

\item{output_dir}{character. Path for output directory}

\item{selected_bands}{numeric. bands selected from input_rast}

\item{nb_clusters}{numeric. number of clusters used in kmeans}

\item{nb_iter}{numeric. nb of iterations averaged to compute diversity indices}

\item{Kmeans_info_save}{character. path where to save Kmeans_info}

\item{clustering}{character. clustering method. default = skm (standard kmeans)}

\item{algorithm}{character. algorithm used in the kmeans clustering}

\item{nbCPU}{numeric. Number of CPUs available}

\item{verbose}{boolean. set true for messages}

\item{progressbar}{boolean. set true for progress bar during clustering}
}
\value{
Kmeans_info
}
\description{
Compute kmeans from random subset of pixels extracted from an image
}
