% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_outliers.R
\name{iqr_outliers}
\alias{iqr_outliers}
\title{This function computes interquartile range (IQR) criterion, which can be used
as a criterion for outlier detection
outlier_range_iqr <- c(range_IQR$1]- weightIQR x iqr, range_IQR(2)+ weightIQR x iqr)
with range_IQR(1) the first quartile and range_IQR(2) the third quartile}
\usage{
iqr_outliers(DistVal, weightIQR = 1.5)
}
\arguments{
\item{DistVal}{numeric. vector of distribution of values}

\item{weightIQR}{numeric. weighting factor applied to IQR to define
lower and upper boundaries for outliers}
}
\value{
outlier_range_iqr numeric. minimum and maximum values beyond which samples are considered outliers
}
\description{
This function computes interquartile range (IQR) criterion, which can be used
as a criterion for outlier detection
outlier_range_iqr <- c(range_IQR$1]- weightIQR x iqr, range_IQR(2)+ weightIQR x iqr)
with range_IQR(1) the first quartile and range_IQR(2) the third quartile
}
