% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR_SpatRaster.R
\name{IQR_SpatRaster}
\alias{IQR_SpatRaster}
\title{This function computes interquartile range (IQR) for a SpatRaster}
\usage{
IQR_SpatRaster(input_rast, weightIQR = 3)
}
\arguments{
\item{input_rast}{SpatRaster.}

\item{weightIQR}{numeric. weighting factor applied to IQR to define
lower and upper boundaries for outliers}
}
\value{
outlier_range_iqr numeric. range beyond which samples identified as outliers
}
\description{
This function computes interquartile range (IQR) for a SpatRaster
}
