% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{jaccard_dissimilarity}
\alias{jaccard_dissimilarity}
\title{Compute Jaccard Dissimilarity}
\usage{
jaccard_dissimilarity(A, B)
}
\arguments{
\item{A}{A numeric matrix of size N x p (N observations, p species).}

\item{B}{A numeric matrix of size M x p (M observations, p species).}
}
\value{
A dissimilarity matrix of size N x M.
}
\description{
Computes the Jaccard dissimilarity between two matrices.
}
\examples{
A <- matrix(c(1, 0, 1, 0, 1, 1), nrow = 2, ncol = 3)
B <- matrix(c(0, 1, 1, 1, 0, 1), nrow = 2, ncol = 3)
jaccard_dissimilarity(A, B)
}
