% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{jaccard_turnover}
\alias{jaccard_turnover}
\title{Compute Jaccard Turnover}
\usage{
jaccard_turnover(A, B)
}
\arguments{
\item{A}{A numeric matrix of size N x p.}

\item{B}{A numeric matrix of size M x p.}
}
\value{
A turnover matrix of size N x M.
}
\description{
Computes the Jaccard turnover between two matrices.
}
\examples{
A <- matrix(c(1, 0, 1, 0, 1, 1), nrow = 2, ncol = 3)
B <- matrix(c(0, 1, 1, 1, 0, 1), nrow = 2, ncol = 3)
jaccard_turnover(A, B)
}
