% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_iter.R
\name{kmeans_iter}
\alias{kmeans_iter}
\title{computes kmeans for an iteration in biodivMapR}
\usage{
kmeans_iter(Kmeans_Centroids, inputvals)
}
\arguments{
\item{Kmeans_Centroids}{numeric. coordinates of cluster centroids}

\item{inputvals}{numeric. chunk of input values to apply kmeans on}
}
\value{
SSchunk spectral species corresponding to the input chunk
}
\description{
computes kmeans for an iteration in biodivMapR
}
