% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_biodivMapR.R
\name{list_shp}
\alias{list_shp}
\title{Get list of shapefiles in a directory}
\usage{
list_shp(x)
}
\arguments{
\item{x}{character or list. Directory containing shapefiles}
}
\value{
list of shapefiles names
}
\description{
Get list of shapefiles in a directory
}
