% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tutorialdataset.R
\name{load_tutorialdataset}
\alias{load_tutorialdataset}
\title{loads tutorial dataset}
\usage{
load_tutorialdataset(tmpdir = NULL)
}
\arguments{
\item{tmpdir}{character. path for directory}
}
\value{
Sel_PC
}
\description{
loads tutorial dataset
}
