% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{Map.Alpha.Diversity}
\alias{Map.Alpha.Diversity}
\title{maps alpha diversity indicators  based on prior selection of PCs}
\usage{
Map.Alpha.Diversity(Input.Image.File, Output.Dir, Spatial.Unit,
  TypePCA = "SPCA", nbclusters = 50, MinSun = 0.25, pcelim = 0.02,
  Index.Alpha = "Shannon", FullRes = TRUE, LowRes = FALSE,
  MapSTD = FALSE, nbCPU = FALSE, MaxRAM = FALSE)
}
\arguments{
\item{Input.Image.File}{character. Path and name of the image to be processed.}

\item{Output.Dir}{character. Output directory.}

\item{Spatial.Unit}{numeric. Size of spatial units (in pixels) to compute diversity.}

\item{TypePCA}{character. Type of PCA (PCA, SPCA, NLPCA...).}

\item{nbclusters}{numeric. Number of clusters defined in k-Means.}

\item{MinSun}{numeric. Minimum proportion of sunlit pixels required to consider plot.}

\item{pcelim}{numeric. Minimum contribution (in \%) required for a spectral species.}

\item{Index.Alpha}{character. Either 'Shannon', 'Simpson' or 'Fisher'.}

\item{FullRes}{boolean. Full resolution.}

\item{LowRes}{boolean. Low resolution.}

\item{nbCPU}{numeric. Number of CPUs to use in parallel.}

\item{MaxRAM}{numeric. MaxRAM maximum size of chunk in GB to limit RAM allocation when reading image file.}
}
\description{
maps alpha diversity indicators  based on prior selection of PCs
}
