% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{Map.Alpha.Diversity.TestnbCluster}
\alias{Map.Alpha.Diversity.TestnbCluster}
\title{maps alpha diversity indicators  based on prior selection of PCs}
\usage{
Map.Alpha.Diversity.TestnbCluster(ImNames, Output.Dir, Spatial.Unit,
  TypePCA = "SPCA", nbclusters = 50, MinSun = 0.25, pcelim = 0.02,
  Index.Alpha = "Shannon", FullRes = TRUE, LowRes = FALSE,
  nbCPU = FALSE, MaxRAM = FALSE)
}
\arguments{
\item{ImNames}{Path and name of the images to be processed}

\item{Output.Dir}{output directory}

\item{Spatial.Unit}{dimensions of the spatial unit}

\item{TypePCA}{Type of PCA (PCA, SPCA, NLPCA...)}

\item{nbclusters}{number of clusters defined in k-Means}

\item{MinSun}{minimum proportion of sunlit pixels required to consider plot}

\item{pcelim}{minimum contribution (in %) required for a spectral species}

\item{Index.Alpha}{'Shannon', 'Simpson, 'Fisher'}
}
\description{
maps alpha diversity indicators  based on prior selection of PCs
}
