% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{Map.Beta.Diversity}
\alias{Map.Beta.Diversity}
\title{maps beta diversity indicator based on spectral species distribution}
\usage{
Map.Beta.Diversity(Input.Image.File, Output.Dir, Spatial.Unit,
  TypePCA = "SPCA", nbclusters = 50, Nb.Units.Ordin = 2000,
  MinSun = 0.25, pcelim = 0.02, scaling = "PCO", FullRes = TRUE,
  LowRes = FALSE, nbCPU = FALSE, MaxRAM = FALSE)
}
\arguments{
\item{Input.Image.File}{Path and name of the image to be processed}

\item{Output.Dir}{output directory}

\item{Spatial.Unit}{dimensions of the spatial unit}

\item{TypePCA}{Type of PCA (PCA, SPCA, NLPCA...)}

\item{nbclusters}{number of clusters defined in k-Means}

\item{Nb.Units.Ordin}{maximum number of spatial units to be processed in NMDS
--> 1000 will be fast but may not capture important patterns if large area
--> 4000 will be slow but may show better ability to capture landscape patterns}

\item{MinSun}{minimum proportion of sunlit pixels required to consider plot}

\item{pcelim}{minimum contribution (in %) required for a spectral species}
}
\description{
maps beta diversity indicator based on spectral species distribution
}
