% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapSpectralSpecies.R
\name{Map.Spectral.Species}
\alias{Map.Spectral.Species}
\title{maps spectral species based on PCA file computed previously}
\usage{
Map.Spectral.Species(Input.Image.File, Output.Dir, PCA.Files,
  TypePCA = "SPCA", nbclusters = 50, nbCPU = 1, MaxRAM = FALSE)
}
\arguments{
\item{Input.Image.File}{Path and name of the image to be processed}

\item{Output.Dir}{output directory}

\item{PCA.Files}{path for PCA file}

\item{TypePCA}{Type of PCA (PCA, SPCA, NLPCA...)}

\item{nbclusters}{number of clusters defined in k-Means}
}
\description{
maps spectral species based on PCA file computed previously
}
