% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{map_alpha_div}
\alias{map_alpha_div}
\title{maps alpha diversity indicators  based on prior selection of PCs}
\usage{
map_alpha_div(Input_Image_File, Output_Dir, window_size,
  TypePCA = "SPCA", nbclusters = 50, MinSun = 0.25, pcelim = 0.02,
  Index_Alpha = "Shannon", FullRes = TRUE, LowRes = FALSE,
  MapSTD = FALSE, nbCPU = FALSE, MaxRAM = FALSE)
}
\arguments{
\item{Input_Image_File}{character. Path and name of the image to be processed.}

\item{Output_Dir}{character. Output directory.}

\item{window_size}{numeric. Size of spatial units (in pixels) to compute diversity.}

\item{TypePCA}{character. Type of PCA (PCA, SPCA, NLPCA...).}

\item{nbclusters}{numeric. Number of clusters defined in k-Means.}

\item{MinSun}{numeric. Minimum proportion of sunlit pixels required to consider plot.}

\item{pcelim}{numeric. Minimum contribution (in \%) required for a spectral species.}

\item{Index_Alpha}{character. Either 'Shannon', 'Simpson' or 'Fisher'.}

\item{FullRes}{boolean. Full resolution.}

\item{LowRes}{boolean. Low resolution.}

\item{MapSTD}{boolean. map of standard deviation of the alpha diversity map (over repetitions)}

\item{nbCPU}{numeric. Number of CPUs to use in parallel.}

\item{MaxRAM}{numeric. MaxRAM maximum size of chunk in GB to limit RAM allocation when reading image file.}
}
\value{
None
}
\description{
maps alpha diversity indicators  based on prior selection of PCs
}
