% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{map_beta_div}
\alias{map_beta_div}
\title{maps beta diversity indicator based on spectral species distribution}
\usage{
map_beta_div(
  Input_Image_File = FALSE,
  Output_Dir = "",
  window_size = 10,
  TypePCA = "SPCA",
  nb_partitions = 20,
  nbclusters = 50,
  Nb_Units_Ordin = 2000,
  MinSun = 0.25,
  pcelim = 0.02,
  scaling = "PCO",
  FullRes = TRUE,
  LowRes = FALSE,
  nbCPU = 1,
  MaxRAM = 0.25,
  ClassifMap = FALSE
)
}
\arguments{
\item{Input_Image_File}{character. Path and name of the image to be processed.}

\item{Output_Dir}{character. Output directory.}

\item{window_size}{numeric. Dimensions of the spatial unit.}

\item{TypePCA}{character. Type of PCA (PCA, SPCA, NLPCA...).}

\item{nb_partitions}{numeric. Number of partitions (repetitions) to be computed then averaged.}

\item{nbclusters}{numeric. Number of clusters defined in k-Means.}

\item{Nb_Units_Ordin}{numeric. Maximum number of spatial units to be processed in NMDS.
--> 1000 will be fast but may not capture important patterns if large area
--> 4000 will be slow but may show better ability to capture landscape patterns}

\item{MinSun}{numeric. Minimum proportion of sunlit pixels required to consider plot.}

\item{pcelim}{numeric. Minimum contribution (in \%) required for a spectral species}

\item{scaling}{character. "PCO" or "NMDS".}

\item{FullRes}{boolean.}

\item{LowRes}{boolean.}

\item{nbCPU}{numeric. Number of CPUs to use in parallel.}

\item{MaxRAM}{numeric. MaxRAM maximum size of chunk in GB to limit RAM allocation when reading image file.}

\item{ClassifMap}{character. If FALSE, perform standard biodivMapR based on SpectralSpecies.
else corresponds to path for a classification map.}
}
\value{
None
}
\description{
maps beta diversity indicator based on spectral species distribution
}
