% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapFunctionalDiversity.R
\name{map_functional_div}
\alias{map_functional_div}
\title{maps functional diversity indicators based on prior selection of PCs}
\usage{
map_functional_div(
  Original_Image_File,
  Functional_File = FALSE,
  Selected_Features = FALSE,
  Output_Dir,
  window_size,
  TypePCA = "SPCA",
  MinSun = 0.25,
  FullRes = FALSE,
  LowRes = TRUE,
  MapSTD = TRUE,
  nbCPU = 1,
  MaxRAM = 0.25,
  SmoothImage = TRUE,
  FDmetric = c("FRic", "FEve", "FDiv")
)
}
\arguments{
\item{Original_Image_File}{character. Path and name of the original input image for biodivMapR.}

\item{Functional_File}{character. Path and name of the image processed to be used to compute functional diversity
--> can be PCA file with. if FALSE: using Original_Image_File}

\item{Selected_Features}{numeric. Contains features to be used from Input_Image_File. using all if FALSE}

\item{Output_Dir}{character. Output directory.}

\item{window_size}{numeric. Size of spatial units (in pixels) to compute diversity.}

\item{TypePCA}{character. Type of PCA (PCA, SPCA, NLPCA...).}

\item{MinSun}{numeric. Minimum proportion of sunlit pixels required to consider plot.}

\item{FullRes}{boolean. Full resolution.}

\item{LowRes}{boolean. Low resolution.}

\item{MapSTD}{boolean. map of standard deviation of the alpha diversity map (over repetitions)}

\item{nbCPU}{numeric. Number of CPUs to use in parallel.}

\item{MaxRAM}{numeric. MaxRAM maximum size of chunk in GB to limit RAM allocation when reading image file.}

\item{SmoothImage}{boolean. set TRUE if you want smooting filter applied to resulting diversity rasters}

\item{FDmetric}{character. Functional diversity metric}
}
\value{
None
}
\description{
maps functional diversity indicators based on prior selection of PCs
}
