% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapPartition.R
\name{map_partition_div}
\alias{map_partition_div}
\title{Partitioning of plant spectral diversity}
\usage{
map_partition_div(
  Original_Image_File,
  Partition_File = FALSE,
  Selected_Features = FALSE,
  Output_Dir,
  window_size,
  TypePCA = "SPCA",
  MinSun = 0.25,
  FullRes = TRUE,
  LowRes = FALSE,
  nbCPU = FALSE,
  MaxRAM = FALSE,
  SmoothImage = TRUE
)
}
\arguments{
\item{Original_Image_File}{character. Path and name of the original input image for biodivMapR.}

\item{Partition_File}{character. Path and name of the input image to perform diversity partitioning on}

\item{Selected_Features}{numeric. Contains features to be used from Input_Image_File. using all if FALSE}

\item{Output_Dir}{character. Output directory.}

\item{window_size}{numeric. Size of spatial units (in pixels) to compute diversity.}

\item{TypePCA}{character. Type of PCA (PCA, SPCA, NLPCA...).}

\item{MinSun}{numeric. Minimum proportion of sunlit pixels required to consider plot.}

\item{FullRes}{boolean. Full resolution.}

\item{LowRes}{boolean. Low resolution.}

\item{nbCPU}{numeric. Number of CPUs to use in parallel.}

\item{MaxRAM}{numeric. MaxRAM maximum size of chunk in GB to limit RAM allocation when reading image file.}

\item{SmoothImage}{boolean. set TRUE if you want smooting filter applied to resulting diversity rasters}
}
\value{
None
}
\description{
Partitioning of plant spectral diversity
}
