% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximize_sfs_criterion.R
\name{maximize_sfs_criterion}
\alias{maximize_sfs_criterion}
\title{define variable maximizing a criterion}
\usage{
maximize_sfs_criterion(CorrSFS, obs_criterion, nbvars2select)
}
\arguments{
\item{CorrSFS}{numeric}

\item{obs_criterion}{character. richness, shannon, simpson or BC}

\item{nbvars2select}{numeric}
}
\value{
SelVar numeric
}
\description{
define variable maximizing a criterion
}
