% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxRows_chunk.R
\name{maxRows_chunk}
\alias{maxRows_chunk}
\title{redefined chunks based on the max number of rows per chunk}
\usage{
maxRows_chunk(blk, maxRows = NULL)
}
\arguments{
\item{blk}{list. obtained from function terra::blocks}

\item{maxRows}{numeric. max number of rows in each block}
}
\value{
blk
}
\description{
redefined chunks based on the max number of rows per chunk
}
