% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Mean.Filter}
\alias{Mean.Filter}
\title{applies mean filter to an image}
\usage{
Mean.Filter(ImageInit, nbi, nbj, SizeFilt)
}
\arguments{
\item{ImageInit}{image defined as 2d matrix}

\item{nbi}{number of lines in image}

\item{nbj}{number of columns in image}

\item{SizeFilt}{size of the window to compute mean filter}
}
\value{
rank of all spectral bands of interest in the image and corresponding wavelength
}
\description{
applies mean filter to an image
}
