% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{mean_filter}
\alias{mean_filter}
\title{applies mean filter to an image}
\usage{
mean_filter(Image, SizeFilt, NA_remove = FALSE)
}
\arguments{
\item{Image}{numeric. image defined as 2d matrix}

\item{SizeFilt}{numeric. size of the window to compute mean filter}

\item{NA_remove}{boolean. Should NA be removed?}
}
\value{
rank of all spectral bands of interest in the image and corresponding wavelength
}
\description{
applies mean filter to an image
}
