% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PerformPCA.R
\name{minmax}
\alias{minmax}
\title{this function performs rescaling and
either defines min and max from each feature in a data set,
or applies the transformation based on a previously defined min and max}
\usage{
minmax(x, mode = "define", MinX = FALSE, MaxX = FALSE)
}
\arguments{
\item{x}{data matrix}

\item{mode}{'define' or 'apply'}

\item{MinX}{if 'apply'}

\item{MaxX}{if 'apply'}
}
\value{
rescaled data, min and max values
}
\description{
this function performs rescaling and
either defines min and max from each feature in a data set,
or applies the transformation based on a previously defined min and max
}
