% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnf.R
\name{mnf}
\alias{mnf}
\title{Function to perform MNF}
\usage{
mnf(X, coordPix = NULL, retx = TRUE)
}
\arguments{
\item{X}{numeric. matrix to apply MNF on}

\item{coordPix}{dataframe to compute noise, cf get_random_subset_from_image}

\item{retx}{boolean.}
}
\value{
results of MNF applied on matrix
}
\description{
Function to perform MNF
}
