% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_from_vrt.R
\name{mosaic_from_vrt}
\alias{mosaic_from_vrt}
\title{mosaic tiles defined in vrt and compress if possible}
\usage{
mosaic_from_vrt(vrt_path, mosaic_path)
}
\arguments{
\item{vrt_path}{character. path for vrt file}

\item{mosaic_path}{character. path for mosaic file}
}
\value{
none
}
\description{
mosaic tiles defined in vrt and compress if possible
}
