% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_tiles.R
\name{mosaic_tiles}
\alias{mosaic_tiles}
\title{produce a mosaic from a series of tiles}
\usage{
mosaic_tiles(
  pattern,
  dir_path,
  vrt_save,
  site_name = NULL,
  idx = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{pattern}{character. identify which files to mosaic}

\item{dir_path}{character. firectory where rasters to mosaic are stored}

\item{vrt_save}{character. where to save vrt}

\item{site_name}{character. name of the site}

\item{idx}{character. index if not provided as pattern}

\item{overwrite}{boolean}
}
\value{
mosaic_path path for the mosaic
}
\description{
produce a mosaic from a series of tiles
}
