% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbRows_chunk.R
\name{nbRows_chunk}
\alias{nbRows_chunk}
\title{adjusts number of rows from chunks}
\usage{
nbRows_chunk(blk, nbRows = NULL)
}
\arguments{
\item{blk}{list.}

\item{nbRows}{numeric}
}
\value{
blk
}
\description{
adjusts number of rows from chunks
}
