% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PerformPCA.R
\name{nlpca}
\alias{nlpca}
\title{Function to perform NLPCA}
\usage{
nlpca(DataSubset)
}
\arguments{
\item{DataSubset}{matrix to apply NLPCA on}
}
\value{
list of PCA parameters (PCs from X, mean, eigenvectors and values)
}
\description{
Function to perform NLPCA
}
