% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{noise}
\alias{noise}
\title{gets noise}
\usage{
noise(X, coordPix = NULL)
}
\arguments{
\item{X}{numeric. data matrix}

\item{coordPix}{numeric.}
}
\value{
rescaled data, min and max values
}
\description{
gets noise
}
