% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{normalize_SSD}
\alias{normalize_SSD}
\title{Performs normalization of spectral species distribution based on cumulative
sum of spectral species}
\usage{
normalize_SSD(SSDList, pcelim = 0.02)
}
\arguments{
\item{SSDList}{numeric. matrix corresponding to element of a list of SSD}

\item{pcelim}{numeric. Minimum contribution in percent required for a spectral species.}
}
\value{
SSD numeric. matrix corresponding to normalized spectral species distribution
}
\description{
Performs normalization of spectral species distribution based on cumulative
sum of spectral species
}
