% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{Ordination.Parallel}
\alias{Ordination.Parallel}
\title{applies results of ordination to full image based on nearest neighbors}
\usage{
Ordination.Parallel(id.sub, coordTotSort, SSD.Path, Sample.Sel,
  Beta.Ordination.sel, Nb.Units.Ordin, NbIter, nbclusters, pcelim)
}
\arguments{
\item{pcelim}{}
}
\value{
list of mean and SD of alpha diversity metrics
}
\description{
applies results of ordination to full image based on nearest neighbors
}
