% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{Ordination.to.NN}
\alias{Ordination.to.NN}
\title{identifies ordination coordinates based on nearest neighbors}
\usage{
Ordination.to.NN(Beta.Ordination.sel, SSD.Path, Sample.Sel, coordTotSort,
  NbIter, nbclusters, pcelim, nbCPU = FALSE)
}
\arguments{
\item{Beta.Ordination.sel}{ordination}

\item{SSD.Path}{ath for spectral species distribution file}

\item{Sample.Sel}{Samples selected during ordination}

\item{coordTotSort}{coordinates of sunlit spatial units}

\item{NbIter}{number of iterations}

\item{nbclusters}{number of clusters}

\item{pcelim}{number of CPUs available}

\item{nbCPU}{number of CPUs available}
}
\value{
Ordination.est coordinates of each spatial unit in ordination space
}
\description{
identifies ordination coordinates based on nearest neighbors
}
