% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{ordination_parallel}
\alias{ordination_parallel}
\title{applies results of ordination to full image based on nearest neighbors}
\usage{
ordination_parallel(
  id.sub,
  coordTotSort,
  SSD_Path,
  Sample_Sel,
  Beta_Ordination_sel,
  Nb_Units_Ordin,
  nb_partitions,
  nbclusters,
  pcelim,
  p = NULL
)
}
\arguments{
\item{id.sub}{numeric. ID of a subset of sunlit spatial units}

\item{coordTotSort}{numeric. coordinates of sunlit spatial units}

\item{SSD_Path}{character. path for spectral species distribution file}

\item{Sample_Sel}{numeric. Samples selected during ordination}

\item{Beta_Ordination_sel}{numeric. ordination of dissimilarity matrix for a selection of spatial units}

\item{Nb_Units_Ordin}{numeric. Maximum number of spatial units to be processed in NMDS.}

\item{nb_partitions}{numeric. Number of partitions (repetitions) to be computed then averaged.}

\item{nbclusters}{numeric. Number of clusters defined in k-Means.}

\item{pcelim}{numeric. Minimum contribution in percents required for a spectral species}

\item{p}{function.}
}
\value{
OutPut list of mean and SD of alpha diversity metrics
}
\description{
applies results of ordination to full image based on nearest neighbors
}
