% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{ordination_to_NN}
\alias{ordination_to_NN}
\title{Identifies ordination coordinates based on nearest neighbors}
\usage{
ordination_to_NN(
  SSD_subset,
  Beta_Ordination_sel,
  Sample_Sel,
  nb_partitions,
  nbclusters,
  pcelim = 0.02,
  nbCPU = 1,
  SamplesPerThread = 2000
)
}
\arguments{
\item{SSD_subset}{numeric. matrix corresponding to Spectral species distribution for a set of windows}

\item{Beta_Ordination_sel}{ordination of dissimilarity matrix for a selection of spatial units}

\item{Sample_Sel}{numeric. Samples selected during ordination}

\item{nb_partitions}{number of k-means then averaged}

\item{nbclusters}{number of clusters}

\item{pcelim}{numeric. Minimum contribution in percent required for a spectral species}

\item{nbCPU}{numeric. number of CPUs available}

\item{SamplesPerThread}{numeric. number of samples to be processed per thread for dissimilarity matrix}
}
\value{
Ordination_est coordinates of each spatial unit in ordination space
}
\description{
Identifies ordination coordinates based on nearest neighbors
}
