% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{ordination_to_NN_list}
\alias{ordination_to_NN_list}
\title{computes nearest neighbors of a SSD subset with samples used for the adjustment
of the ordination, then gets the coordinates of each window in the ordination space}
\usage{
ordination_to_NN_list(
  SSD_subset,
  Sample_Sel,
  Beta_Ordination_sel,
  nb_partitions,
  nbclusters,
  pcelim,
  p = NULL
)
}
\arguments{
\item{SSD_subset}{numeric. subset of spectral species distribution file}

\item{Sample_Sel}{numeric. Samples selected during ordination}

\item{Beta_Ordination_sel}{numeric. ordination of dissimilarity matrix for a selection of spatial units}

\item{nb_partitions}{numeric. Number of partitions (repetitions) to be computed then averaged.}

\item{nbclusters}{numeric. Number of clusters defined in k-Means.}

\item{pcelim}{numeric. Minimum contribution in percents required for a spectral species}

\item{p}{function.}
}
\value{
OutPut list of mean and SD of alpha diversity metrics
}
\description{
computes nearest neighbors of a SSD subset with samples used for the adjustment
of the ordination, then gets the coordinates of each window in the ordination space
}
