% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pco.R
\name{pco}
\alias{pco}
\title{computes PCoA using the cmdscale function
original source: package labdsv https://rdrr.io/cran/labdsv/src/R/pco.R}
\usage{
pco(dis, k = 3)
}
\arguments{
\item{dis}{matrix of BC dissimilarity between the kernels excluded from Ordination (rows)}

\item{k}{numeric. number of dimensions for PCoA}
}
\value{
list
}
\description{
computes PCoA using the cmdscale function
original source: package labdsv https://rdrr.io/cran/labdsv/src/R/pco.R
}
