% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_FilterData.R
\name{Perform.Radiometric.Filtering}
\alias{Perform.Radiometric.Filtering}
\title{Performs radiometric filtering based on three criteria: NDVI, NIR reflectance, Blue reflectance}
\usage{
Perform.Radiometric.Filtering(Image.Path, Mask.Path, Output.Dir,
  TypePCA = "SPCA", NDVI.Thresh = 0.5, Blue.Thresh = 500,
  NIR.Thresh = 1500, Blue = 480, Red = 700, NIR = 835)
}
\arguments{
\item{Image.Path}{Path of the image to be processed}

\item{Mask.Path}{Path of the mask corresponding to the image}

\item{Output.Dir}{output directory}

\item{TypePCA}{Type of PCA: "PCA" or "SPCA"}

\item{NDVI.Thresh}{NDVI threshold applied to produce a mask (select pixels with NDVI>NDVI.Thresh)}

\item{Blue.Thresh}{Blue threshold applied to produce a mask (select pixels with Blue refl < Blue.Thresh --> filter clouds) refl expected between 0 and 10000}

\item{NIR.Thresh}{NIR threshold applied to produce a mask (select pixels with NIR refl < NIR.Thresh) refl expected between 0 and 10000}

\item{Mask.Path}{}
}
\value{
ImPathShade = updated shademask file
}
\description{
Performs radiometric filtering based on three criteria: NDVI, NIR reflectance, Blue reflectance
}
