% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PerformPCA.R
\name{predict.from.NLPCA}
\alias{predict.from.NLPCA}
\title{applies NLPCA to data}
\usage{
\method{predict}{from.NLPCA}(Data, model.AutoEncod)
}
\arguments{
\item{Data}{matrix to apply NLPCA on}

\item{model.AutoEncod}{model defined previously}
}
\value{
Image.Chunk after NLPCA
}
\description{
applies NLPCA to data
}
