% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{prepare_HDR_SSD}
\alias{prepare_HDR_SSD}
\title{prepares Spectral species distribution (SSD) file and header}
\usage{
prepare_HDR_SSD(HDR_SS, SSD_Path, nbclusters, window_size)
}
\arguments{
\item{HDR_SS}{list. header of the spectral species file}

\item{SSD_Path}{character. path for spectral species file}

\item{nbclusters}{numeric. number of clusters}

\item{window_size}{numeric. window size}
}
\value{
HDR_SSD
}
\description{
prepares Spectral species distribution (SSD) file and header
}
