% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{prepare_HDR_Sunlit}
\alias{prepare_HDR_Sunlit}
\title{prepare for writing sunlit proportion file}
\usage{
prepare_HDR_Sunlit(HDR_SSD, Sunlit_Path)
}
\arguments{
\item{HDR_SSD}{list. header of the spectral species distribution file}

\item{Sunlit_Path}{character. path for spectral species file}
}
\value{
HDR_Sunlit
}
\description{
prepare for writing sunlit proportion file
}
