% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapSpectralSpecies.R
\name{print_error_message}
\alias{print_error_message}
\title{prints an error message if problems occur}
\usage{
print_error_message(def_error, optarg)
}
\arguments{
\item{def_error}{character. nature of the error}

\item{optarg}{character. optional additional info required for the error message}
}
\value{
none.
}
\description{
prints an error message if problems occur
}
