% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_biodivMapR.R
\name{projection.file}
\alias{projection.file}
\title{get projection of a raster or a vector}
\usage{
projection.file(file, type = "raster")
}
\arguments{
\item{file}{path for a raster or vector (shapefile)}

\item{type}{'raster' or 'vector'}
}
\value{
projection
}
\description{
get projection of a raster or a vector
}
