% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiometric_filtering.R
\name{radiometric_filtering}
\alias{radiometric_filtering}
\title{Performs radiometric filtering based on three criteria: NDVI, NIR reflectance, Blue reflectance}
\usage{
radiometric_filtering(
  input_raster_path,
  output_dir,
  input_rast_wl,
  input_mask_path = NULL,
  NDVI_Thresh = 0.65,
  Blue_Thresh = 500,
  NIR_Thresh = 1500,
  Blue = 480,
  Red = 670,
  NIR = 835,
  maxRows = 1000,
  filetype = "GTiff",
  maskfilename = "mask_update"
)
}
\arguments{
\item{input_raster_path}{character. Path of the image to be processed}

\item{output_dir}{character. Path for output directory}

\item{input_rast_wl}{numeric. spectral bands used to identify relevant bands}

\item{input_mask_path}{character. Path of the mask corresponding to the image}

\item{NDVI_Thresh}{numeric. NDVI threshold applied to produce a mask
(select pixels with NDVI>NDVI_Thresh)}

\item{Blue_Thresh}{numeric. Blue threshold applied to produce a mask
(select pixels with Blue refl < Blue_Thresh --> filter clouds)}

\item{NIR_Thresh}{numeric. NIR threshold applied to produce a mask
(select pixels with NIR refl > NIR_Thresh)}

\item{Blue}{numeric. central wl corresponding to the blue spectral band}

\item{Red}{numeric. central wl corresponding to the red spectral band}

\item{NIR}{numeric. central wl corresponding to the NIR spectral band}

\item{maxRows}{numeric. max number of rows in each block}

\item{filetype}{character. GDAL driver}

\item{maskfilename}{character. name of the updated mask file}
}
\value{
mask_update = updated mask file
}
\description{
Performs radiometric filtering based on three criteria: NDVI, NIR reflectance, Blue reflectance
}
