% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiometricfilter_chunk.R
\name{radiometricfilter_chunk}
\alias{radiometricfilter_chunk}
\title{Performs radiometric filtering based on three criteria: NDVI, NIR reflectance, Blue reflectance}
\usage{
radiometricfilter_chunk(blk, r_in, thresholds)
}
\arguments{
\item{blk}{character. Path of the image to be processed}

\item{r_in}{character. Path of the mask corresponding to the image}

\item{thresholds}{character. Path for output directory}
}
\value{
mask = updated mask values
}
\description{
Performs radiometric filtering based on three criteria: NDVI, NIR reflectance, Blue reflectance
}
