% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randperm.R
\name{randperm}
\alias{randperm}
\title{performs random permutation for k samples among the vector defined by a}
\usage{
randperm(a, k)
}
\arguments{
\item{a}{numeric. vctor of values}

\item{k}{numeric. number of samples}
}
\value{
list
}
\description{
performs random permutation for k samples among the vector defined by a
}
