% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_sample_sfs.R
\name{rast_sample_sfs}
\alias{rast_sample_sfs}
\title{extract informaton from raster to prepare for SFS}
\usage{
rast_sample_sfs(input_raster, input_mask, nb_pix, nb_iter)
}
\arguments{
\item{input_raster}{SpatRaster or list of SpatRaster}

\item{input_mask}{SpatRaster corresponding to mask}

\item{nb_pix}{numeric.}

\item{nb_iter}{numeric}
}
\value{
list
}
\description{
extract informaton from raster to prepare for SFS
}
