% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapSpectralSpecies.R
\name{RdistList}
\alias{RdistList}
\title{Compute distance between each pixel of input data and each of the nbClusters x nbIter centroids}
\usage{
RdistList(InputData, CentroidsArray, nbClusters, nbIter)
}
\arguments{
\item{nbIter}{}
}
\value{
ResDist
}
\description{
Compute distance between each pixel of input data and each of the nbClusters x nbIter centroids
}
