% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Read.Bin.Subset}
\alias{Read.Bin.Subset}
\title{reads a subset from a binary image}
\usage{
Read.Bin.Subset(Byte.Start, nbLines, lenBin, ImPath, ImBand, jpix,
  nbChannels, Image.Format)
}
\arguments{
\item{Byte.Start}{location of byte where to start reading in the image}

\item{nbLines}{number of lines to read}

\item{lenBin}{number of elements to read}

\item{ImPath}{path for the image}

\item{ImBand}{bands of interest}

\item{jpix}{number of columns in the image}

\item{nbChannels}{total number of channels in the image}

\item{Image.Format}{type of data (INT/FLOAT)}
}
\value{
data corresponding to the subset in original 3D format
}
\description{
reads a subset from a binary image
}
