% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{read.ENVI.header}
\alias{read.ENVI.header}
\title{Reads ENVI hdr file}
\usage{
read.ENVI.header(headerFpath)
}
\arguments{
\item{headerFpath}{Path of the hdr file}
}
\value{
list of the content of the hdr file
}
\description{
Reads ENVI hdr file
}
