% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Read.Image.Bands}
\alias{Read.Image.Bands}
\title{read specific image bands from image}
\usage{
Read.Image.Bands(ImPath, Header, ImBand)
}
\arguments{
\item{ImPath}{Path of the image to read}

\item{Header}{Header for the image}

\item{ImBand}{Bands to be read}
}
\value{
Image.Subset information corresponding to ImBand
}
\description{
read specific image bands from image
}
