% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Read.Image.Subset}
\alias{Read.Image.Subset}
\title{reads subset of an image}
\usage{
Read.Image.Subset(ImPath, HDR, Byte.Start, lenBin, nbLines, Image.Format,
  ImgFormat)
}
\arguments{
\item{ImPath}{path for the image}

\item{HDR}{header information corresponding to the image}

\item{Byte.Start}{location of byte where to start reading in the image}

\item{lenBin}{number of elements to read}

\item{nbLines}{number of lines to read}

\item{Image.Format}{type of data (INT/FLOAT)}

\item{ImgFormat}{should output be 2D or 3D (original image format)?}
}
\value{
data corresponding to the subset in original 3D format
}
\description{
reads subset of an image
}
