% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R,
%   R/Lib_Validation_DiversityMappR.R
\name{Read.Image.Zipfile}
\alias{Read.Image.Zipfile}
\title{This image reads information in a zipfile}
\usage{
Read.Image.Zipfile(Zip.Path)

Read.Image.Zipfile(Zip.Path)
}
\arguments{
\item{Zip.Path}{path for the zipped binary image file in .BIL format and}

\item{HDR}{Header corresponding}

\item{Zip.Path}{path for the zipped binary image file in .BIL format and}

\item{HDR}{Header corresponding}
}
\value{
Img.Data vecotr containing all image data

Img.Data vecotr containing all image data
}
\description{
This image reads information in a zipfile

This function reads information directly from a zipfile
}
