% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{read_BIL_image_subset}
\alias{read_BIL_image_subset}
\title{reads subset of an ENVI BIL image}
\usage{
read_BIL_image_subset(
  ImPath,
  HDR,
  Byte_Start,
  lenBin,
  nbLines,
  Image_Format,
  ImgFormat
)
}
\arguments{
\item{ImPath}{path for the image}

\item{HDR}{header information corresponding to the image}

\item{Byte_Start}{location of byte where to start reading in the image}

\item{lenBin}{number of elements to read}

\item{nbLines}{number of lines to read}

\item{Image_Format}{type of data (INT/FLOAT)}

\item{ImgFormat}{should output be 2D or 3D (original image format)?}
}
\value{
data corresponding to the subset in original 3D format
}
\description{
reads subset of an ENVI BIL image
}
