% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{read_bin_subset}
\alias{read_bin_subset}
\title{reads a subset from a binary image}
\usage{
read_bin_subset(
  Byte_Start,
  nbLines,
  lenBin,
  ImPath,
  ImBand,
  jpix,
  nbChannels,
  Image_Format
)
}
\arguments{
\item{Byte_Start}{numeric. location of byte where to start reading in the image}

\item{nbLines}{numeric. number of lines to read}

\item{lenBin}{numeric. number of elements to read}

\item{ImPath}{character. path for the image}

\item{ImBand}{numeric. bands of interest}

\item{jpix}{numeric. number of columns in the image}

\item{nbChannels}{numeric. total number of channels in the image}

\item{Image_Format}{character. type of data (INT/FLOAT)}
}
\value{
data corresponding to the subset in original 3D format
}
\description{
reads a subset from a binary image
}
