% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_envi_header.R
\name{read_envi_header}
\alias{read_envi_header}
\title{Reads ENVI hdr file}
\usage{
read_envi_header(hdr_path)
}
\arguments{
\item{hdr_path}{character. Path of the hdr file}
}
\value{
list of the content of the hdr file
}
\description{
Reads ENVI hdr file
}
