% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{read_image_subset}
\alias{read_image_subset}
\title{reads subset of lines from an image}
\usage{
read_image_subset(ImPath, HDR, Line_Start, Lines_To_Read, ImgFormat = "3D")
}
\arguments{
\item{ImPath}{path for the image}

\item{HDR}{header information corresponding to the image}

\item{Line_Start}{which line to start reading}

\item{Lines_To_Read}{number of lines to read}

\item{ImgFormat}{character. is it a matrix (2D) or a raster (3D)?}
}
\value{
data corresponding to the subset in original 3D format
}
\description{
reads subset of lines from an image
}
