% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ContinuumRemoval.R
\name{repmat}
\alias{repmat}
\title{R equivalent of repmat (matlab)}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{initial matrix}

\item{m}{nb of replications in row dimension}

\item{n}{nb of replications in column dimension}
}
\value{
matrix with values replicated and tiles
}
\description{
R equivalent of repmat (matlab)
}
