% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_DiversityMappR.R
\name{Reproject.Vector}
\alias{Reproject.Vector}
\title{reprojects a vector file and saves it}
\usage{
Reproject.Vector(Initial.File, Projection, Reprojected.File)
}
\arguments{
\item{Initial.File}{path for a shapefile to be reprojected}

\item{Projection}{projection to be applied to Initial.File}

\item{Reprojected.File}{path for the reprojected shapefile}
}
\description{
reprojects a vector file and saves it
}
