% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_biodivMapR.R
\name{reproject_vector}
\alias{reproject_vector}
\title{reprojects a vector file and saves it}
\usage{
reproject_vector(Initial.File, Projection, Reprojected.File)
}
\arguments{
\item{Initial.File}{path for a shapefile to be reprojected}

\item{Projection}{projection to be applied to Initial.File}

\item{Reprojected.File}{path for the reprojected shapefile}
}
\value{
None
}
\description{
reprojects a vector file and saves it
}
