% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Revert.Resolution.HDR}
\alias{Revert.Resolution.HDR}
\title{revert resolution in a HDR file}
\usage{
Revert.Resolution.HDR(HDR, Spatial.Unit)
}
\arguments{
\item{HDR}{information read from a header file}

\item{Spatial.Unit}{multiplying factor for initial resolution}
}
\value{
updated HDR information
}
\description{
revert resolution in a HDR file
}
