% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{revert_resolution_HDR}
\alias{revert_resolution_HDR}
\title{revert resolution in a HDR file}
\usage{
revert_resolution_HDR(HDR, window_size)
}
\arguments{
\item{HDR}{information read from a header file}

\item{window_size}{numeric. multiplying factor for initial resolution}
}
\value{
updated HDR information
}
\description{
revert resolution in a HDR file
}
