% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_invariant_bands.R
\name{rm_invariant_bands}
\alias{rm_invariant_bands}
\title{remove constant bands}
\usage{
rm_invariant_bands(DataMatrix, Spectral)
}
\arguments{
\item{DataMatrix}{numeric. each variable is a column}

\item{Spectral}{list. summary of spectral information: which spectral bands selected from initial data}
}
\value{
updated DataMatrix and Spectral
}
\description{
remove constant bands
}
