% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_to_linear.R
\name{row_to_linear}
\alias{row_to_linear}
\title{get max index for each row and convert into linear index}
\usage{
row_to_linear(MM, nbi, nbj)
}
\arguments{
\item{MM}{numeric.}

\item{nbi}{numeric.}

\item{nbj}{numeric.}
}
\value{
MaxCont
}
\description{
get max index for each row and convert into linear index
}
