% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ContinuumRemoval.R
\name{RowToLinear}
\alias{RowToLinear}
\title{get max index for each row and convert into linear index}
\usage{
RowToLinear(MM, nbi, nbj)
}
\arguments{
\item{MM}{numeric.}

\item{nbi}{numeric.}

\item{nbj}{numeric.}
}
\value{
MaxCont
}
\description{
get max index for each row and convert into linear index
}
