% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_biodivMapR_plot.R
\name{run_biodivMapR_plot}
\alias{run_biodivMapR_plot}
\title{apply biodivMapR to a set of plots identified by a field '\emph{id}'
produced with preprocS2 function get_s2_tiling}
\usage{
run_biodivMapR_plot(
  id,
  feature_dir,
  mask_dir = NULL,
  list_features,
  Kmeans_info,
  Beta_info,
  output_dir,
  selected_bands = NULL,
  window_size,
  alpha_metrics = "shannon",
  Hill_order = 1,
  beta_metrics = "bray",
  fd_metrics = NULL,
  pcelim = 0.02,
  maxRows = NULL,
  nbCPU = 1,
  min_sun = 0.25,
  filetype = "GTiff",
  moving_window = FALSE,
  buffer_mw = TRUE,
  p = NULL,
  ...
)
}
\arguments{
\item{id}{character. ID for plot}

\item{feature_dir}{directory where spectral indices are for each plot}

\item{mask_dir}{character.}

\item{list_features}{character.}

\item{Kmeans_info}{list. kmeans description obtained from function get_kmeans}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics}

\item{output_dir}{path where to save outputs}

\item{selected_bands}{numeric. bands selected from input data}

\item{window_size}{numeric. window size for square plots}

\item{alpha_metrics}{list. alpha diversity metrics}

\item{Hill_order}{numeric. Hill order}

\item{beta_metrics}{list. beta diversity metrics}

\item{fd_metrics}{character. list of functional metrics}

\item{pcelim}{numeric. min proportion of pix to consider spectral species}

\item{maxRows}{numeric. maximum number or rows to process}

\item{nbCPU}{numeric. Number of CPUs available}

\item{min_sun}{numeric. minimum amount of sunlit pixels in the plots}

\item{filetype}{character. gdal driver for output raster}

\item{moving_window}{boolean. should process be moving window (much longer)}

\item{buffer_mw}{boolean. should buffer be applied on moving windows?}

\item{p}{list. progressbar}

\item{...}{list. additional parameters}
}
\value{
none
}
\description{
apply biodivMapR to a set of plots identified by a field '\emph{id}'
produced with preprocS2 function get_s2_tiling
}
