% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{RW_bytes}
\alias{RW_bytes}
\title{identifies bytes where to read or write for each piece of an image in a given file}
\usage{
RW_bytes(SeqRW, piece)
}
\arguments{
\item{SeqRW}{list. coordinates corresponding to spectral species file}

\item{piece}{numeric. identifier of the piece of image}
}
\value{
RW0 list of bytes coordinates for each read and write
}
\description{
identifies bytes where to read or write for each piece of an image in a given file
}
