% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_exact_raster.R
\name{sample_exact_raster}
\alias{sample_exact_raster}
\title{sample exact number of pixels from a raster}
\usage{
sample_exact_raster(
  extent_area = extent_area,
  nb_samples = nb_samples,
  input_rast = input_rast,
  input_mask = input_mask
)
}
\arguments{
\item{extent_area}{extent}

\item{nb_samples}{numeric. number of samples to be extracted}

\item{input_rast}{SpatRaster. raster to extract data from}

\item{input_mask}{SpatRaster. mask corresponding to raster to extract data from}
}
\value{
rast_sample dataframe. pixel/plot info extracted from input_rast
}
\description{
sample exact number of pixels from a raster
}
