% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_plots_alpha.R
\name{sample_from_plots_alpha}
\alias{sample_from_plots_alpha}
\title{get samples for alpha diversity mapping}
\usage{
sample_from_plots_alpha(
  feature_dir,
  list_features,
  plot_names,
  nbCPU = 1,
  nb_pix_valid,
  mask_dir = NULL,
  nb_samples_alpha = 1e+05,
  method = "regular"
)
}
\arguments{
\item{feature_dir}{character. directory where features to be used by biodivMapR are stored.}

\item{list_features}{list.}

\item{plot_names}{character. plot names}

\item{nbCPU}{numeric.}

\item{nb_pix_valid}{list.}

\item{mask_dir}{character.}

\item{nb_samples_alpha}{numeric.}

\item{method}{character. method for terra::spatSample ('random' or 'regular')}
}
\value{
samples_alpha_terra
}
\description{
get samples for alpha diversity mapping
}
