% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_raster.R
\name{sample_from_raster}
\alias{sample_from_raster}
\title{sample pixels or plots from raster data}
\usage{
sample_from_raster(
  extent_area,
  nbSamples,
  input_rast,
  input_mask = NULL,
  window_size = NULL,
  capstyle = "square"
)
}
\arguments{
\item{extent_area}{.}

\item{nbSamples}{numeric. number of samples to be extracted}

\item{input_rast}{SpatRaster. raster to extract data from}

\item{input_mask}{SpatRaster. mask corresponding to raster to extract data from}

\item{window_size}{numeric. window size for square plots}

\item{capstyle}{character. shape of the plot, see terra::buffer for info on options for capstyle}
}
\value{
rast_sample dataframe. pixel/plot info extracted from input_rast
}
\description{
sample pixels or plots from raster data
}
