% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_plots_from_raster.R
\name{sample_plots_from_raster}
\alias{sample_plots_from_raster}
\title{sample pixels or plots from raster data}
\usage{
sample_plots_from_raster(
  extent_area,
  nbSamples,
  input_rast,
  MinSun = 0.25,
  input_mask = NULL,
  window_size = NULL
)
}
\arguments{
\item{extent_area}{.}

\item{nbSamples}{numeric. number of samples to be extracted}

\item{input_rast}{SpatRaster. raster to extract data from}

\item{MinSun}{numeric. minimum amount of sunlit pixels in the plots}

\item{input_mask}{SpatRaster. mask corresponding to raster to extract data from}

\item{window_size}{numeric. window size for square plots}
}
\value{
rast_sample dataframe. pixel/plot info extracted from input_rast
}
\description{
sample pixels or plots from raster data
}
