% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_diversity_maps_mw.R
\name{save_diversity_maps_mw}
\alias{save_diversity_maps_mw}
\title{save diversity maps as raster data once moving window process done}
\usage{
save_diversity_maps_mw(
  input_raster_path,
  ab_div_metrics,
  alpha_metrics = "shannon",
  Hill_order = 1,
  Beta_info = NULL,
  beta_metrics = "bray",
  fd_metrics = NULL,
  input_rast,
  output_dir,
  output_raster_name = NULL,
  window_size,
  filetype = "GTiff",
  buffer_mw = TRUE
)
}
\arguments{
\item{input_raster_path}{list. list of paths corresponding to input rasters}

\item{ab_div_metrics}{list produced from get_raster_diversity}

\item{alpha_metrics}{list. alpha diversity metrics: richness, shannon, simpson, hill}

\item{Hill_order}{numeric. Hill order}

\item{Beta_info}{list.}

\item{beta_metrics}{list beta diversity metrics:}

\item{fd_metrics}{character. list of functional metrics}

\item{input_rast}{list. path for input rasters}

\item{output_dir}{character.}

\item{output_raster_name}{character.}

\item{window_size}{numeric. window size for square plots}

\item{filetype}{character. GDAL driver}

\item{buffer_mw}{boolean. should buffer be applied on moving windows?}
}
\value{
diversity_maps character. path for diversity maps
}
\description{
save diversity maps as raster data once moving window process done
}
