% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_alphabeta.R
\name{scatter_alphabeta}
\alias{scatter_alphabeta}
\title{produces a figure summarizing alpha and beta diversity on scatterplots}
\usage{
scatter_alphabeta(
  alpha,
  PCoA_1,
  PCoA_2,
  PCoA_3,
  classes,
  filename = NULL,
  palette = "Set1"
)
}
\arguments{
\item{alpha}{numeric. alpha diversity estimated}

\item{PCoA_1}{numeric. first component obtained from PCoA}

\item{PCoA_2}{numeric. second component obtained from PCoA}

\item{PCoA_3}{numeric. third component obtained from PCoA}

\item{classes}{character. classes of vegetation types/species communities}

\item{filename}{character. path where to save figure}

\item{palette}{character. palette to be used.
--> either name for a palette in RColorBrewer::brewer.pal or set of colors}
}
\value{
none
}
\description{
produces a figure summarizing alpha and beta diversity on scatterplots
}
