% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PerformPCA.R
\name{Select.Components}
\alias{Select.Components}
\title{Check if principal components are properly selected as expected by the method}
\usage{
Select.Components(Input.Image.File, Output.Dir, PCA.Files,
  TypePCA = "SPCA", File.Open = FALSE)
}
\arguments{
\item{Input.Image.File}{path for image to be processed}

\item{Output.Dir}{output directory}

\item{PCA.Files}{path of PCA files}

\item{TypePCA}{Type of PCA: "PCA" or "SPCA"}
}
\value{
nothing
}
\description{
Check if principal components are properly selected as expected by the method
}
