% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PerformPCA.R
\name{select_PCA_components}
\alias{select_PCA_components}
\title{Check if principal components are properly selected as expected by the method}
\usage{
select_PCA_components(Input_Image_File, Output_Dir, PCA_Files,
  TypePCA = "SPCA", File_Open = FALSE)
}
\arguments{
\item{Input_Image_File}{path for image to be processed}

\item{Output_Dir}{output directory}

\item{PCA_Files}{path of PCA files}

\item{TypePCA}{Type of PCA: "PCA" or "SPCA"}

\item{File_Open}{Boolean. automatically opens file if TRUE}
}
\value{
nothing
}
\description{
Check if principal components are properly selected as expected by the method
}
