% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options_biodivMapR.R
\name{set_options_biodivMapR}
\alias{set_options_biodivMapR}
\title{set options}
\usage{
set_options_biodivMapR(fun, options = NULL)
}
\arguments{
\item{fun}{character. name of the function which has optional parameters}

\item{options}{list. including
\itemize{
\item nb_clusters numeric. number of clusters
\item nb_samples_alpha numeric. number of samples to compute alpha diversity
\item nb_samples_beta numeric. number of samples to compute beta diversity
\item alphametrics character.
\item Hill_order numeric.
\item FDmetric character.
\item nb_iter numeric. Number of iterations required to compute diversity
\item pcelim numeric. minimum proportion of pixels to consider spectral species
\item maxRows numeric. maximum number of rows
\item moving_window boolean. should moving window be used?
\item mosaic_output boolean. set TRUE if outputs need to be mosaiced
\item weightIRQ numeric. IQR applied to filter out features to be used
}}
}
\value{
options with default values when missing
}
\description{
set options
}
