% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{sorensen_dissimilarity}
\alias{sorensen_dissimilarity}
\title{Compute Sorensen Dissimilarity}
\usage{
sorensen_dissimilarity(A, B)
}
\arguments{
\item{A}{A numeric matrix of size N x p.}

\item{B}{A numeric matrix of size M x p.}
}
\value{
A dissimilarity matrix of size N x M.
}
\description{
Computes the Sorensen dissimilarity between two matrices.
}
\examples{
A <- matrix(c(1, 0, 1, 0, 1, 1), nrow = 2, ncol = 3)
B <- matrix(c(0, 1, 1, 1, 0, 1), nrow = 2, ncol = 3)
sorensen_dissimilarity(A, B)
}
