% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specdiv_functions.R
\name{specdiv}
\alias{specdiv}
\title{Partitioning spectral diversity ----}
\usage{
specdiv(cube, fact = 40, prop = 0.5, n = 1)
}
\arguments{
\item{cube}{dataframe. data cube corresponding to raster}

\item{fact}{numeric. window size}

\item{prop}{numeric. proportion of valid pixels in a window under which window is discarded}

\item{n}{numeric.}
}
\value{
out list. includes rasters and mean alpha, beta and gamma
}
\description{
Partitioning spectral diversity ----
}
