% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralspecies_per_polygon.R
\name{spectralspecies_per_polygon}
\alias{spectralspecies_per_polygon}
\title{get spectral species corresponding to polygns in a SpatVector object}
\usage{
spectralspecies_per_polygon(
  SpatVector,
  input_rast,
  Kmeans_info,
  SelectBands = NULL,
  input_mask = NULL,
  MinSun = 0.25
)
}
\arguments{
\item{SpatVector}{SpatVector.}

\item{input_rast}{SpatRaster.}

\item{Kmeans_info}{list. kmeans description obtained from function get_kmeans}

\item{SelectBands}{numeric. bands selected from input data}

\item{input_mask}{SpatRaster.}

\item{MinSun}{numeric. minimum amount of sunlit pixels in the plots}
}
\value{
list
}
\description{
get spectral species corresponding to polygns in a SpatVector object
}
