% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralspecies_per_rast_sample.R
\name{spectralspecies_per_rast_sample}
\alias{spectralspecies_per_rast_sample}
\title{get spectral species corresponding to raster sample extracted with extract}
\usage{
spectralspecies_per_rast_sample(
  Kmeans_info,
  rast_sample,
  selected_bands = NULL
)
}
\arguments{
\item{Kmeans_info}{list. kmeans description obtained from function get_kmeans}

\item{rast_sample}{dataframe.}

\item{selected_bands}{numeric. bands selected from input data}
}
\value{
list
}
\description{
get spectral species corresponding to raster sample extracted with extract
}
