% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Split.Image}
\alias{Split.Image}
\title{defines the number of pieces resulting from image split}
\usage{
Split.Image(HDR, LimitSizeGb = FALSE)
}
\arguments{
\item{HDR}{information extracted from a header}

\item{LimitSizeGb}{maximum size of individual pieces of an image (in Gb)}
}
\value{
nbPieces number of pieces
}
\description{
defines the number of pieces resulting from image split
}
