% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Split.Pixel.Samples}
\alias{Split.Pixel.Samples}
\title{splits a set of pixels to be sampled in an image based on number of lines, not number of samples}
\usage{
Split.Pixel.Samples(coordPix, Lines.Per.Read)
}
\arguments{
\item{coordPix}{coordinates of pixels to sample}

\item{Lines.Per.Read}{max number of lines per read for memory concerns}
}
\value{
coordPix.List list of pixel coordinates
}
\description{
splits a set of pixels to be sampled in an image based on number of lines, not number of samples
}
