% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_line.R
\name{split_line}
\alias{split_line}
\title{split information from a line}
\usage{
split_line(x, separator, trim.blank = TRUE)
}
\arguments{
\item{x}{character.}

\item{separator}{character}

\item{trim.blank}{boolean.}
}
\value{
list.
}
\description{
split information from a line
}
