% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{split_pixel_samples}
\alias{split_pixel_samples}
\title{splits a set of pixels to be sampled in an image based on number of lines, not number of samples}
\usage{
split_pixel_samples(coordPix, Lines_Per_Read)
}
\arguments{
\item{coordPix}{numeric. coordinates of pixels to sample}

\item{Lines_Per_Read}{numeric. max number of lines per read for memory concerns}
}
\value{
coordPix_List list of pixel coordinates
}
\description{
splits a set of pixels to be sampled in an image based on number of lines, not number of samples
}
