% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specdiv_functions.R
\name{sum_squares_beta}
\alias{sum_squares_beta}
\title{SS beta ----}
\usage{
sum_squares_beta(Y, m)
}
\arguments{
\item{Y}{numeric. data matrix}

\item{m}{numeric. unused}
}
\value{
out
}
\description{
SS beta ----
}
