% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Update.Shademask}
\alias{Update.Shademask}
\title{updates an existing shade mask}
\usage{
Update.Shademask(ImPathShade, Header, Mask, ImPathShade.Update)
}
\arguments{
\item{ImPathShade}{original shade mask (may not exist)}

\item{Header}{header correpondingproviding general info about data format}

\item{Mask}{data to be used in the mask}

\item{ImPathShade.Update}{path for teh updated shade mask}
}
\value{
ImPathShade.Update
}
\description{
updates an existing shade mask
}
