% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mask_from_tiles.R
\name{update_mask_from_tiles}
\alias{update_mask_from_tiles}
\title{update mask based on IQR for a series of rasters
produced with preprocS2 function get_s2_tiling}
\usage{
update_mask_from_tiles(plotID, listfiles, iqr_si, mask_dir, p = NULL)
}
\arguments{
\item{plotID}{character.}

\item{listfiles}{character.}

\item{iqr_si}{character.}

\item{mask_dir}{list.}

\item{p}{numeric.}
}
\value{
filename
}
\description{
update mask based on IQR for a series of rasters
produced with preprocS2 function get_s2_tiling
}
