% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{update_shademask}
\alias{update_shademask}
\title{updates an existing mask}
\usage{
update_shademask(MaskPath, HDR, Mask, MaskPath_Update)
}
\arguments{
\item{MaskPath}{character. path for original mask (may not exist)}

\item{HDR}{list. header correpondingproviding general info about data format}

\item{Mask}{numeric. data to be used in the mask}

\item{MaskPath_Update}{character. path for the updated mask}
}
\value{
MaskPath_Update
}
\description{
updates an existing mask
}
