% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{VectorInRasterFootprint}
\alias{VectorInRasterFootprint}
\title{This function identify plots from a shapefile matching with a raster footprint}
\usage{
VectorInRasterFootprint(
  PathVect,
  PathRaster,
  buffer = NULL,
  PathVectOut = NULL,
  WriteShp = TRUE
)
}
\arguments{
\item{PathVect}{character. path for input vector file}

\item{PathRaster}{character. path for a raster file}

\item{buffer}{numeric. buffer to be applied around the original points/polygons, in meters}

\item{PathVectOut}{character. path for output vector file}

\item{WriteShp}{boolean. Set TRUE to write file}
}
\value{
vectSP list. OGR object corresponding to final vector data
}
\description{
This function identify plots from a shapefile matching with a raster footprint
}
