% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_coords_nn.R
\name{weighted_coords_nn}
\alias{weighted_coords_nn}
\title{Compute weighted coordinates of a spatial unit based on nearest neighbors
used during PCoA}
\usage{
weighted_coords_nn(nn, knn, pcoa_train)
}
\arguments{
\item{nn}{list. coordinates & ID of nearest neighbors from pcoa_train matrix}

\item{knn}{number of neighbors}

\item{pcoa_train}{PCoA coordinates of reference samples}
}
\value{
ordin_est estimated PCoA position based on nearest neighbors from
PCoA space defined by training set
}
\description{
Compute weighted coordinates of a spatial unit based on nearest neighbors
used during PCoA
}
