% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedCoordsNN.R
\name{WeightedCoordsNN}
\alias{WeightedCoordsNN}
\title{Compute the weighted coordinates of a spatial unit based on nearest neighbors used during PCoA}
\usage{
WeightedCoordsNN(NN, knn, PCoA_train)
}
\arguments{
\item{NN}{list. coordinates and ID of nearest neighbors from PCoA_train matrix}

\item{knn}{number of neighbors}

\item{PCoA_train}{PCoA coordinates of reference samples}
}
\value{
estimated NMDS position based on nearest neighbors from NMDS
}
\description{
Compute the weighted coordinates of a spatial unit based on nearest neighbors used during PCoA
}
