% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Where.To.Read.Kernel}
\alias{Where.To.Read.Kernel}
\title{defines which byte should be read for each part of an image split in nbPieces}
\usage{
Where.To.Read.Kernel(HDR, nbPieces, SE.Size)
}
\arguments{
\item{HDR}{header info}

\item{nbPieces}{number of pieces resulting from image split}
}
\value{
location of the bytes corresponding to beginning and end of each piece, and corresponding number of lines
}
\description{
defines which byte should be read for each part of an image split in nbPieces
}
