% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Where.To.Write.Kernel}
\alias{Where.To.Write.Kernel}
\title{defines which byte should be written for each part of an image split in nbPieces}
\usage{
Where.To.Write.Kernel(HDR.SS, HDR.SSD, nbPieces, SE.Size)
}
\arguments{
\item{HDR.SS}{header info for SpectralSpecies file}

\item{HDR.SSD}{header info for SpectralSpecies_Distribution file}

\item{nbPieces}{number of pieces resulting from image split}

\item{SE.Size}{}
}
\value{
location of the bytes corresponding to beginning and end of each piece, and corresponding number of lines
}
\description{
defines which byte should be written for each part of an image split in nbPieces
}
