% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{where_to_read}
\alias{where_to_read}
\title{defines which byte should be read for each part of an image split in nbPieces}
\usage{
where_to_read(HDR, nbPieces)
}
\arguments{
\item{HDR}{list. header info}

\item{nbPieces}{numeric. number of pieces resulting from image split}
}
\value{
location of the bytes corresponding to beginning and end of each piece, and corresponding number of lines
}
\description{
defines which byte should be read for each part of an image split in nbPieces
}
