% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{where_to_write_kernel}
\alias{where_to_write_kernel}
\title{defines which byte should be written for each part of an image split in nbPieces}
\usage{
where_to_write_kernel(HDR_SS, HDR_SSD, nbPieces, SE_Size)
}
\arguments{
\item{HDR_SS}{list. header info for SpectralSpecies file}

\item{HDR_SSD}{list. header info for SpectralSpecies_Distribution file}

\item{nbPieces}{numeric. number of pieces resulting from image split}

\item{SE_Size}{numeric.}
}
\value{
location of the bytes corresponding to beginning and end of each piece, and corresponding number of lines
}
\description{
defines which byte should be written for each part of an image split in nbPieces
}
