% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_alpha_beta.R
\name{which_alpha_beta}
\alias{which_alpha_beta}
\title{identifies which alpha/beta metrics should be used for next processes}
\usage{
which_alpha_beta(obs_criterion)
}
\arguments{
\item{obs_criterion}{character. richness, shannon, simpson, hill, BC}
}
\value{
list including alpha_metrics and beta
}
\description{
identifies which alpha/beta metrics should be used for next processes
}
