% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{write.ENVI.header}
\alias{write.ENVI.header}
\title{writes ENVI hdr file}
\usage{
write.ENVI.header(header, headerFpath)
}
\arguments{
\item{header}{content to be written}

\item{headerFpath}{Path of the hdr file}
}
\description{
writes ENVI hdr file
}
