% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{Write.Image.Alpha}
\alias{Write.Image.Alpha}
\title{Writes image of alpha diversity indicator (1 band) and smoothed alpha diversity}
\usage{
Write.Image.Alpha(Image, HDR.SSD, ImagePath, Spatial.Unit, Index,
  FullRes = TRUE, LowRes = FALSE)
}
\arguments{
\item{Image}{2D matrix of image to be written}

\item{HDR.SSD}{hdr template derived from SSD to modify}

\item{ImagePath}{path of image file to be written}

\item{Spatial.Unit}{spatial units use dto compute diversiy (in pixels)}

\item{Index}{name of the index (eg. Shannon)}

\item{FullRes}{should full resolution image be written (original pixel size)}

\item{LowRes}{should low resolution image be written (one value per spatial unit)}
}
\description{
Writes image of alpha diversity indicator (1 band) and smoothed alpha diversity
}
