% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{Write.Image.Beta}
\alias{Write.Image.Beta}
\title{Writes image of beta diversity indicator (3 bands) resulting from BC + NMDS}
\usage{
Write.Image.Beta(Image, HDR.SSD, ImagePath, Spatial.Unit, FullRes = TRUE,
  LowRes = FALSE)
}
\arguments{
\item{Image}{2D matrix of image to be written}

\item{HDR.SSD}{hdr template derived from SSD to modify}

\item{ImagePath}{path of image file to be written}

\item{Spatial.Unit}{spatial units use dto compute diversiy (in pixels)}

\item{FullRes}{should full resolution image be written (original pixel size)}

\item{LowRes}{should low resolution image be written (one value per spatial unit)}
}
\description{
Writes image of beta diversity indicator (3 bands) resulting from BC + NMDS
}
