% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Write_Big_Image}
\alias{Write_Big_Image}
\title{write an image which size is > 2**31-1}
\usage{
Write_Big_Image(ImgWrite, ImagePath, HDR, Image_Format)
}
\arguments{
\item{ImgWrite}{numeric. Image as array}

\item{ImagePath}{character. Path where the image should be written}

\item{HDR}{list. Image header}

\item{Image_Format}{list. description of data format corresponding to ENVI type}
}
\value{
None
}
\description{
write an image which size is > 2**31-1
}
