% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{write_ENVI_header}
\alias{write_ENVI_header}
\title{writes ENVI hdr file}
\usage{
write_ENVI_header(HDR, HDRpath)
}
\arguments{
\item{HDR}{content to be written}

\item{HDRpath}{Path of the hdr file}
}
\value{

}
\description{
writes ENVI hdr file
}
