% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{Write_Image_NativeRes}
\alias{Write_Image_NativeRes}
\title{write an image resulting from "window processing" at native spatial resolution
(assuming square windows & origin at top left corner)}
\usage{
Write_Image_NativeRes(Image, ImagePath, HDR, window_size)
}
\arguments{
\item{Image}{numeric. Image corresponding to a 2D matrix or 3D array}

\item{ImagePath}{character. Path where the image should be written}

\item{HDR}{list. Image header}

\item{window_size}{numeric. window size used to generate Image}
}
\value{
None
}
\description{
write an image resulting from "window processing" at native spatial resolution
(assuming square windows & origin at top left corner)
}
