% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{write_raster}
\alias{write_raster}
\title{Writes a matrix or an array into a ENVI BIL raster}
\usage{
write_raster(
  Image,
  HDR,
  ImagePath,
  window_size,
  FullRes = TRUE,
  LowRes = FALSE,
  SmoothImage = FALSE
)
}
\arguments{
\item{Image}{numeric. matrix or array of image to be written}

\item{HDR}{hdr template}

\item{ImagePath}{path of image file to be written}

\item{window_size}{spatial units use dto compute diversiy (in pixels)}

\item{FullRes}{should full resolution image be written (original pixel size)}

\item{LowRes}{should low resolution image be written (one value per spatial unit)}

\item{SmoothImage}{boolean. set TRUE if you want smooting filter applied to resulting diversity rasters}
}
\value{
None
}
\description{
Writes a matrix or an array into a ENVI BIL raster
}
