% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{write_StarsStack}
\alias{write_StarsStack}
\title{This function writes a stars object into a raster file}
\usage{
write_StarsStack(StarsObj, dsn, BandNames = NULL, datatype = "Float32")
}
\arguments{
\item{StarsObj}{list. stars object containing raster data. Can be produced with function Crop_n_resample_S2}

\item{dsn}{character. path where to store the image produced}

\item{BandNames}{character. vector of band names to be assigned to the raster}

\item{datatype}{character. should be Int16 or Float64 for example}
}
\value{
None
}
\description{
This function writes a stars object into a raster file
}
