% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabeta_window.R
\name{alphabeta_window}
\alias{alphabeta_window}
\title{compute alpha and beta diversity metrics from pixel data corresponding to
spectral species extracted from a window}
\usage{
alphabeta_window(
  SSwindow,
  nbclusters,
  Beta_info,
  alphametrics,
  pcelim = 0.02,
  Hill_order = 1,
  p = NULL
)
}
\arguments{
\item{SSwindow}{dataframe. spectral species corresponding to a raster subset
(window = elementary spatial unit of process)}

\item{nbclusters}{numeric. number of clusters used in kmeans}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics from training set}

\item{alphametrics}{list. alpha diversity metrics: richness, shannon, simpson}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{Hill_order}{numeric. Hill order}

\item{p}{list. progressor object for progress bar}
}
\value{
list of alpha and beta diversity metrics
}
\description{
compute alpha and beta diversity metrics from pixel data corresponding to
spectral species extracted from a window
}
