% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabeta_window_classif.R
\name{alphabeta_window_classif}
\alias{alphabeta_window_classif}
\title{compute alpha and beta diversity metrics from pixel data corresponding to
spectral species extracted from a window}
\usage{
alphabeta_window_classif(
  SSwindow,
  nb_clusters,
  Beta_info = NULL,
  alpha_metrics,
  pcelim = 0.02,
  Hill_order = 1
)
}
\arguments{
\item{SSwindow}{dataframe. spectral species corresponding to a raster subset
(window = elementary spatial unit of process)}

\item{nb_clusters}{numeric. number of clusters used in kmeans}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics}

\item{alpha_metrics}{list. alpha diversity metrics}

\item{pcelim}{numeric. min proportion of pixels to consider spectral species}

\item{Hill_order}{numeric. Hill order}
}
\value{
list of alpha and beta diversity metrics
}
\description{
compute alpha and beta diversity metrics from pixel data corresponding to
spectral species extracted from a window
}
