% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabeta_window_list.R
\name{alphabeta_window_list}
\alias{alphabeta_window_list}
\title{apply alphabeta_window to a list of lists}
\usage{
alphabeta_window_list(
  SSwindow,
  nbclusters,
  Beta_info = NULL,
  alphametrics = "shannon",
  pcelim = 0.02
)
}
\arguments{
\item{SSwindow}{dataframe. spectral species corresponding to a raster subset
(window = elementary spatial unit of process)}

\item{nbclusters}{numeric. number of clusters used in kmeans}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics from training set}

\item{alphametrics}{list. alpha diversity metrics: richness, shannon, simpson}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}
}
\value{
list of alpha and beta diversity metrics
}
\description{
apply alphabeta_window to a list of lists
}
