% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_continuum_removal.R
\name{apply_continuum_removal}
\alias{apply_continuum_removal}
\title{prepares data to run multithreaded continuum removal}
\usage{
apply_continuum_removal(Spectral_Data, Spectral)
}
\arguments{
\item{Spectral_Data}{numeric. initial data matrix (nb samples x nb bands)}

\item{Spectral}{list. information about spectral bands}
}
\value{
samples from image and updated number of pixels to sample if necessary
}
\description{
prepares data to run multithreaded continuum removal
}
