% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ContinuumRemoval.R
\name{apply_continuum_removal}
\alias{apply_continuum_removal}
\title{prepares data to run multithreaded continuum removal}
\usage{
apply_continuum_removal(Spectral_Data, Spectral, nbCPU = 1)
}
\arguments{
\item{Spectral_Data}{numeric. initial data matrix (nb samples x nb bands)}

\item{Spectral}{list. information about spectral bands}

\item{nbCPU}{numeric. number of CPUs to be used in parallel}
}
\value{
samples from image and updated number of pixels to sample if necessary
}
\description{
prepares data to run multithreaded continuum removal
}
