% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapSpectralSpecies.R
\name{apply_kmeans}
\alias{apply_kmeans}
\title{Applies Kmeans clustering to PCA image and writes spectral species map}
\usage{
apply_kmeans(
  PCA_Path,
  PC_Select,
  Input_Mask_File,
  Kmeans_info,
  Spectral_Species_Path,
  nbCPU = 1,
  MaxRAM = 0.25
)
}
\arguments{
\item{PCA_Path}{path for the PCA image}

\item{PC_Select}{PCs selected from PCA}

\item{Input_Mask_File}{Path for the mask}

\item{Kmeans_info}{information about kmeans computed in previous step}

\item{Spectral_Species_Path}{path for spectral species file to be written}

\item{nbCPU}{numeric. number of CPU to work with in multisession task}

\item{MaxRAM}{numeric. size of image pieces to be read at once}
}
\value{
None
}
\description{
Applies Kmeans clustering to PCA image and writes spectral species map
}
