% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_metrics_sfs.R
\name{beta_metrics_sfs}
\alias{beta_metrics_sfs}
\title{compute beta metrics during sfs}
\usage{
beta_metrics_sfs(
  obs2optimize,
  SSValid,
  nb_clusters,
  pcelim,
  nb_iter,
  nbPlots_total,
  Attributes,
  corrMethod,
  IDwindow
)
}
\arguments{
\item{obs2optimize}{numeric. number of clusters used in kmeans}

\item{SSValid}{numeric}

\item{nb_clusters}{numeric}

\item{pcelim}{numeric}

\item{nb_iter}{numeric}

\item{nbPlots_total}{numeric}

\item{Attributes}{list}

\item{corrMethod}{character}

\item{IDwindow}{numeric}
}
\value{
assessed value and correlation
}
\description{
compute beta metrics during sfs
}
