% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR-package.R, R/biodivMapR2-package.R
\docType{package}
\name{biodivMapR-package}
\alias{biodivMapR-package}
\title{biodivMapR: biodivMapR: an R package for a- and ß-diversity mapping using remotely-sensed images}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

this packages allows processing image data based on the method described in the following publication: Féret, J.-B., Asner, G.P., 2014. Mapping tropical forest canopy diversity using high-fidelity imaging spectroscopy. Ecol. Appl. 24, 1289–1296. https://doi.org/10.1890/13-1824.1 It expects a raster file as input, with ENVI HDR header file including spectral bands of the sensor, if optical sensor.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

this packages allows processing image data based on the method described in the following publication: Féret, J.-B., Asner, G.P., 2014. Mapping tropical forest canopy diversity using high-fidelity imaging spectroscopy. Ecol. Appl. 24, 1289–1296. https://doi.org/10.1890/13-1824.1 It expects a raster file as input, with ENVI HDR header file including spectral bands of the sensor, if optical sensor.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jbferet/biodivMapR}
  \item Report bugs at \url{https://github.com/jbferet/biodivMapR/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/jbferet/biodivMapR}
  \item Report bugs at \url{https://github.com/jbferet/biodivMapR/issues}
}

}
\author{
\strong{Maintainer}: Jean-Baptiste Feret \email{jb.feret@teledetection.fr} (\href{https://orcid.org/0000-0002-0151-1334}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Florian de Boissieu \email{florian.deboissieu@inrae.fr} (\href{https://orcid.org/0000-0002-2185-9952}{ORCID}) (clean code, format as package, build website)
}

}
\keyword{internal}
