% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_chunk.R
\name{biodivMapR_chunk}
\alias{biodivMapR_chunk}
\title{apply biodivMapR (computes clusters + diversity metrics) to an image chunk}
\usage{
biodivMapR_chunk(
  blk,
  r_in,
  window_size,
  Kmeans_info,
  Beta_info = NULL,
  alpha_metrics = "shannon",
  Hill_order = 1,
  fd_metrics = NULL,
  selected_bands = NULL,
  pcelim = 0.02,
  nbCPU = 1,
  min_sun = 0.25,
  p = NULL
)
}
\arguments{
\item{blk}{list. rows and number of rows to read from}

\item{r_in}{list. path of rasters to read from}

\item{window_size}{numeric. window size for square plots}

\item{Kmeans_info}{list. kmeans description obtained from function get_kmeans}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics}

\item{alpha_metrics}{list. alpha diversity metrics}

\item{Hill_order}{numeric. Hill order}

\item{fd_metrics}{character. list of functional metrics}

\item{selected_bands}{numeric. bands selected from input data}

\item{pcelim}{numeric. min proportion of pixels to consider spectral species}

\item{nbCPU}{numeric. Number of CPUs available}

\item{min_sun}{numeric. minimum amount of sunlit pixels in the plots}

\item{p}{list. progressor object for progress bar}
}
\value{
Shannon index correspnding to the distribution
}
\description{
apply biodivMapR (computes clusters + diversity metrics) to an image chunk
}
