% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_chunk.R
\name{biodivMapR_chunk}
\alias{biodivMapR_chunk}
\title{apply biodivMapR (computes clusters + diversity metrics) to an image chunk}
\usage{
biodivMapR_chunk(
  blk,
  r_in,
  window_size,
  Kmeans_info,
  Beta_info = NULL,
  alphametrics = "shannon",
  SelectBands = NULL,
  pcelim = 0.02,
  nbCPU = 1,
  MinSun = 0.25,
  p = NULL
)
}
\arguments{
\item{blk}{list. rows and number of rows to read from}

\item{r_in}{list. path of rasters to read from}

\item{window_size}{numeric. window size for square plots}

\item{Kmeans_info}{list. kmeans description obtained from function get_kmeans}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics from training set}

\item{alphametrics}{list. alpha diversity metrics: richness, shannon, simpson}

\item{SelectBands}{numeric. bands selected from input data}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{nbCPU}{numeric. Number of CPUs available}

\item{MinSun}{numeric. minimum amount of sunlit pixels in the plots}

\item{p}{list. progressor object for progress bar}
}
\value{
Shannon index correspnding to the distribution
}
\description{
apply biodivMapR (computes clusters + diversity metrics) to an image chunk
}
