% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_full.R
\name{biodivMapR_full}
\alias{biodivMapR_full}
\title{computes diversity metrics from raster}
\usage{
biodivMapR_full(
  input_raster_path,
  input_mask_path = NULL,
  output_dir,
  window_size,
  selected_bands = NULL,
  Kmeans_info_save = NULL,
  Kmeans_info_read = NULL,
  Beta_info_save = NULL,
  Beta_info_read = NULL,
  nbCPU = 1,
  options = NULL
)
}
\arguments{
\item{input_raster_path}{character. path for the input rasters}

\item{input_mask_path}{character. path for mask file}

\item{output_dir}{character. path for the output files}

\item{window_size}{numeric. window size for square plots}

\item{selected_bands}{numeric. bands selected from input_rast}

\item{Kmeans_info_save}{character. path where to save Kmeans_info}

\item{Kmeans_info_read}{character. path where to read Kmeans_info}

\item{Beta_info_save}{character. path where to save Beta_info}

\item{Beta_info_read}{character. path where to read Beta_info}

\item{nbCPU}{numeric. Number of CPUs available}

\item{maxRows}{numeric. max number of rows processed once by each CPU}

\item{nb_clusters}{numeric. number of clusters used in kmeans}

\item{nb_samples_beta}{numeric. number of samples to compute beta diversity}

\item{alpha_metrics}{list. alpha diversity metrics: richness, shannon, simpson}

\item{beta_metrics}{boolean. set TRUE to compute beta diversity}

\item{Hill_order}{numeric. Hill order}

\item{fd_metrics}{character. list of functional metrics}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{nb_iter}{numeric. nb of iterations averaged to compute diversity indices}

\item{min_sun}{numeric. minimum amount of sunlit pixels in the plots}

\item{nb_samples_alpha}{numeric. max number of pixels extracted for kmeans}

\item{dimPCoA}{numeric. number of dimensions of PCoA}

\item{progressbar}{boolean. set true for progress bar during clustering}

\item{filetype}{character. driver for output diversity raster data}

\item{moving_window}{boolean. should diversity be computed on moving window?}
}
\value{
Kmeans_info and Beta_info
}
\description{
computes diversity metrics from raster
}
