% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_full_tiles.R
\name{biodivMapR_full_tiles}
\alias{biodivMapR_full_tiles}
\title{computes diversity metrics from raster}
\usage{
biodivMapR_full_tiles(
  dsn_grid,
  ID_aoi,
  feature_dir,
  list_features,
  mask_dir = NULL,
  output_dir,
  window_size,
  plots,
  nbsamples_alpha = 1e+05,
  nbsamples_beta = 2000,
  nbCPU = 1,
  pcelim = 0.02,
  maxRows = 1200,
  MovingWindow = F,
  siteName = NULL
)
}
\arguments{
\item{dsn_grid}{character. path for the tiling grid}

\item{ID_aoi}{list.}

\item{feature_dir}{character. path where to get features}

\item{list_features}{character. list of features}

\item{mask_dir}{character. path for masks}

\item{output_dir}{character. path where to save results}

\item{window_size}{numeric. number of clusters used in kmeans}

\item{plots}{list. list of sf plots}

\item{nbsamples_alpha}{numeric. number of samples to compute beta diversity from}

\item{nbsamples_beta}{numeric. number of samples to compute beta diversity from}

\item{nbCPU}{numeric. Number of CPUs available}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{maxRows}{numeric. maximum number of rows}

\item{MovingWindow}{boolean. should moving window be used?}

\item{siteName}{character. name for the output files}
}
\value{
mosaic_path
}
\description{
computes diversity metrics from raster
}
