% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_full_tiles.R
\name{biodivMapR_full_tiles}
\alias{biodivMapR_full_tiles}
\title{computes diversity metrics from raster}
\usage{
biodivMapR_full_tiles(
  dsn_grid,
  feature_dir,
  list_features,
  mask_dir = NULL,
  output_dir,
  window_size,
  plots,
  nb_clusters = 50,
  nb_samples_alpha = 1e+05,
  nb_samples_beta = 2000,
  alphametrics = "shannon",
  Hill_order = 1,
  FDmetric = NULL,
  nbCPU = 1,
  nb_iter = 10,
  pcelim = 0.02,
  maxRows = 1200,
  moving_window = FALSE,
  siteName = NULL
)
}
\arguments{
\item{dsn_grid}{character. path for the tiling grid}

\item{feature_dir}{character. path where to get features}

\item{list_features}{character. list of features}

\item{mask_dir}{character. path for masks}

\item{output_dir}{character. path where to save results}

\item{window_size}{numeric. number of clusters used in kmeans}

\item{plots}{list. list of sf plots}

\item{nb_clusters}{numeric. number of clusters}

\item{nb_samples_alpha}{numeric. number of samples to compute alpha diversity}

\item{nb_samples_beta}{numeric. number of samples to compute beta diversity}

\item{alphametrics}{character.}

\item{Hill_order}{numeric.}

\item{FDmetric}{character.}

\item{nbCPU}{numeric. Number of CPUs available}

\item{nb_iter}{numeric. Number of iterations required to compute diversity}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{maxRows}{numeric. maximum number of rows}

\item{moving_window}{boolean. should moving window be used?}

\item{siteName}{character. name for the output files}
}
\value{
mosaic_path
}
\description{
computes diversity metrics from raster
}
