% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_opt_clusters.R
\name{biodivMapR_opt_clusters}
\alias{biodivMapR_opt_clusters}
\title{apply biodivMapR on a test set for different numbers of clusters}
\usage{
biodivMapR_opt_clusters(
  input_raster_path,
  obs_vect,
  obs2optimize,
  selected_bands,
  obs_criterion = "shannon",
  input_mask_path = NULL,
  outputdir = "./",
  nb_clusters = 50,
  min_sun = 0.25,
  nb_iter = 10,
  pcelim = 0.02,
  verbose = TRUE,
  nb_repetitions = 50,
  nb_samples_alpha = 1e+05,
  Hill_order = 1,
  algorithm = "Hartigan-Wong",
  nbCPU = 1
)
}
\arguments{
\item{input_raster_path}{SpatRaster or list of SpatRaster}

\item{obs_vect}{SpatVector or SpatVectorCollection}

\item{obs2optimize}{numeric .list of ground obs diversity metrics
corresponding to obs_vect.
Expected values: richness, shannon, simpson, BC}

\item{selected_bands}{numeric. bnds to select from input_raster}

\item{obs_criterion}{character. richness, shannon, simpson or BC}

\item{input_mask_path}{SpatRaster corresponding to mask}

\item{outputdir}{character. output directory}

\item{nb_clusters}{numeric.}

\item{min_sun}{numeric.}

\item{nb_iter}{numeric.}

\item{pcelim}{numeric.}

\item{verbose}{boolean.}

\item{nb_repetitions}{numeric.}

\item{nb_samples_alpha}{numeric.}

\item{Hill_order}{numeric.}

\item{algorithm}{character.}

\item{nbCPU}{numeric.}
}
\value{
list including performances (correlation) of SFS with additional
features and assessed diversity metrics corresponding to each step
}
\description{
apply biodivMapR on a test set for different numbers of clusters
}
