% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_sample.R
\name{biodivMapR_sample}
\alias{biodivMapR_sample}
\title{sample pixels and plots to produce alpha and beta models, including
clustering, dissimilarity & PCoA}
\usage{
biodivMapR_sample(
  feature_dir,
  list_features,
  mask_dir = NULL,
  output_dir,
  window_size,
  plot_names,
  nb_clusters = 50,
  nb_samples_alpha = 1e+05,
  beta_metrics = "bray",
  compute_beta = TRUE,
  nb_samples_beta = 2000,
  pcelim = 0.02,
  nbCPU = 1,
  nb_iter = 10,
  weightIQR = 4,
  Kmeans_path = NULL,
  Beta_path = NULL
)
}
\arguments{
\item{feature_dir}{character. path where to get features}

\item{list_features}{character. list of features}

\item{mask_dir}{character. path for masks}

\item{output_dir}{character. path where to save results}

\item{window_size}{numeric. number of clusters used in kmeans}

\item{plot_names}{character. plot names}

\item{nb_clusters}{numeric. number of clusters}

\item{nb_samples_alpha}{numeric. number of samples to compute alpha diversity}

\item{beta_metrics}{vector. list of beta diversity metrics}

\item{compute_beta}{boolean. set TRUE to compute beta diversity}

\item{nb_samples_beta}{numeric. number of samples to compute beta diversity}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{nbCPU}{numeric. Number of CPUs available}

\item{nb_iter}{numeric. Number of iterations required to compute diversity}

\item{weightIQR}{numeric. IQR applied to filter out features to be used}

\item{Kmeans_path}{character. path for Kmeans_info}

\item{Beta_path}{character. path for Beta_info}
}
\value{
Kmeans_info, Beta_info and ID_aoi
}
\description{
sample pixels and plots to produce alpha and beta models, including
clustering, dissimilarity & PCoA
}
