% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_sdm.R
\name{biodivMapR_sdm}
\alias{biodivMapR_sdm}
\title{computes diversity metrics from raster corresponding to SDM}
\usage{
biodivMapR_sdm(
  input_raster_path,
  output_dir,
  input_mask_path = NULL,
  site_name = NULL,
  alpha_metrics = "shannon",
  beta_metrics = "bray",
  nb_samples_beta = 1000
)
}
\arguments{
\item{input_raster_path}{character. path for the input rasters}

\item{output_dir}{character. path for the output files}

\item{input_mask_path}{character. path for mask file}

\item{site_name}{character. name of site}

\item{alpha_metrics}{character. list of alpha metrics}

\item{beta_metrics}{character. list of beta metrics}

\item{nb_samples_beta}{numeric. number of samples to compute beta diversity}
}
\value{
diversity_maps_ground
}
\description{
computes diversity metrics from raster corresponding to SDM
}
