% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_SFS.R
\name{biodivMapR_SFS}
\alias{biodivMapR_SFS}
\title{performs SFS to identify combination of input variables maximizing a criterion}
\usage{
biodivMapR_SFS(
  input_raster,
  obs_vect,
  obs2optimize,
  obs_criterion = "shannon",
  corrMethod = "spearman",
  input_mask = NULL,
  Hill_order = 1,
  nbclusters = 50,
  MinSun = 0.25,
  nbPix = 1e+05,
  nbIter = 20,
  pcelim = 0.02,
  verbose = T,
  nbWorkers = 1,
  nbCPU = 1
)
}
\arguments{
\item{input_raster}{SpatRaster or list of SpatRaster}

\item{obs_vect}{SpatVector or SpatVectorCollection}

\item{obs2optimize}{numeric .list of ground obs diversity metrics
corresponding to obs_vect.
Expected values: richness, shannon, simpson, hill, BC}

\item{obs_criterion}{character. richness, shannon, simpson or BC}

\item{corrMethod}{character. select between method available for cor.test}

\item{input_mask}{SpatRaster corresponding to mask}

\item{Hill_order}{numeric.}

\item{nbclusters}{numeric.}

\item{MinSun}{numeric.}

\item{nbPix}{numeric.}

\item{nbIter}{numeric.}

\item{pcelim}{numeric.}

\item{verbose}{boolean.}

\item{nbWorkers}{numeric.}

\item{nbCPU}{numeric.}
}
\value{
list including performances (correlation) of SFS with additional
features and assessed diversity metrics corresponding to each step
}
\description{
performs SFS to identify combination of input variables maximizing a criterion
}
