% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivMapR_tiling.R
\name{biodivMapR_tiling}
\alias{biodivMapR_tiling}
\title{computes diversity metrics from a set of raster corresponding to
tiles as produced with preprocS2}
\usage{
biodivMapR_tiling(
  feature_dir,
  list_features,
  mask_dir = NULL,
  output_dir,
  window_size,
  plots,
  nbCPU = 1,
  site_name = NULL,
  options = NULL
)
}
\arguments{
\item{feature_dir}{character. path where to get features}

\item{list_features}{character. list of features}

\item{mask_dir}{character. path for masks}

\item{output_dir}{character. path where to save results}

\item{window_size}{numeric. number of clusters used in kmeans}

\item{plots}{list. list of sf plots}

\item{nbCPU}{numeric. Number of CPUs available}

\item{site_name}{character. name for the output files}

\item{options}{list. including
\itemize{
\item nb_clusters numeric. number of clusters
\item nb_samples_alpha numeric. number of samples to compute alpha diversity
\item nb_samples_beta numeric. number of samples to compute beta diversity
\item alpha_metrics character.
\item Hill_order numeric.
\item fd_metrics character.
\item nb_iter numeric. Number of iterations required to compute diversity
\item pcelim numeric. minimum proportion of pixels to consider spectral species
\item maxRows numeric. maximum number of rows
\item moving_window boolean. should moving window be used?
\item mosaic_output boolean. set TRUE if outputs need to be mosaiced
\item weightIQR numeric. IQR applied to filter out features to be used
}}
}
\value{
mosaic_path
}
\description{
computes diversity metrics from a set of raster corresponding to
tiles as produced with preprocS2
}
