% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_CheckConvertData.R
\name{check_data}
\alias{check_data}
\title{Checks if the data to be processed has the format type expected}
\usage{
check_data(DataMatrix, Spectral)
}
\arguments{
\item{Raster.Path}{full path for the raster to be converted}

\item{Mask}{is the raster a mask?}
}
\description{
Checks if the data to be processed has the format type expected
}
