% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_CheckConvertData.R
\name{check_data}
\alias{check_data}
\title{Checks if the data to be processed has the format type expected}
\usage{
check_data(Raster_Path, Mask = FALSE)
}
\arguments{
\item{Raster_Path}{character. full path for the raster to be converted}

\item{Mask}{boolean. Set true if the raster is a mask}
}
\value{
nothing
}
\description{
Checks if the data to be processed has the format type expected
}
