% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tiff_list.R
\name{clean_tiff_list}
\alias{clean_tiff_list}
\title{computes diversity metrics from raster}
\usage{
clean_tiff_list(list_files)
}
\arguments{
\item{list_files}{vector. list of file names corresponding to tiff}
}
\value{
list_files vector. list of file names withoout aux.xml and tfw files
}
\description{
computes diversity metrics from raster
}
