% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mask_IQR.R
\name{compute_mask_iqr}
\alias{compute_mask_iqr}
\title{compute mask based on interquartile range criterion applied on input rasters}
\usage{
compute_mask_iqr(
  input_raster_path,
  output_mask_path,
  input_mask_path = NULL,
  weightIQR = 3,
  filetype = "GTiff"
)
}
\arguments{
\item{input_raster_path}{character. path for input rasters.}

\item{output_mask_path}{character. path for output mask}

\item{input_mask_path}{character. path for optional input mask}

\item{weightIQR}{numeric. weight to define SD range for IQR}

\item{filetype}{character. GDAL driver}
}
\value{
none
}
\description{
compute mask based on interquartile range criterion applied on input rasters
}
