% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mask_iqr_tiles.R
\name{compute_mask_iqr_tiles}
\alias{compute_mask_iqr_tiles}
\title{computes a mask for a collection of tiles based on interquartile range filtering
interquartile is computed for all data included in teh collection
samples out of a range defined by
quartile_1-weightIQR x (quartile_3 - quartile_1) and
quartile_3+weightIQR x (quartile_3-quartile_1)
are then considered as outliers}
\usage{
compute_mask_iqr_tiles(
  feature_dir,
  feature_list,
  mask_dir,
  plots,
  weightIQR = 4,
  filetype = "GTiff",
  nbCPU = 1,
  nb_pixstats = 5e+06
)
}
\arguments{
\item{feature_dir}{character.}

\item{feature_list}{character.}

\item{mask_dir}{character.}

\item{plots}{list.}

\item{weightIQR}{numeric.}

\item{filetype}{character.}

\item{nbCPU}{numeric.}

\item{nb_pixstats}{numeric.}
}
\value{
mask_path_list
}
\description{
computes a mask for a collection of tiles based on interquartile range filtering
interquartile is computed for all data included in teh collection
samples out of a range defined by
quartile_1-weightIQR x (quartile_3 - quartile_1) and
quartile_3+weightIQR x (quartile_3-quartile_1)
are then considered as outliers
}
