% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mask_IQR_tiles.R
\name{compute_mask_iqr_tiles}
\alias{compute_mask_iqr_tiles}
\title{get samples for alpha and beta diversity mapping}
\usage{
compute_mask_iqr_tiles(
  feature_dir,
  feature_list,
  mask_dir,
  plots,
  weightIRQ = 4,
  filetype = "GTiff",
  nbCPU = 1,
  nb_pixstats = 5e+06
)
}
\arguments{
\item{feature_dir}{character.}

\item{feature_list}{character.}

\item{mask_dir}{character.}

\item{plots}{list.}

\item{weightIRQ}{numeric.}

\item{filetype}{character.}

\item{nbCPU}{numeric.}

\item{nb_pixstats}{numeric.}
}
\value{
mask_path_list
}
\description{
get samples for alpha and beta diversity mapping
}
