% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{compute_NMDS}
\alias{compute_NMDS}
\title{computes NMDS}
\usage{
compute_NMDS(MatBCdist)
}
\arguments{
\item{MatBCdist}{BC dissimilarity matrix}
}
\value{
BetaNMDS_sel
}
\description{
computes NMDS
}
