% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{compute_NN_from_ordination}
\alias{compute_NN_from_ordination}
\title{compute the nearest neighbors among kernels used in NMDS}
\usage{
compute_NN_from_ordination(MatBC3, knn, BetaDiversity0)
}
\arguments{
\item{MatBC3}{matrix of BC dissimilarity between the kernels excluded from Ordination (rows)}

\item{knn}{numeric. number of neighbors}

\item{BetaDiversity0}{numeric. matrix of ordinated coordinates computed from dissimilarity matrix}
}
\value{
Ordin_est estimated NMDS position based on nearest neighbors from NMDS
}
\description{
compute the nearest neighbors among kernels used in NMDS
}
