% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapAlphaDiversity.R
\name{convert_PCA_to_SSD}
\alias{convert_PCA_to_SSD}
\title{Convert PCA into SSD based on previous clustering}
\usage{
convert_PCA_to_SSD(
  ReadWrite,
  Spectral_Species_Path,
  HDR_SS,
  HDR_SSD,
  HDR_Sunlit,
  window_size,
  nbclusters,
  Input_Mask_File = FALSE,
  HDR_Mask = FALSE,
  MinSun = 0.25,
  pcelim = 0.02,
  Index_Alpha = c("Shannon"),
  SSD_Path,
  Sunlit_Path,
  p = NULL,
  nbCPU = 1,
  nbPieces = 1,
  progressbar = FALSE
)
}
\arguments{
\item{ReadWrite}{list. includes byte coordinates where to read and write from files}

\item{Spectral_Species_Path}{character. path for Spectral_Species raster file}

\item{HDR_SS}{list. HDR file for spectral species}

\item{HDR_SSD}{list. HDR file for spectral species distribution}

\item{HDR_Sunlit}{list. HDR file for spectral species distribution}

\item{window_size}{numeric. window size}

\item{nbclusters}{numeric. number of clusters}

\item{Input_Mask_File}{character. input mask file (useful when using classification map)}

\item{HDR_Mask}{list. HDR file for mask}

\item{MinSun}{numeric. Minimum proportion of sunlit pixels required to consider plot.}

\item{pcelim}{numeric. Minimum contribution (in \%) required for a spectral species.}

\item{Index_Alpha}{list. list of spectral metrics to be computed}

\item{SSD_Path}{character. path for spectral species distribution file}

\item{Sunlit_Path}{character. path for sunlit proportion file}

\item{p}{list. progressor object for progress bar}

\item{nbCPU}{numeric. Number of CPUs to use in parallel.}

\item{nbPieces}{numeric. number of pieces to split file read into}

\item{progressbar}{boolean. should progress bar be displayed (set to TRUE only if no conflict of parallel process)}
}
\value{
list of mean and SD of alpha diversity metrics
}
\description{
Convert PCA into SSD based on previous clustering
}
