% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_border_tile.R
\name{crop_border_tile}
\alias{crop_border_tile}
\title{save diversity maps as raster data once moving window process done}
\usage{
crop_border_tile(rast_obj, window_size)
}
\arguments{
\item{rast_obj}{spatRaster.}

\item{window_size}{numeric. window size for square plots}
}
\value{
rast_crop
}
\description{
save diversity maps as raster data once moving window process done
}
