% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PerformPCA.R
\name{define_pixels_per_iter}
\alias{define_pixels_per_iter}
\title{defines the number of pixels per iteration based on a trade-off between image size and sample size per iteration}
\usage{
define_pixels_per_iter(ImNames, nb_partitions)
}
\arguments{
\item{ImNames}{Path and name of the images to be processed}

\item{nb_partitions}{number of iterations peformed to average diversity indices}
}
\value{
Pix_Per_Partition number of pixels per iteration
}
\description{
defines the number of pixels per iteration based on a trade-off between image size and sample size per iteration
}
