% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_pixels_per_iter.R
\name{define_pixels_per_iter}
\alias{define_pixels_per_iter}
\title{defines the number of pixels per iteration}
\usage{
define_pixels_per_iter(
  input_rast,
  input_mask = NULL,
  nb_iter = 10,
  nb_pix = 1e+05
)
}
\arguments{
\item{input_rast}{SpatRaster.}

\item{input_mask}{SpatRaster.}

\item{nb_iter}{numeric. nb of iterations averaged to compute diversity indices}

\item{nb_pix}{numeric. maximum number of pixels to extract for kmeans}
}
\value{
Pix_Per_Iter number of pixels per iteration
}
\description{
defines the number of pixels per iteration
}
