% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_biodivMapR.R
\name{diversity_from_plots}
\alias{diversity_from_plots}
\title{gets alpha diversity indicators from plot}
\usage{
diversity_from_plots(
  Raster_SpectralSpecies,
  Plots,
  nbclusters = 50,
  Raster_Functional = FALSE,
  Selected_Features = FALSE,
  Name_Plot = FALSE,
  Hellinger = FALSE,
  pcelim = 0.02
)
}
\arguments{
\item{Raster_SpectralSpecies}{character. path for the SpectralSpecies file computed from DiverstyMapping method}

\item{Plots}{list. list of paths corresponding to shapefiles defining polygons in the raster}

\item{nbclusters}{numeric. Number of clusters defined in k-Means.}

\item{Raster_Functional}{character. path for the raster file used to compute functional diversity}

\item{Selected_Features}{numeric. selected features for Raster_Functional. use all features if set to FALSE}

\item{Name_Plot}{character. Name of the plots defined in the shapefiles}

\item{Hellinger}{boolean. set TRUE to compute Hellinger distance matrices based on Euclidean distances}

\item{pcelim}{numeric. Discard spectral species when contribution is < pcelim (between 0 and 1)}
}
\value{
alpha and beta diversity metrics
}
\description{
gets alpha diversity indicators from plot
}
