% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_spectral_domains.R
\name{exclude_spectral_domains}
\alias{exclude_spectral_domains}
\title{define water vapor bands based on spectral sampling of original image}
\usage{
exclude_spectral_domains(input_rast, Excluded_WL = NULL)
}
\arguments{
\item{input_rast}{character. path of the image}

\item{Excluded_WL}{numeric. spectral domains corresponding to water vapor absorption}
}
\value{
bands corresponding to atmospheric water absorption domain
}
\description{
define water vapor bands based on spectral sampling of original image
}
