% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_cluster_range.R
\name{explore_cluster_range}
\alias{explore_cluster_range}
\title{explore performances for a range of cluster number}
\usage{
explore_cluster_range(
  repet,
  input_raster_path,
  input_vector_path,
  input_mask_path,
  selected_bands,
  obs_criterion = "shannon",
  alpha_metrics,
  beta_metrics,
  output_dir = "./",
  nbClust_list,
  min_sun = 0.25,
  nb_iter = 10,
  pcelim = 0.02,
  nb_samples_alpha = 1e+05,
  nb_samples_beta = 1000,
  Hill_order = 1,
  algorithm = "Hartigan-Wong",
  overwrite = TRUE,
  p = NULL
)
}
\arguments{
\item{repet}{character identifier of the repetition}

\item{input_raster_path}{SpatRaster or list of SpatRaster}

\item{input_vector_path}{SpatVector or SpatVectorCollection}

\item{input_mask_path}{SpatRaster corresponding to mask}

\item{selected_bands}{numeric. bnds to select from input_raster}

\item{obs_criterion}{character. richness, shannon, simpson or BC}

\item{alpha_metrics}{character. richness, shannon, simpson}

\item{beta_metrics}{character.}

\item{output_dir}{character. output directory}

\item{nbClust_list}{numeric.}

\item{min_sun}{numeric.}

\item{nb_iter}{numeric.}

\item{pcelim}{numeric.}

\item{nb_samples_alpha}{numeric.}

\item{nb_samples_beta}{numeric.}

\item{Hill_order}{numeric.}

\item{algorithm}{character.}

\item{overwrite}{boolean.}

\item{p}{list}
}
\value{
list divIndex_est
}
\description{
explore performances for a range of cluster number
}
