% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_kmeans.R
\name{explore_kmeans}
\alias{explore_kmeans}
\title{Compute kmeans from random subset of pixels extracted from an image and a list
of values for k}
\usage{
explore_kmeans(
  input_rast,
  input_mask = NULL,
  selected_bands = NULL,
  nbClust_list = 50,
  nb_iter = 10,
  nb_samples_alpha = 1e+05,
  algorithm = "Hartigan-Wong",
  nbCPU = 1,
  progressbar = FALSE
)
}
\arguments{
\item{input_rast}{character. Path and name of the image to be processed.}

\item{input_mask}{character. Path and name of the mask corresponding to the image}

\item{selected_bands}{numeric. bands selected from input_rast}

\item{nbClust_list}{numeric. number of clusters used in kmeans}

\item{nb_iter}{numeric. nb of iterations averaged to compute diversity indices}

\item{nb_samples_alpha}{numeric. max number of pixels to extract for kmeans}

\item{algorithm}{character. algorithm used in the kmeans clustering}

\item{nbCPU}{numeric. Number of CPUs available}

\item{progressbar}{boolean. set true for progress bar during clustering}
}
\value{
kmeans_info
}
\description{
Compute kmeans from random subset of pixels extracted from an image and a list
of values for k
}
