% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_biodivMapR.R
\name{extract_pixels_coordinates}
\alias{extract_pixels_coordinates}
\title{Extracts pixels coordinates from raster intersecting a vector.}
\usage{
extract_pixels_coordinates(x, y)
}
\arguments{
\item{x}{terra::SpatRaster or character. Raster or raster path.}

\item{y}{terra::SpatVector or character. Vector or vector path.}
}
\value{
data.frame. Pixel indices row, col of intersecting parts.
}
\description{
Extracts pixels coordinates from raster intersecting a vector.
}
