% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_prior_CR.R
\name{filter_prior_CR}
\alias{filter_prior_CR}
\title{Filter data prior to continuum removal:
\itemize{
\item values are expected to be real reflectance values between 0 and 10000
\item negative values may occur, so a +100 value is applied to avoid negative
\item possibly remaining negative values are set to 0
\item constant spectra are eliminated
}}
\usage{
filter_prior_CR(Minit, Spectral_Bands)
}
\arguments{
\item{Minit}{initial data matrix, n rows = n samples, p cols = p spectral bands}

\item{Spectral_Bands}{numeric. central wavelength for the spectral bands}
}
\value{
list. updated Minit
}
\description{
Filter data prior to continuum removal:
\itemize{
\item values are expected to be real reflectance values between 0 and 10000
\item negative values may occur, so a +100 value is applied to avoid negative
\item possibly remaining negative values are set to 0
\item constant spectra are eliminated
}
}
