% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_window.R
\name{functional_window}
\alias{functional_window}
\title{compute functional diversity metrics from pixel data corresponding to
features extracted from a window}
\usage{
functional_window(SSwindow, FDmetric = c("FRic", "FEve", "FDiv"), p = NULL)
}
\arguments{
\item{SSwindow}{dataframe. spectral species corresponding to a raster subset
(window = elementary spatial unit of process)}

\item{FDmetric}{list. alpha diversity metrics: richness, shannon, simpson}

\item{p}{list. progressor object for progress bar}
}
\value{
list of alpha and beta diversity metrics
}
\description{
compute functional diversity metrics from pixel data corresponding to
features extracted from a window
}
