% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_Validation_DiversityMappR.R
\name{gdal_polygonizeR}
\alias{gdal_polygonizeR}
\title{build a vector file from raster footprint
borrowed from https://johnbaumgartner.wordpress.com/2012/07/26/getting-rasters-into-shape-from-r/}
\usage{
gdal_polygonizeR(x, outshape = NULL, gdalformat = "ESRI Shapefile",
  pypath = NULL, readpoly = TRUE, quiet = TRUE)
}
\arguments{
\item{x}{path for a raster or raster object}

\item{outshape}{path for a vector to be written}

\item{pypath}{}

\item{readpoly}{}

\item{quiet}{}
}
\description{
build a vector file from raster footprint
borrowed from https://johnbaumgartner.wordpress.com/2012/07/26/getting-rasters-into-shape-from-r/
}
