% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha_from_SSD.R
\name{get_alpha_from_SSD}
\alias{get_alpha_from_SSD}
\title{computes alpha diversity metrics from SSD}
\usage{
get_alpha_from_SSD(
  SSD,
  nbPix_Sunlit,
  alphametrics = "shannon",
  pcelim = 0.02,
  Hill_order = 1
)
}
\arguments{
\item{SSD}{numeric. spectral species distribution}

\item{nbPix_Sunlit}{numeric.}

\item{alphametrics}{list. alpha diversity metrics: richness, shannon, simpson}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{Hill_order}{numeric. Hill order}
}
\value{
Shannon index correspnding to the distribution
}
\description{
computes alpha diversity metrics from SSD
}
