% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha_from_ssd.R
\name{get_alpha_from_ssd}
\alias{get_alpha_from_ssd}
\title{computes alpha diversity metrics from ssd}
\usage{
get_alpha_from_ssd(
  ssd,
  nb_pix_sunlit,
  alpha_metrics = "shannon",
  pcelim = 0.02,
  hill_order = 1
)
}
\arguments{
\item{ssd}{numeric. spectral species distribution}

\item{nb_pix_sunlit}{numeric.}

\item{alpha_metrics}{list. alpha diversity metrics: richness, shannon, simpson}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{hill_order}{numeric. Hill order}
}
\value{
Shannon index correspnding to the distribution
}
\description{
computes alpha diversity metrics from ssd
}
