% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_list.R
\name{get_cluster_list}
\alias{get_cluster_list}
\title{get cluster list for biodivMapR_opt_clusters}
\usage{
get_cluster_list(nb_clusters)
}
\arguments{
\item{nb_clusters}{numeric. number of clusters}
}
\value{
list including alpha_metrics and beta
}
\description{
get cluster list for biodivMapR_opt_clusters
}
