% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diversity_from_plots.R
\name{get_diversity_from_plots}
\alias{get_diversity_from_plots}
\title{computes diversity metrics from validation plots}
\usage{
get_diversity_from_plots(
  input_rast,
  validation_vect,
  Kmeans_info,
  Beta_info = NULL,
  input_mask = NULL,
  SelectBands = NULL,
  MinSun = 0.25,
  pcelim = 0.02,
  nbCPU = 1
)
}
\arguments{
\item{input_rast}{SpatRaster}

\item{validation_vect}{SpatVector}

\item{Kmeans_info}{list. kmeans description obtained from function get_kmeans}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics from training set}

\item{input_mask}{SpatRaster}

\item{SelectBands}{numeric. bands selected from input_rast}

\item{MinSun}{numeric. minimum amount of sunlit pixels in the plots}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{nbCPU}{numeric. Number of CPUs available}
}
\value{
SpatVector including diversity metrics and BC dissimilarity for the plots
}
\description{
computes diversity metrics from validation plots
}
