% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diversity_from_plots_cluster.R
\name{get_diversity_from_plots_cluster}
\alias{get_diversity_from_plots_cluster}
\title{computes diversity metrics from validation plots}
\usage{
get_diversity_from_plots_cluster(
  input_raster_path,
  Hill_order = 1,
  Kmeans_info,
  Beta_info = NULL,
  input_mask_path = NULL,
  fd_metrics = NULL,
  selected_bands = NULL,
  rast_sample = NULL,
  AttributeTable = NULL,
  alpha_metrics = c("richness", "shannon", "simpson", "hill"),
  min_sun = 0.25,
  pcelim = 0.02,
  nbCPU = 1,
  getBeta = TRUE,
  verbose = FALSE,
  p = NULL
)
}
\arguments{
\item{Hill_order}{numeric. Hill order}

\item{Kmeans_info}{list. kmeans description obtained from function get_kmeans}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics from training set}

\item{fd_metrics}{character.}

\item{selected_bands}{numeric. bands selected from input_rast}

\item{rast_sample}{dataframe.}

\item{AttributeTable}{dataframe.}

\item{alpha_metrics}{character.}

\item{min_sun}{numeric. minimum amount of sunlit pixels in the plots}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{nbCPU}{numeric. Number of CPUs available}

\item{getBeta}{boolean. set true if computation of beta required}

\item{verbose}{boolean. set true for messages}

\item{p}{object}

\item{input_rast}{SpatRaster}

\item{input_mask}{SpatRaster}
}
\value{
SpatVector including diversity metrics and BC dissimilarity for the plots
}
\description{
computes diversity metrics from validation plots
}
