% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functional_diversity.R
\name{get_functional_diversity}
\alias{get_functional_diversity}
\title{get functional diversity metrics from dataframe
This function was inspired from FD package}
\usage{
get_functional_diversity(
  spectraits,
  fd_metrics = c("FRic", "FEve", "FDiv"),
  p = NULL
)
}
\arguments{
\item{spectraits}{numeric. dataframe containing species in rows and trait values in columns}

\item{fd_metrics}{character. Functional diversity metric}

\item{p}{list. progressor object for progress bar}
}
\value{
fd_metricss
}
\description{
get functional diversity metrics from dataframe
This function was inspired from FD package
}
