% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{get_HDR_name}
\alias{get_HDR_name}
\title{get hdr name from image file name, assuming it is BIL format}
\usage{
get_HDR_name(ImPath)
}
\arguments{
\item{ImPath}{path of the image}
}
\value{
corresponding hdr
}
\description{
get hdr name from image file name, assuming it is BIL format
}
