% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HDR_name.R
\name{get_HDR_name}
\alias{get_HDR_name}
\title{get hdr name from image file name, assuming it is BIL format}
\usage{
get_HDR_name(ImPath, showWarnings = TRUE)
}
\arguments{
\item{ImPath}{character. ath of the image}

\item{showWarnings}{boolean. set TRUE if warning because HDR does not exist}
}
\value{
corresponding hdr
}
\description{
get hdr name from image file name, assuming it is BIL format
}
