% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Hill.R
\name{get_Hill}
\alias{get_Hill}
\title{computes hill number from a distribution
The Hill numbers quantify biodiversity. The importance of the abundance
distribution increases with increasing Hill order.
For q=0, the Hill number is the richness, for q=1, it is the exponential
Shannon entropy and for q=2, it is the inverse Simpson index.
Note that the Hill order can also be a fraction, e.g. 0.5.}
\usage{
get_Hill(Distrib, q = 1)
}
\arguments{
\item{Distrib}{Distribution}

\item{q}{numeric. Hill order}
}
\value{
hill number corresponding to the distribution
}
\description{
computes hill number from a distribution
The Hill numbers quantify biodiversity. The importance of the abundance
distribution increases with increasing Hill order.
For q=0, the Hill number is the richness, for q=1, it is the exponential
Shannon entropy and for q=2, it is the inverse Simpson index.
Note that the Hill order can also be a fraction, e.g. 0.5.
}
