% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kmeans.R
\name{get_kmeans}
\alias{get_kmeans}
\title{computes k-means from nb_iter subsets taken from dataPCA}
\usage{
get_kmeans(
  rast_sample,
  nb_iter,
  nb_clusters = 50,
  nbCPU = 1,
  algorithm = "Hartigan-Wong",
  progressbar = TRUE
)
}
\arguments{
\item{rast_sample}{numeric. initial dataset sampled from PCA image}

\item{nb_iter}{numeric. nb of iterations averaged to compute diversity indices}

\item{nb_clusters}{numeric. number of clusters used in kmeans}

\item{nbCPU}{numeric. Number of CPUs available}

\item{algorithm}{character. algorithm used in the kmeans clustering}

\item{progressbar}{boolean. set true for progress bar during clustering}
}
\value{
list of centroids and parameters needed to center/reduce data
}
\description{
computes k-means from nb_iter subsets taken from dataPCA
}
