% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_si_tiles_from_raster.R
\name{get_si_tiles_from_raster}
\alias{get_si_tiles_from_raster}
\title{computes spectral indices over a spatial extent from a raster}
\usage{
get_si_tiles_from_raster(
  aoi,
  aoi_ID,
  rastobj,
  si_list,
  output_dir,
  maskobj = NULL,
  sensor_name = "sentinel-2",
  radiometric_filter = list(cloudMask = NULL, shadeMask = NULL, NDVIMask = NULL),
  p = NULL,
  overwrite = FALSE,
  siteName = NULL
)
}
\arguments{
\item{aoi}{spatVector.}

\item{aoi_ID}{character.}

\item{rastobj}{spatRaster}

\item{si_list}{character.}

\item{output_dir}{character.}

\item{maskobj}{spatRaster}

\item{sensor_name}{character.}

\item{radiometric_filter}{list.}

\item{p}{object}

\item{overwrite}{boolean.}

\item{siteName}{character.}
}
\value{
xxx
}
\description{
computes spectral indices over a spatial extent from a raster
}
