% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spectralSpecies.R
\name{get_spectralSpecies}
\alias{get_spectralSpecies}
\title{compute spectral species from inputdata}
\usage{
get_spectralSpecies(inputdata, Kmeans_info, selected_bands = NULL, nbCPU = 1)
}
\arguments{
\item{inputdata}{dataframe. information extracted from raster data}

\item{Kmeans_info}{list. obtained from function get_kmeans}

\item{selected_bands}{numeric. bands selected from inputdata}

\item{nbCPU}{numeric. Number of CPUs available}
}
\value{
SSwindow results of kmeans per window
}
\description{
compute spectral species from inputdata
}
