% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssd_full.R
\name{get_ssd_full}
\alias{get_ssd_full}
\title{get spectral species distribution for all clusters}
\usage{
get_ssd_full(ssd, nb_clusters, pcelim = 0.02)
}
\arguments{
\item{ssd}{dataframe. spectral species distribution}

\item{nb_clusters}{numeric. number of clusters used in kmeans}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}
}
\value{
SSwindow results of kmeans per window
}
\description{
get spectral species distribution for all clusters
}
