% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valid_pixels_from_tiles.R
\name{get_valid_pixels_from_tiles}
\alias{get_valid_pixels_from_tiles}
\title{get valid pixels from list of tiles}
\usage{
get_valid_pixels_from_tiles(feature_dir, plots, nbCPU = 1, mask_dir = NULL)
}
\arguments{
\item{feature_dir}{character. directory where features to be used by biodivMapR are stored.}

\item{plots}{list.}

\item{nbCPU}{numeric.}

\item{mask_dir}{character.}
}
\value{
nb_pix_valid
}
\description{
get valid pixels from list of tiles
}
