% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{getBCdiss}
\alias{getBCdiss}
\title{Computes BC dissimilarity for a given matrix}
\usage{
getBCdiss(Mat1, pcelim = 0.02, p = NULL)
}
\arguments{
\item{Mat1}{numeric. matrix of spectral species distribution}

\item{pcelim}{numeric. Minimum contribution in percent required for a spectral species.}

\item{p}{list. progressor object for progress bar}
}
\value{
BCtmp numeric. BC dissimilarity matrix corresponding to Mat1 and Mat2
}
\description{
Computes BC dissimilarity for a given matrix
}
