% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{ind2sub2}
\alias{ind2sub2}
\title{convert image coordinates from index to X-Y
image coordinates are given as index = (ID.col-1) * total.lines + ID.row}
\usage{
ind2sub2(Raster, Image.Index)
}
\arguments{
\item{Raster}{image raster object}

\item{Image.Index}{coordinates corresponding to the raster}
}
\description{
convert image coordinates from index to X-Y
image coordinates are given as index = (ID.col-1) * total.lines + ID.row
}
