% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kmeans.R
\name{init_kmeans}
\alias{init_kmeans}
\title{Compute kmeans from random subset of pixels extracted from an image}
\usage{
init_kmeans(
  input_rast,
  output_dir,
  input_mask = NULL,
  SelectBands = NULL,
  nbclusters = 50,
  nbIter = 20,
  Kmeans_info_save = NULL,
  Kmeans_info_read = NULL,
  maxPixel_kmeans = 1e+05,
  algorithm = "Hartigan-Wong",
  nbCPU = 1,
  verbose = T,
  progressbar = T
)
}
\arguments{
\item{input_rast}{character. Path and name of the image to be processed.}

\item{output_dir}{character. Path for output directory}

\item{input_mask}{character. Path and name of the mask corresponding to the image}

\item{SelectBands}{numeric. bands selected from input_rast}

\item{nbclusters}{numeric. number of clusters used in kmeans}

\item{nbIter}{numeric. nb of iterations averaged to compute diversity indices}

\item{Kmeans_info_save}{character. path where to save Kmeans_info}

\item{Kmeans_info_read}{character. path where to read Kmeans_info}

\item{maxPixel_kmeans}{numeric. max number of pixels to extract for kmeans}

\item{algorithm}{character. algorithm used in the kmeans clustering}

\item{nbCPU}{numeric. Number of CPUs available}

\item{verbose}{boolean. set true for messages}

\item{progressbar}{boolean. set true for progress bar during clustering}
}
\value{
Kmeans_info
}
\description{
Compute kmeans from random subset of pixels extracted from an image
}
