% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapSpectralSpecies.R
\name{init_kmeans}
\alias{init_kmeans}
\title{computes k-means from nb_partitions subsets taken from dataPCA}
\usage{
init_kmeans(dataPCA, Pix_Per_Partition, nb_partitions, nbclusters, nbCPU = 1)
}
\arguments{
\item{dataPCA}{numeric. initial dataset sampled from PCA image}

\item{Pix_Per_Partition}{numeric. number of pixels per iteration}

\item{nb_partitions}{numeric. number of k-means then averaged}

\item{nbclusters}{numeric.number of clusters used in kmeans}

\item{nbCPU}{numeric. Number of CPUs available}
}
\value{
list of centroids and parameters needed to center/reduce data
}
\description{
computes k-means from nb_partitions subsets taken from dataPCA
}
