% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_PCoA.R
\name{init_PCoA}
\alias{init_PCoA}
\title{initialize PCoA for beta diversity mapping}
\usage{
init_PCoA(
  input_rast,
  output_dir,
  window_size,
  Kmeans_info,
  SelectBands = NULL,
  input_mask = NULL,
  nbSamples = 1000,
  MinSun = 0.25,
  pcelim = 0.02,
  dimPCoA = 3,
  nbCPU = 1,
  Beta_info_save = NULL,
  Beta_info_read = NULL,
  verbose = T
)
}
\arguments{
\item{input_rast}{SpatRaster. raster to extract data from}

\item{output_dir}{character. Path for output directory}

\item{window_size}{numeric. window size for square plots}

\item{Kmeans_info}{list. obtained from prepare_init_kmeans}

\item{SelectBands}{numeric. bands selected from input_rast}

\item{input_mask}{SpatRaster. mask corresponding to raster to extract data from}

\item{nbSamples}{numeric. number of samples to be extracted}

\item{MinSun}{numeric. minimum percentage of sunlit pixels}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{dimPCoA}{numeric.}

\item{nbCPU}{numeric. Number of CPUs available}

\item{Beta_info_save}{character. path where to save Beta_info}

\item{Beta_info_read}{character. path where to read Beta_info}

\item{verbose}{boolean. set true for messages}
}
\value{
list including spectral species distribution & BC diss matrix per plot, BetaPCO model
}
\description{
initialize PCoA for beta diversity mapping
}
