% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_PCoA_samples.R
\name{init_PCoA_samples}
\alias{init_PCoA_samples}
\title{initialize PCoA for beta diversity mapping based on samples extracted from images}
\usage{
init_PCoA_samples(
  rast_sample,
  output_dir,
  Kmeans_info,
  SelectBands = NULL,
  pcelim = 0.02,
  dimPCoA = 3,
  nbCPU = 1,
  Beta_info_save = NULL,
  verbose = T
)
}
\arguments{
\item{rast_sample}{data frame containing samples to use}

\item{output_dir}{character. Path for output directory}

\item{Kmeans_info}{list. obtained from prepare_init_kmeans}

\item{SelectBands}{numeric. bands selected from input_rast}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{dimPCoA}{numeric.}

\item{nbCPU}{numeric. Number of CPUs available}

\item{Beta_info_save}{character. path where to save Beta_info}

\item{verbose}{boolean. set true for messages}
}
\value{
list including spectral species distribution & BC diss matrix per plot, BetaPCO model
}
\description{
initialize PCoA for beta diversity mapping based on samples extracted from images
}
