% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR_outliers.R
\name{IQR_outliers}
\alias{IQR_outliers}
\title{This function computes interquartile range (IQR) criterion, which can be used
as a criterion for outlier detection}
\usage{
IQR_outliers(DistVal, weightIRQ = 1.5)
}
\arguments{
\item{DistVal}{numeric. vector of distribution of values}

\item{weightIRQ}{numeric. weighting factor applied to IRQ to define
lower and upper boundaries for outliers}
}
\value{
outlier_IQR numeric. band numbers of original sensor corresponding to S2
}
\description{
This function computes interquartile range (IQR) criterion, which can be used
as a criterion for outlier detection
}
