% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR_SpatRaster.R
\name{IQR_SpatRaster}
\alias{IQR_SpatRaster}
\title{This function computes interquartile range (IQR) for a SpatRaster}
\usage{
IQR_SpatRaster(input_rast, weightIRQ = 3)
}
\arguments{
\item{input_rast}{SpatRaster.}

\item{weightIRQ}{numeric. weighting factor applied to IRQ to define
lower and upper boundaries for outliers}
}
\value{
outlier_IQR numeric. band numbers of original sensor corresponding to S2
}
\description{
This function computes interquartile range (IQR) for a SpatRaster
}
