% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_progressr.R
\name{kmeans_progressr}
\alias{kmeans_progressr}
\title{applies results of ordination to full image based on nearest neighbors}
\usage{
kmeans_progressr(
  x,
  centers,
  iter.max,
  nstart,
  algorithm = "Hartigan-Wong",
  p = NULL
)
}
\arguments{
\item{x}{numeric.}

\item{centers}{numeric.}

\item{iter.max}{numeric.}

\item{nstart}{numeric.}

\item{algorithm}{character. algorithm used in the kmeans clustering}

\item{p}{function.}
}
\value{
results of kmeans
}
\description{
applies results of ordination to full image based on nearest neighbors
}
