% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapSpectralSpecies.R
\name{map_spectral_species}
\alias{map_spectral_species}
\title{maps spectral species based on PCA file computed previously}
\usage{
map_spectral_species(Input_Image_File, Output_Dir, PCA_Files, ImPathShade,
  Pix_Per_Partition, nb_partitions, TypePCA = "SPCA", nbclusters = 50,
  nbCPU = 1, MaxRAM = FALSE)
}
\arguments{
\item{Input_Image_File}{character. Path of the image to be processed}

\item{Output_Dir}{character. Path for output directory}

\item{PCA_Files}{character. Path of the PCA image}

\item{ImPathShade}{character. Path of the mask corresponding to the image}

\item{Pix_Per_Partition}{number of pixels for each partition}

\item{nb_partitions}{number of partition}

\item{TypePCA}{character. Type of PCA: choose either "PCA" or "SPCA"}

\item{nbclusters}{number of clusters defined in k-Means}

\item{nbCPU}{numeric. Number of CPUs to use in parallel.}

\item{MaxRAM}{numeric. MaxRAM maximum size of chunk in GB to limit RAM allocation when reading image file.}
}
\description{
maps spectral species based on PCA file computed previously
}
