% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapSpectralSpecies.R
\name{map_spectral_species}
\alias{map_spectral_species}
\title{maps spectral species based on PCA file computed previously}
\usage{
map_spectral_species(
  Input_Image_File,
  Output_Dir,
  SpectralSpace_Output,
  Input_Mask_File = FALSE,
  nbclusters = 50,
  nbCPU = 1,
  MaxRAM = 0.25,
  Kmeans_Only = FALSE,
  SelectedPCs = FALSE,
  SpectralFilter = NULL,
  Kmeans_info_path = NULL,
  Kmeans_info = NULL
)
}
\arguments{
\item{Input_Image_File}{character. Path of the image to be processed}

\item{Output_Dir}{character. Path for output directory}

\item{SpectralSpace_Output}{list. list of variables produced from function perform_PCA}

\item{Input_Mask_File}{character. Path of the mask corresponding to the image}

\item{nbclusters}{numeric. number of clusters defined in k-Means}

\item{nbCPU}{numeric. Number of CPUs to use in parallel.}

\item{MaxRAM}{numeric. MaxRAM maximum size of chunk in GB to limit RAM allocation when reading image file.}

\item{Kmeans_Only}{boolean. set to TRUE if computation of kmeans without production of spectral species map}

\item{SelectedPCs}{numeric. Define PCs to be selected. Set to FALSE if you want to use the "Selected_Components.txt" file}

\item{SpectralFilter}{list. information about spectral band location}

\item{Kmeans_info_path}{character. path to a Kmeans_info Rdata file computed from previous process}

\item{Kmeans_info}{list. Kmeans_info list computed from previous process
(central wavelength), bands to keep...}
}
\value{
Kmeans_info
}
\description{
maps spectral species based on PCA file computed previously
}
