% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_tiles.R
\name{mosaic_tiles}
\alias{mosaic_tiles}
\title{compute spectral indices and composite for a given period of time}
\usage{
mosaic_tiles(pattern, dir_path, vrt_save, site_name = NULL, overwrite = FALSE)
}
\arguments{
\item{pattern}{character. identify which files to mosaic}

\item{dir_path}{character. firectory where rasters to mosaic are stored}

\item{vrt_save}{character. where to save vrt}

\item{site_name}{character. name of the site}

\item{overwrite}{boolean}
}
\value{
mosaic_path
}
\description{
compute spectral indices and composite for a given period of time
}
