% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{ordination_to_NN}
\alias{ordination_to_NN}
\title{identifies ordination coordinates based on nearest neighbors}
\usage{
ordination_to_NN(
  Beta_Ordination_sel,
  SSD_Path,
  Sample_Sel,
  coordTotSort,
  nb_partitions,
  nbclusters,
  pcelim,
  nbCPU = FALSE
)
}
\arguments{
\item{Beta_Ordination_sel}{ordination}

\item{SSD_Path}{ath for spectral species distribution file}

\item{Sample_Sel}{Samples selected during ordination}

\item{coordTotSort}{coordinates of sunlit spatial units}

\item{nb_partitions}{number of k-means then averaged}

\item{nbclusters}{number of clusters}

\item{pcelim}{number of CPUs available}

\item{nbCPU}{number of CPUs available}
}
\value{
Ordination_est coordinates of each spatial unit in ordination space
}
\description{
identifies ordination coordinates based on nearest neighbors
}
