% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{ordination_to_NN}
\alias{ordination_to_NN}
\title{Identifies ordination coordinates based on nearest neighbors}
\usage{
ordination_to_NN(
  Beta_Ordination_sel,
  SSD_Path,
  Sample_Sel,
  coordTotSort,
  nb_partitions,
  nbclusters,
  pcelim,
  nbCPU = 1
)
}
\arguments{
\item{Beta_Ordination_sel}{ordination of dissimilarity matrix for a selection of spatial units}

\item{SSD_Path}{character. path for spectral species distribution file}

\item{Sample_Sel}{numeric. Samples selected during ordination}

\item{coordTotSort}{numeric. coordinates of sunlit spatial units}

\item{nb_partitions}{number of k-means then averaged}

\item{nbclusters}{number of clusters}

\item{pcelim}{numeric. Minimum contribution in percent required for a spectral species}

\item{nbCPU}{numeric. number of CPUs available}
}
\value{
Ordination_est coordinates of each spatial unit in ordination space
}
\description{
Identifies ordination coordinates based on nearest neighbors
}
