% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{performs PCA on matrix X}
\usage{
pca(X, type = "SPCA")
}
\arguments{
\item{X}{matrix to apply PCA on}

\item{type}{PCA (no rescale) or SPCA (rescale)}
}
\value{
list of PCA parameters (PCs from X, mean, eigenvectors and values)
}
\description{
performs PCA on matrix X
}
