% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_PCA.R
\name{perform_PCA}
\alias{perform_PCA}
\title{Performs PCA for all images and create PCA file with either all or a selection of PCs}
\usage{
perform_PCA(
  input_raster_path,
  output_dir,
  input_rast_wl = NULL,
  input_mask_path = NULL,
  Continuum_Removal = TRUE,
  TypePCA = "SPCA",
  NbPCs_To_Keep = 30,
  Excluded_WL = NULL,
  nbPix_PCA = 1e+06,
  nbIter = 20,
  maxRows = 100,
  filetype = "GTiff"
)
}
\arguments{
\item{input_raster_path}{character. path for image to be processed}

\item{output_dir}{character. Path for output directory}

\item{input_rast_wl}{numeric. spectral bands corresponding to input_raster_path}

\item{input_mask_path}{character. path for mask corresponding to the image}

\item{Continuum_Removal}{boolean. Set to TRUE if continuum removal should be applied}

\item{TypePCA}{character. Type of PCA: choose either "PCA" or "SPCA"}

\item{NbPCs_To_Keep}{numeric. number of components to ke saved in the PCA file. default = 30 if set to FALSE (or nb PC if <30)}

\item{Excluded_WL}{numeric. Water Vapor Absorption domains (in nanometers, min and max WL).
Can also be used to exclude spectific domains.
dims = N x 2 (N = number of domains to be eliminated)}

\item{nbPix_PCA}{numeric. number of pixels to use to compute PCA}

\item{nbIter}{numeric. nb of iterations averaged to compute diversity indices}

\item{maxRows}{numeric. max number of rows in each block}

\item{filetype}{character. gdal driver}
}
\value{
list of paths corresponding to resulting PCA files
}
\description{
Performs PCA for all images and create PCA file with either all or a selection of PCs
}
