% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PerformPCA.R
\name{perform_PCA}
\alias{perform_PCA}
\title{Performs PCA for all images and create PCA file with either all or a selection of PCs}
\usage{
perform_PCA(ImPath, ImPathShade, Output_Dir, Continuum_Removal = TRUE,
  TypePCA = "SPCA", FilterPCA = FALSE, Excluded_WL = FALSE,
  nb_partitions = 20, nbCPU = 1, MaxRAM = 0.25)
}
\arguments{
\item{ImPath}{character. Path of the image to be processed}

\item{ImPathShade}{character. Path of the mask corresponding to the image}

\item{Output_Dir}{character. Path for output directory}

\item{Continuum_Removal}{boolean. Set to TRUE if continuum removal should be applied}

\item{TypePCA}{character. Type of PCA: choose either "PCA" or "SPCA"}

\item{FilterPCA}{boolean. Set to TRUE if 2nd filtering based on PCA is required}

\item{Excluded_WL}{numeric. Water Vapor Absorption domains (in nanometers, min and max WL). Can also be used to exclude spectific domains. dims = N x 2 (N = number of domains to be eliminated)}

\item{nb_partitions}{numeric. Number of repetitions to estimate diversity from the raster (averaging repetitions).}

\item{nbCPU}{numeric. Number fo CPUs in use.}

\item{MaxRAM}{numeric. Maximum size of chunk in GB to limit RAM allocation when reading image file.}
}
\value{
list of paths corresponding to resulting PCA files
}
\description{
Performs PCA for all images and create PCA file with either all or a selection of PCs
}
