% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_FilterData.R
\name{perform_radiometric_filtering}
\alias{perform_radiometric_filtering}
\title{Performs radiometric filtering based on three criteria: NDVI, NIR reflectance, Blue reflectance}
\usage{
perform_radiometric_filtering(Image_Path, Mask_Path, Output_Dir,
  TypePCA = "SPCA", NDVI_Thresh = 0.5, Blue_Thresh = 500,
  NIR_Thresh = 1500, Blue = 480, Red = 700, NIR = 835)
}
\arguments{
\item{Image_Path}{character. Path of the image to be processed}

\item{Mask_Path}{character. Path of the mask corresponding to the image}

\item{Output_Dir}{character. Path for output directory}

\item{TypePCA}{character. Type of PCA: choose either "PCA" or "SPCA"}

\item{NDVI_Thresh}{numeric. NDVI threshold applied to produce a mask (select pixels with NDVI>NDVI_Thresh)}

\item{Blue_Thresh}{numeric. Blue threshold applied to produce a mask (select pixels with Blue refl < Blue_Thresh --> filter clouds) refl expected between 0 and 10000}

\item{NIR_Thresh}{numeric. NIR threshold applied to produce a mask (select pixels with NIR refl < NIR_Thresh) refl expected between 0 and 10000}

\item{Blue}{numeric. central wavelength corresponding to the blue spectral band (in nanometers)}

\item{Red}{numeric. central wavelength corresponding to the red spectral band (in nanometers)}

\item{NIR}{numeric. central wavelength corresponding to the NIR spectral band (in nanometers)}
}
\value{
MaskPath = updated mask file
}
\description{
Performs radiometric filtering based on three criteria: NDVI, NIR reflectance, Blue reflectance
}
