% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_win_ID.R
\name{produce_win_ID}
\alias{produce_win_ID}
\title{design a matrix with window ID based on an original raster and window size in pixels}
\usage{
produce_win_ID(inputdata, blk, window_size)
}
\arguments{
\item{inputdata}{numeric. image chunk}

\item{blk}{list.}

\item{window_size}{numeric. window size for square plots}
}
\value{
blk
}
\description{
design a matrix with window ID based on an original raster and window size in pixels
}
