% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_CheckConvertData.R
\name{raster2BIL}
\alias{raster2BIL}
\title{converts a raster into BIL format as expected by DivMapping codes}
\usage{
raster2BIL(Raster_Path, Sensor = "unknown", Output_Dir = FALSE,
  Convert_Integer = TRUE, Multiplying_Factor = 1,
  Multiplying_Factor_Last = 1)
}
\arguments{
\item{Raster_Path}{character. Full path for the raster to be converted}

\item{Sensor}{character. Name of the sensor. a .hdr template for the sensor should be provided in extdata/HDR}

\item{Output_Dir}{character. Path to output directory.}

\item{Convert_Integer}{boolean. Should data be converted into integer ?}

\item{Multiplying_Factor}{numeric. Multiplying factor (eg convert real reflectance values between 0 and 1 into integer between 0 and 10000).}

\item{Multiplying_Factor_Last}{numeric. Multiplying factor for last band.}
}
\value{
Output_Path path for the image converted into ENVI BIL format
}
\description{
converts a raster into BIL format as expected by DivMapping codes
}
