% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_biodivMapR.R
\name{run_biodivMapR}
\alias{run_biodivMapR}
\title{apply biodivMapR to an individual set of rasters (one site)}
\usage{
run_biodivMapR(
  input_raster_path,
  input_mask_path = NULL,
  Kmeans_info,
  Beta_info,
  output_dir,
  output_raster_name,
  selected_bands = NULL,
  window_size,
  alphametrics = "shannon",
  Hill_order = 1,
  FDmetric = NULL,
  pcelim = 0.02,
  maxRows = NULL,
  nbCPU = 1,
  min_sun = 0.25,
  filetype = "GTiff",
  moving_window = FALSE
)
}
\arguments{
\item{input_raster_path}{character.}

\item{input_mask_path}{directory where spectral indices are for each plot}

\item{Kmeans_info}{list. kmeans description obtained from function get_kmeans}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics}

\item{output_dir}{path where to save outputs}

\item{output_raster_name}{raster file names}

\item{selected_bands}{numeric. bands selected from input data}

\item{window_size}{numeric. window size for square plots}

\item{alphametrics}{list. alpha diversity metrics: richness, shannon, simpson}

\item{Hill_order}{numeric. Hill order}

\item{FDmetric}{character. list of functional metrics}

\item{pcelim}{numeric. min proportion of pixels to consider spectral species}

\item{maxRows}{numeric. maximum number or rows to process}

\item{nbCPU}{numeric. Number of CPUs available}

\item{min_sun}{numeric. minimum proportion of sunlit pixels}

\item{filetype}{character. gdal driver name}

\item{moving_window}{boolean. should process be moving window (much longer)}
}
\value{
none
}
\description{
apply biodivMapR to an individual set of rasters (one site)
}
