% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_biodivMapR_plot.R
\name{run_biodivMapR_plot}
\alias{run_biodivMapR_plot}
\title{apply biodivMapR to a set of plots identified by a field '\emph{id}'
produced with preprocS2 function get_s2_tiling}
\usage{
run_biodivMapR_plot(
  id,
  feature_dir,
  mask_dir = NULL,
  list_features,
  Kmeans_info,
  Beta_info,
  output_dir,
  SelectBands = NULL,
  window_size,
  alphametrics = "shannon",
  Hill_order = 1,
  FDmetric = NULL,
  pcelim = 0.02,
  maxRows = NULL,
  nbCPU = 1,
  MinSun = 0.25,
  filetype = "GTiff",
  MovingWindow = F,
  p = NULL
)
}
\arguments{
\item{id}{character. ID for plot}

\item{feature_dir}{directory where spectral indices are for each plot}

\item{mask_dir}{character.}

\item{list_features}{character.}

\item{Kmeans_info}{list. kmeans description obtained from function get_kmeans}

\item{Beta_info}{list. BC dissimilarity & associated beta metrics from training set}

\item{output_dir}{path where to save outputs}

\item{SelectBands}{numeric. bands selected from input data}

\item{window_size}{numeric. window size for square plots}

\item{alphametrics}{list. alpha diversity metrics: richness, shannon, simpson}

\item{Hill_order}{numeric. Hill order}

\item{FDmetric}{character. list of functional metrics}

\item{pcelim}{numeric. minimum proportion of pixels to consider spectral species}

\item{maxRows}{numeric. maximum number or rows to process}

\item{nbCPU}{numeric. Number of CPUs available}

\item{MinSun}{numeric. minimum amount of sunlit pixels in the plots}

\item{filetype}{character. gdal driver for output raster}

\item{MovingWindow}{boolean. should process be moving window (much longer)}

\item{p}{list. progressbar}
}
\value{
none
}
\description{
apply biodivMapR to a set of plots identified by a field '\emph{id}'
produced with preprocS2 function get_s2_tiling
}
