% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_plots.R
\name{sample_from_plots}
\alias{sample_from_plots}
\title{get samples for alpha and beta diversity mapping}
\usage{
sample_from_plots(
  feature_dir,
  list_features,
  plot_names,
  mask_dir = NULL,
  window_size,
  nbCPU = 1,
  Kmeans_info = NULL,
  compute_beta = TRUE,
  nb_samples_alpha = 1e+05,
  nb_samples_beta = 2000
)
}
\arguments{
\item{feature_dir}{character.}

\item{list_features}{character.}

\item{plot_names}{character. plot names}

\item{mask_dir}{character.}

\item{window_size}{numeric.}

\item{nbCPU}{numeric.}

\item{Kmeans_info}{list.}

\item{compute_beta}{boolean. set TRUE to compute beta diversity}

\item{nb_samples_alpha}{numeric.}

\item{nb_samples_beta}{numeric.}
}
\value{
samples_alpha_beta
}
\description{
get samples for alpha and beta diversity mapping
}
