% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_plots.R
\name{sample_from_plots}
\alias{sample_from_plots}
\title{get samples for alpha and beta diversity mapping}
\usage{
sample_from_plots(
  feature_dir,
  list_features,
  plots,
  mask_dir = NULL,
  window_size,
  nbCPU = 1,
  nbsamples_alpha = 1e+05,
  nbsamples_beta = 2000
)
}
\arguments{
\item{feature_dir}{character.}

\item{list_features}{character.}

\item{plots}{list.}

\item{mask_dir}{character.}

\item{window_size}{numeric.}

\item{nbCPU}{numeric.}

\item{nbsamples_alpha}{numeric.}

\item{nbsamples_beta}{numeric.}
}
\value{
samples_alpha_beta
}
\description{
get samples for alpha and beta diversity mapping
}
