% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_plots_beta.R
\name{sample_from_plots_beta}
\alias{sample_from_plots_beta}
\title{get samples for beta diversity mapping}
\usage{
sample_from_plots_beta(
  feature_dir,
  list_features,
  plots,
  nbPixValid,
  mask_dir = NULL,
  window_size,
  nbsamples_beta = 2000,
  nbCPU = 1
)
}
\arguments{
\item{feature_dir}{character.}

\item{list_features}{character.}

\item{plots}{list.}

\item{nbPixValid}{numeric.}

\item{mask_dir}{character.}

\item{window_size}{numeric.}

\item{nbsamples_beta}{numeric.}

\item{nbCPU}{numeric.}
}
\value{
samples_beta_terra
}
\description{
get samples for beta diversity mapping
}
