% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_raster.R
\name{sample_raster}
\alias{sample_raster}
\title{sample pixels from raster data}
\usage{
sample_raster(input_rast, pix2extract, xy = F, prog = F)
}
\arguments{
\item{input_rast}{SpatRaster or list of SpatRaster}

\item{pix2extract}{dataframe. lat and lon for pixels to extract}

\item{xy}{boolean. set TRUE to get coordinates for each pixel sampled}

\item{prog}{boolean. set TRUE to display progression bar}
}
\value{
rast_sample dataframe. pixel/plot info extracted from input_rast
}
\description{
sample pixels from raster data
}
