% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_diversity_maps.R
\name{save_diversity_maps}
\alias{save_diversity_maps}
\title{save diversity maps as raster data}
\usage{
save_diversity_maps(
  ab_div_metrics,
  alpha_metrics = "shannon",
  Hill_order = 1,
  beta_metrics = TRUE,
  fd_metrics = NULL,
  input_rast,
  output_dir,
  output_raster_name = NULL,
  window_size,
  filetype = "GTiff"
)
}
\arguments{
\item{ab_div_metrics}{list produced from get_raster_diversity}

\item{alpha_metrics}{list. alpha diversity metrics}

\item{Hill_order}{numeric. Hill order}

\item{beta_metrics}{boolean. set TRUE to compute beta diversity}

\item{fd_metrics}{character. list of functional metrics}

\item{input_rast}{list. path for input rasters}

\item{output_dir}{character.}

\item{output_raster_name}{character.}

\item{window_size}{numeric. window size for square plots}

\item{filetype}{character. GDAL driver}
}
\value{
blk
}
\description{
save diversity maps as raster data
}
