% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_diversity_maps_tile.R
\name{save_diversity_maps_tile}
\alias{save_diversity_maps_tile}
\title{save diversity maps as raster data}
\usage{
save_diversity_maps_tile(
  input_raster_path,
  ab_div_metrics,
  alphametrics = "shannon",
  Hill_order = 1,
  FDmetric = NULL,
  input_rast,
  output_dir,
  output_raster_name = NULL,
  window_size,
  filetype = "GTiff"
)
}
\arguments{
\item{input_raster_path}{list. list of paths corresponding to input rasters}

\item{ab_div_metrics}{list produced from get_raster_diversity}

\item{alphametrics}{list. alpha diversity metrics}

\item{Hill_order}{numeric. Hill order}

\item{FDmetric}{character. list of functional metrics}

\item{input_rast}{list. path for input rasters}

\item{output_dir}{character.}

\item{output_raster_name}{character.}

\item{window_size}{numeric. window size for square plots}

\item{filetype}{character. GDAL driver}
}
\value{
blk
}
\description{
save diversity maps as raster data
}
