% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_pca_components.R
\name{select_PCA_components}
\alias{select_PCA_components}
\title{Check if principal components are properly selected as expected by the method}
\usage{
select_PCA_components(pca_rast_path, output_dir, File_Open = TRUE)
}
\arguments{
\item{pca_rast_path}{character. Path of the PCA image}

\item{output_dir}{character. Path for output directory}

\item{File_Open}{Boolean. Set to TRUE for file to open automatically}
}
\value{
Sel_PC
}
\description{
Check if principal components are properly selected as expected by the method
}
