% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PerformPCA.R
\name{select_PCA_components}
\alias{select_PCA_components}
\title{Check if principal components are properly selected as expected by the method}
\usage{
select_PCA_components(Input_Image_File, Output_Dir, PCA_Files,
  TypePCA = "SPCA", File_Open = FALSE)
}
\arguments{
\item{Input_Image_File}{character. Path of the image to be processed}

\item{Output_Dir}{character. Path for output directory}

\item{PCA_Files}{character. Path of the PCA image}

\item{TypePCA}{character. Type of PCA: choose either "PCA" or "SPCA"}

\item{File_Open}{Boolean. Set to TRUE for file to open automatically}
}
\value{
Sel_PC
}
\description{
Check if principal components are properly selected as expected by the method
}
