% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_species_full_tiles.R
\name{spectral_species_full_tiles}
\alias{spectral_species_full_tiles}
\title{computes spectral species from raster}
\usage{
spectral_species_full_tiles(
  feature_dir,
  list_features,
  mask_dir = NULL,
  output_dir,
  plots,
  nbCPU = 1,
  site_name = NULL,
  overwrite = TRUE,
  options = NULL
)
}
\arguments{
\item{feature_dir}{character. path where to get features}

\item{list_features}{character. list of features}

\item{mask_dir}{character. path for masks}

\item{output_dir}{character. path where to save results}

\item{plots}{list. list of sf plots}

\item{nbCPU}{numeric. Number of CPUs available}

\item{site_name}{character. name for the output files}

\item{overwrite}{boolean.}

\item{options}{list. including
\itemize{
\item nb_clusters numeric. number of clusters
\item nb_samples_alpha numeric. number of samples to compute alpha diversity
\item mosaic_output boolean. set TRUE if outputs need to be mosaiced
\item weightIQR numeric. IQR applied to filter out features to be used
}}
}
\value{
ss_path character. path for the spectral species file
}
\description{
computes spectral species from raster
}
