% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralspecies_per_rast_sample.R
\name{spectralspecies_per_rast_sample}
\alias{spectralspecies_per_rast_sample}
\title{get spectral species corresponding to raster sample extracted with extract}
\usage{
spectralspecies_per_rast_sample(
  input_rast,
  Kmeans_info,
  selected_bands = NULL,
  fd_metrics = NULL,
  input_mask = NULL,
  rast_sample = NULL,
  AttributeTable = NULL,
  min_sun = 0.25
)
}
\arguments{
\item{input_rast}{SpatRaster.}

\item{Kmeans_info}{list. kmeans description obtained from function get_kmeans}

\item{selected_bands}{numeric. bands selected from input data}

\item{fd_metrics}{boolean. should functional diversity be computed as well?}

\item{input_mask}{SpatRaster.}

\item{rast_sample}{dataframe.}

\item{AttributeTable}{dataframe.}

\item{min_sun}{numeric. minimum amount of sunlit pixels in the plots}
}
\value{
list
}
\description{
get spectral species corresponding to raster sample extracted with extract
}
