% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_chunk.R
\name{split_chunk}
\alias{split_chunk}
\title{split chunk into subsets to prepare for parallel processing}
\usage{
split_chunk(SSchunk, nbCPU)
}
\arguments{
\item{SSchunk}{list. obtained from get_spectralSpecies}

\item{nbCPU}{numeric. Number of CPUs available}
}
\value{
estimated NMDS position based on nearest neighbors from NMDS
}
\description{
split chunk into subsets to prepare for parallel processing
}
