% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_line.R
\name{split_line}
\alias{split_line}
\title{ENVI functions}
\usage{
split_line(x, separator, trim.blank = TRUE)
}
\arguments{
\item{x}{character.}

\item{separator}{character}

\item{trim.blank}{boolean.}
}
\value{
list.
}
\description{
based on https://github.com/cran/hyperSpec/blob/master/R/read.ENVI.R
added wavelength, fwhm, ... to header reading
Title
}
