% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_ImageProcess.R
\name{sub2ind}
\alias{sub2ind}
\title{convert image coordinates from X-Y to index}
\usage{
sub2ind(HDR.Raster, Pixels)
}
\arguments{
\item{Pixels}{coordinates corresponding to the raster}
}
\value{
Image.Index
}
\description{
convert image coordinates from X-Y to index
}
