% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_MapBetaDiversity.R
\name{WeightedCoordsNN}
\alias{WeightedCoordsNN}
\title{Compute the weighted coordinates of a spatial unit based on nearest neighbors used during PCoA}
\usage{
WeightedCoordsNN(NN, knn, BetaDiversity0)
}
\arguments{
\item{NN}{list. coordinates and ID of nearest neighbors from BetaDiversity0 matrix}

\item{knn}{number of neighbors}

\item{BetaDiversity0}{PCoA coordinates of reference samples}
}
\value{
estimated NMDS position based on nearest neighbors from NMDS
}
\description{
Compute the weighted coordinates of a spatial unit based on nearest neighbors used during PCoA
}
